Refactor doltlite.c into focused command modules
Split the ~5.8k-line doltlite.c god file into command-sized modules,
matching the existing doltlite_*.c layout and the prolly_btree split:

- doltlite_core: txn seal, catalog flush, create-commit, conflict report
- doltlite_add / doltlite_commit_cmd / doltlite_reset / doltlite_merge_cmd
- doltlite_cherry_pick / doltlite_revert / doltlite_rebase / doltlite_config
- thin doltlite.c: doltliteRegister only

Promote shared helpers into doltlite_internal.h, wire main.mk and the
amalgamation list. No intentional behavior change.
