+++ AGENTS.md
@@ -73,13 +73,17 @@ make DOLTLITE_PROLLY=0 sqlite3   # stock SQLite, for oracle/perf comparison
   diff, three-way merge, chunker, cache, hashing) and `src/chunk_*.c`
   (content-addressed chunk store, WAL, refs, staging, file format).
 - **Version-control surfaces** — `src/doltlite_*.c`, roughly one file per
-  feature: `doltlite_commit`, `doltlite_branch` (branch/checkout/rebase),
-  `doltlite_merge` (+ `doltlite_merge_constraints`), `doltlite_diff` /
-  `doltlite_diff_stat` / `doltlite_diff_table`, `doltlite_log`,
-  `doltlite_history`, `doltlite_blame`, `doltlite_tag`, `doltlite_conflicts`,
-  `doltlite_constraint_violations`, `doltlite_schemas` / `doltlite_schema_diff`,
-  `doltlite_patch`, `doltlite_status`, `doltlite_workspace`, `doltlite_ignore`,
-  `doltlite_hashof`, `doltlite_gc`, `doltlite_remote` / `doltlite_http_remote` /
+  feature: `doltlite_commit` (wire format), `doltlite_commit_cmd` /
+  `doltlite_add` / `doltlite_reset` / `doltlite_merge_cmd` /
+  `doltlite_cherry_pick` / `doltlite_revert` / `doltlite_rebase` /
+  `doltlite_config` (SQL commands), `doltlite_core` (txn seal, catalog flush,
+  create-commit helpers), `doltlite_branch` (branch/checkout), `doltlite_merge`
+  (+ `doltlite_merge_constraints`), `doltlite_diff` / `doltlite_diff_stat` /
+  `doltlite_diff_table`, `doltlite_log`, `doltlite_history`, `doltlite_blame`,
+  `doltlite_tag`, `doltlite_conflicts`, `doltlite_constraint_violations`,
+  `doltlite_schemas` / `doltlite_schema_diff`, `doltlite_patch`,
+  `doltlite_status`, `doltlite_workspace`, `doltlite_ignore`, `doltlite_hashof`,
+  `doltlite_gc`, `doltlite_remote` / `doltlite_http_remote` /
   `doltlite_remotesrv`. Internal header: `src/doltlite_internal.h`; surfaces are
   registered in `src/doltlite.c`.
