+++ src/doltlite.c
@@ -1,64 +1,7 @@
-
 #ifdef DOLTLITE_PROLLY
 
 #include "sqliteInt.h"
-#include "prolly_hash.h"
-#include "prolly_hashset.h"
-#include "chunk_store.h"
-#include "prolly_cursor.h"
-#include "prolly_cache.h"
-#include "prolly_diff.h"
-#include "doltlite_commit.h"
-#include "doltlite_record.h"
 #include "doltlite_internal.h"
-#include "doltlite_name_index.h"
-#include <stddef.h>
-#include "doltlite_ignore.h"
-
-#include <string.h>
-#include <ctype.h>
-#include <time.h>
-
-#ifdef _WIN32
-static const char *dlWinStrptime(
-  const char *zDate,
-  const char *zFmt,
-  struct tm *pTm
-){
-  int year = 0, month = 0, day = 0;
-  int hour = 0, minute = 0, second = 0;
-  char sep = 0;
-  int n = 0;
-
-  if( strcmp(zFmt, "%Y-%m-%dT%H:%M:%S")==0 ){
-    n = sscanf(zDate, "%d-%d-%dT%d:%d:%d%c",
-               &year, &month, &day, &hour, &minute, &second, &sep);
-    if( n!=6 ) return 0;
-  }else if( strcmp(zFmt, "%Y-%m-%d %H:%M:%S")==0 ){
-    n = sscanf(zDate, "%d-%d-%d %d:%d:%d%c",
-               &year, &month, &day, &hour, &minute, &second, &sep);
-    if( n!=6 ) return 0;
-  }else{
-    return 0;
-  }
-
-  memset(pTm, 0, sizeof(*pTm));
-  pTm->tm_year = year - 1900;
-  pTm->tm_mon = month - 1;
-  pTm->tm_mday = day;
-  pTm->tm_hour = hour;
-  pTm->tm_min = minute;
-  pTm->tm_sec = second;
-  pTm->tm_isdst = 0;
-  return zDate + sqlite3Strlen30(zDate);
-}
-
-static time_t dlWinTimegm(struct tm *pTm){
-  return _mkgmtime(pTm);
-}
-#define strptime dlWinStrptime
-#define timegm dlWinTimegm
-#endif
 
 extern int doltliteLogRegister(sqlite3 *db);
 extern int doltliteCommitAncestorsRegister(sqlite3 *db);
@@ -68,5704 +11,30 @@ extern int doltliteSchemasRegister(sqlite3 *db);
 extern int doltliteDiffStatRegister(sqlite3 *db);
 extern int doltliteBranchRegister(sqlite3 *db);
 extern int doltliteConflictsRegister(sqlite3 *db);
-extern int doltliteRegisterConflictTables(sqlite3 *db);
 extern int doltliteTagRegister(sqlite3 *db);
 extern int doltliteGcRegister(sqlite3 *db);
 extern int doltliteRegisterDiffTables(sqlite3 *db);
 extern int doltliteRegisterWorkspaceTables(sqlite3 *db);
 extern int doltliteAncestorRegister(sqlite3 *db);
 extern int doltliteRegisterAtTables(sqlite3 *db);
-extern int doltliteRegisterAtTablesForCatalog(sqlite3 *db, const ProllyHash *pCatHash);
 extern int doltliteRegisterHistoryTables(sqlite3 *db);
 extern int doltliteRegisterBlameTables(sqlite3 *db);
-extern int doltliteRefreshConstraintViolationTables(sqlite3 *db);
 extern int doltliteSchemaDiffRegister(sqlite3 *db);
 extern int doltlitePatchRegister(sqlite3 *db);
 extern int doltliteRemoteSqlRegister(sqlite3 *db);
-
-extern int doltliteFindAncestor(sqlite3 *db, const ProllyHash *h1,
-                                 const ProllyHash *h2, ProllyHash *pAnc);
-
-extern void doltliteSetTableSchemaHash(sqlite3 *db, Pgno iTable, const ProllyHash *pH);
-
-void doltliteTxnStateClear(DoltliteTxnState *p){
-  sqlite3_free(p->zSessionBranch);
-  memset(p, 0, sizeof(*p));
-}
-
-int doltliteSaveTxnState(sqlite3 *db, DoltliteTxnState *p){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  int rc;
-
-  memset(p, 0, sizeof(*p));
-  if( !cs ) return SQLITE_ERROR;
-
-  memcpy(&p->refsHash, refsTableGetHash(&cs->refs), sizeof(ProllyHash));
-
-  p->zSessionBranch = sqlite3_mprintf("%s", doltliteGetSessionBranch(db));
-  if( !p->zSessionBranch ){
-    doltliteTxnStateClear(p);
-    return SQLITE_NOMEM;
-  }
-  doltliteGetSessionHead(db, &p->sessionHead);
-  doltliteGetSessionStaged(db, &p->sessionStaged);
-  doltliteGetSessionMergeState(db, &p->sessionIsMerging,
-                               &p->sessionMergeCommit,
-                               &p->sessionConflictsCatalog);
-  doltliteGetSessionConstraintViolationsCatalog(
-      db, &p->sessionConstraintViolationsCatalog);
-
-  rc = doltliteFlushCatalogToHash(db, &p->sessionCatalogHash);
-  if( rc!=SQLITE_OK ){
-    doltliteTxnStateClear(p);
-  }
-  return rc;
-}
-
-int doltliteRestoreTxnState(sqlite3 *db, DoltliteTxnState *p){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  int rc;
-
-  if( !cs ) return SQLITE_ERROR;
-
-  refsTableSetHash(&cs->refs, &p->refsHash);
-  if( prollyHashIsEmpty(&p->refsHash) ){
-    chunkStoreClearRefs(cs);
-  }else{
-    rc = chunkStoreReloadRefs(cs);
-    if( rc!=SQLITE_OK ) return rc;
-  }
-
-  rc = doltliteSwitchCatalog(db, &p->sessionCatalogHash);
-  if( rc!=SQLITE_OK ) return rc;
-
-  doltliteSetSessionBranch(db, p->zSessionBranch);
-  doltliteSetSessionHead(db, &p->sessionHead);
-  doltliteSetSessionStaged(db, &p->sessionStaged);
-  doltliteSetSessionMergeState(db, p->sessionIsMerging,
-                               &p->sessionMergeCommit,
-                               &p->sessionConflictsCatalog);
-  doltliteSetSessionConstraintViolationsCatalog(
-      db, &p->sessionConstraintViolationsCatalog);
-  return SQLITE_OK;
-}
-
-static int doltliteRestoreTxnStateOnFailure(
-  sqlite3 *db,
-  DoltliteTxnState *pSaved,
-  int opRc
-){
-  int rc = doltliteRestoreTxnState(db, pSaved);
-  doltliteTxnStateClear(pSaved);
-  return rc==SQLITE_OK ? opRc : rc;
-}
-
-static int doltliteRefreshAndConfirmHead(
-  sqlite3 *db,
-  ChunkStore *cs,
-  const ProllyHash *pExpectedHead
-){
-  const char *zBranch;
-  ProllyHash branchTip;
-  int found = 0;
-  int rc;
-
-  rc = chunkStoreLockAndRefresh(cs);
-  if( rc!=SQLITE_OK ) return rc;
-
-  /* Refresh in-memory state so a subsequent advance builds on the current
-  ** view. This is a no-op when we hold the lock reentrantly (a VC op already
-  ** in a write transaction), so it cannot be the basis of the CAS below. */
-  rc = chunkStoreForceRefresh(cs);
-  if( rc!=SQLITE_OK ){
-    chunkStoreUnlock(cs);
-    return rc;
-  }
-
-  /* Compare against the authoritative on-disk tip, read directly rather than
-  ** from the in-memory refs. The in-memory tip can be stale here: force-refresh
-  ** above is suppressed under a reentrant lock, and even the lock-time heuristic
-  ** can miss a peer commit when WAL reuse leaves the file size unchanged. A
-  ** stale tip would let this advance clobber the peer's commit (lost update). */
-  zBranch = doltliteGetSessionBranch(db);
-  rc = chunkStoreReadDiskBranchTip(cs, zBranch, &branchTip, &found);
-  if( rc!=SQLITE_OK ){
-    chunkStoreUnlock(cs);
-    return rc;
-  }
-  if( found && prollyHashCompare(&branchTip, pExpectedHead)!=0 ){
-    chunkStoreUnlock(cs);
-    return SQLITE_BUSY;
-  }
-  return SQLITE_OK;
-}
-
-int doltliteHasUncommittedChanges(sqlite3 *db){
-  ProllyHash headCatHash, stagedHash, workingCatHash;
-  u8 *wCatData = 0; int nWCat = 0;
-  int rc;
-
-  rc = doltliteGetHeadCatalogHash(db, &headCatHash);
-  if( rc!=SQLITE_OK ) return 0;
-  if( prollyHashIsEmpty(&headCatHash) ){
-    sqlite3_stmt *pStmt = 0;
-    int hasUserTables = 0;
-    rc = sqlite3_prepare_v2(db,
-      "SELECT 1 FROM sqlite_master "
-      "WHERE type='table' "
-      "AND name NOT LIKE 'sqlite_%' "
-      "AND name NOT LIKE 'dolt_%' "
-      "LIMIT 1",
-      -1, &pStmt, 0);
-    if( rc==SQLITE_OK && sqlite3_step(pStmt)==SQLITE_ROW ){
-      hasUserTables = 1;
-    }
-    sqlite3_finalize(pStmt);
-    return hasUserTables;
-  }
-
-  doltliteGetSessionStaged(db, &stagedHash);
-  if( !prollyHashIsEmpty(&stagedHash)
-   && prollyHashCompare(&headCatHash, &stagedHash)!=0 ){
-    return 1;
-  }
-
-  {
-    ChunkStore *cs = doltliteGetChunkStore(db);
-    if( cs ){
-      rc = doltliteFlushAndSerializeCatalog(db, &wCatData, &nWCat);
-      if( rc==SQLITE_OK ){
-        rc = chunkStorePut(cs, wCatData, nWCat, &workingCatHash);
-        sqlite3_free(wCatData);
-        if( rc==SQLITE_OK && prollyHashCompare(&headCatHash, &workingCatHash)!=0 ){
-          return 1;
-        }
-      }
-    }
-    return 0;
-  }
-}
-
-void doltliteUpdateSchemaHashes(sqlite3 *db){
-  sqlite3_stmt *pStmt = 0;
-  /* One scan of sqlite_master covers every table and index; both key their
-  ** catalog entry by rootpage and canonicalize by their own name. */
-  if( sqlite3_prepare_v2(
-        db,
-        "SELECT name, rootpage, sql "
-        "FROM main.sqlite_master "
-        "WHERE type IN ('table','index') AND sql IS NOT NULL",
-        -1, &pStmt, 0
-      )==SQLITE_OK ){
-    while( sqlite3_step(pStmt)==SQLITE_ROW ){
-      const char *zName = (const char*)sqlite3_column_text(pStmt, 0);
-      Pgno iRoot = (Pgno)sqlite3_column_int(pStmt, 1);
-      const char *zCreate = (const char*)sqlite3_column_text(pStmt, 2);
-      if( zName && zCreate ){
-        ProllyHash h;
-        char *zCanon = doltliteCanonicalizeSchemaSql(zCreate, zName);
-        if( zCanon ){
-          prollyHashCompute(zCanon, (int)strlen(zCanon), &h);
-          sqlite3_free(zCanon);
-          doltliteSetTableSchemaHash(db, iRoot, &h);
-        }
-      }
-    }
-    sqlite3_finalize(pStmt);
-  }
-}
-
-int doltliteLoadLiveSchemaSql(
-  sqlite3 *db,
-  const char *zType,
-  const char *zDb,
-  const char *zName,
-  const char *zTblName,
-  char **pzSql
-){
-  sqlite3_stmt *pStmt = 0;
-  char *zQuery = 0;
-  int rc;
-
-  *pzSql = 0;
-  if( !db || !zType || !zName ) return SQLITE_OK;
-
-  if( zTblName && zTblName[0] ){
-    zQuery = sqlite3_mprintf(
-      "SELECT sql FROM \"%w\".sqlite_master "
-      "WHERE type=%Q AND name=%Q AND tbl_name=%Q",
-      zDb ? zDb : "main", zType, zName, zTblName
-    );
-  }else{
-    zQuery = sqlite3_mprintf(
-      "SELECT sql FROM \"%w\".sqlite_master "
-      "WHERE type=%Q AND name=%Q",
-      zDb ? zDb : "main", zType, zName
-    );
-  }
-  if( !zQuery ) return SQLITE_NOMEM;
-
-  rc = sqlite3_prepare_v2(db, zQuery, -1, &pStmt, 0);
-  sqlite3_free(zQuery);
-  if( rc!=SQLITE_OK ) return rc;
-
-  rc = sqlite3_step(pStmt);
-  if( rc==SQLITE_ROW ){
-    const char *zSql = (const char*)sqlite3_column_text(pStmt, 0);
-    if( zSql ){
-      *pzSql = sqlite3_mprintf("%s", zSql);
-      if( !*pzSql ){
-        sqlite3_finalize(pStmt);
-        return SQLITE_NOMEM;
-      }
-    }
-    rc = SQLITE_OK;
-  }else if( rc==SQLITE_DONE ){
-    rc = SQLITE_OK;
-  }
-  sqlite3_finalize(pStmt);
-  return rc;
-}
-
-int doltliteMutateRefs(sqlite3 *db, DoltliteRefsMutation xMutate, void *pArg){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  int rc;
-
-  if( !cs ) return SQLITE_ERROR;
-
-  rc = chunkStoreLockAndRefresh(cs);
-  if( rc!=SQLITE_OK ) return rc;
-
-  /* xMutate edits the in-memory refs and serializeRefs rewrites the whole refs
-  ** blob, so a stale view would drop a peer's concurrent ref change (e.g. a
-  ** branch delete clobbering main's just-merged advance). Reload persisted
-  ** refs first rather than trusting the lock-time change heuristic. */
-  rc = chunkStoreForceRefresh(cs);
-  if( rc!=SQLITE_OK ){
-    chunkStoreUnlock(cs);
-    return rc;
-  }
-
-  rc = xMutate(db, cs, pArg);
-  if( rc==SQLITE_OK ){
-    rc = chunkStoreSerializeRefs(cs);
-    if( rc==SQLITE_OK ) rc = chunkStoreCommit(cs);
-  }
-
-  chunkStoreUnlock(cs);
-  return rc;
-}
-
-int doltliteFlushCatalogToHash(sqlite3 *db, ProllyHash *pHash){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  u8 *catData = 0;
-  int nCatData = 0;
-  int rc;
-  rc = doltliteFlushAndSerializeCatalog(db, &catData, &nCatData);
-  if( rc!=SQLITE_OK ) return rc;
-  rc = chunkStorePut(cs, catData, nCatData, pHash);
-  sqlite3_free(catData);
-  return rc;
-}
-
-static int doltlitePrepareCatalogForPersistence(sqlite3 *db){
-  UNUSED_PARAMETER(db);
-  return SQLITE_OK;
-}
-
-void freeSchemaMergeActions(SchemaMergeAction *a, int n){
-  int i, j;
-  for(i=0; i<n; i++){
-    for(j=0; j<a[i].nAddColumns; j++){
-      sqlite3_free(a[i].azAddColumns[j]);
-    }
-    sqlite3_free(a[i].azAddColumns);
-    sqlite3_free(a[i].zTableName);
-  }
-  sqlite3_free(a);
-}
-
-static int doltliteCreateAndStoreCommitWithTime(
-  sqlite3 *db,
-  const ProllyHash *pParent,
-  const ProllyHash *pCatalog,
-  const char *zMessage,
-  const char *zAuthorName,
-  const char *zAuthorEmail,
-  const ProllyHash *aExtraParents,
-  int nExtraParents,
-  i64 explicitTimestamp,
-  ProllyHash *pCommitHash
-);
-
-static int doltliteCreateAndStoreCommit(
-  sqlite3 *db,
-  const ProllyHash *pParent,
-  const ProllyHash *pCatalog,
-  const char *zMessage,
-  const char *zAuthorName,
-  const char *zAuthorEmail,
-  const ProllyHash *aExtraParents,
-  int nExtraParents,
-  ProllyHash *pCommitHash
-){
-  return doltliteCreateAndStoreCommitWithTime(db, pParent, pCatalog, zMessage,
-      zAuthorName, zAuthorEmail, aExtraParents, nExtraParents, 0, pCommitHash);
-}
-
-static int doltliteCreateAndStoreCommitWithTime(
-  sqlite3 *db,
-  const ProllyHash *pParent,
-  const ProllyHash *pCatalog,
-  const char *zMessage,
-  const char *zAuthorName,
-  const char *zAuthorEmail,
-  const ProllyHash *aExtraParents,
-  int nExtraParents,
-  i64 explicitTimestamp,
-  ProllyHash *pCommitHash
-){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  DoltliteCommit c;
-  u8 *commitData = 0;
-  int nCommitData = 0;
-  int rc, i;
-
-  memset(&c, 0, sizeof(c));
-  memcpy(&c.parentHash, pParent, sizeof(ProllyHash));
-  memcpy(&c.catalogHash, pCatalog, sizeof(ProllyHash));
-  c.timestamp = explicitTimestamp ? explicitTimestamp : (i64)time(0);
-  c.zName  = sqlite3_mprintf("%s", zAuthorName  ? zAuthorName  : doltliteGetAuthorName(db));
-  c.zEmail = sqlite3_mprintf("%s", zAuthorEmail ? zAuthorEmail : doltliteGetAuthorEmail(db));
-  c.zMessage = sqlite3_mprintf("%s", zMessage);
-
-  if( nExtraParents > 0 && aExtraParents ){
-    c.aParents[0] = *pParent;
-    for(i=0; i<nExtraParents && (i+1)<DOLTLITE_MAX_PARENTS; i++){
-      c.aParents[i+1] = aExtraParents[i];
-    }
-    c.nParents = 1 + (nExtraParents < DOLTLITE_MAX_PARENTS-1
-                       ? nExtraParents : DOLTLITE_MAX_PARENTS-1);
-  }
-
-  rc = doltliteCommitSerialize(&c, &commitData, &nCommitData);
-  if( rc==SQLITE_OK ) rc = chunkStorePut(cs, commitData, nCommitData, pCommitHash);
-  sqlite3_free(commitData);
-  doltliteCommitClear(&c);
-  return rc;
-}
-
-static int doltliteAdvanceBranch(
-  sqlite3 *db,
-  const ProllyHash *pNewHead,
-  const ProllyHash *pCatalogHash,
-  const ProllyHash *pWorkingCatHash
-){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  const char *branch = doltliteGetSessionBranch(db);
-  DoltliteTxnState saved;
-  int rc;
-
-  rc = doltliteSaveTxnState(db, &saved);
-  if( rc!=SQLITE_OK ) return rc;
-
-  if( refsTableBranchCount(&cs->refs)==0 ){
-    rc = chunkStoreAddBranch(cs, branch, pNewHead);
-    if( rc==SQLITE_OK ){
-      rc = chunkStoreSetDefaultBranch(cs, branch);
-    }
-  }else{
-    rc = chunkStoreUpdateBranch(cs, branch, pNewHead);
-  }
-  if( rc!=SQLITE_OK ){
-    return doltliteRestoreTxnStateOnFailure(db, &saved, rc);
-  }
-
-  doltliteSetSessionHead(db, pNewHead);
-  doltliteSetSessionStaged(db, pCatalogHash);
-  if( pWorkingCatHash && !prollyHashIsEmpty(pWorkingCatHash) ){
-    rc = doltliteSwitchCatalog(db, pWorkingCatHash);
-  }else{
-    rc = doltliteSwitchCatalog(db, pCatalogHash);
-  }
-  if( rc!=SQLITE_OK ){
-    return doltliteRestoreTxnStateOnFailure(db, &saved, rc);
-  }
-
-  rc = doltlitePersistWorkingSetWithHash(db, pWorkingCatHash);
-  if( rc!=SQLITE_OK ){
-    return doltliteRestoreTxnStateOnFailure(db, &saved, rc);
-  }
-
-  doltliteTxnStateClear(&saved);
-  return SQLITE_OK;
-}
-
-static int doltlitePersistOrSaveWorkingSet(sqlite3 *db){
-  if( doltliteVcTxnMode(db)==DOLTLITE_VC_TXN_AUTOCOMMIT_LIKE ){
-    return doltlitePersistWorkingSet(db);
-  }
-  return doltliteSaveWorkingSet(db);
-}
-
-static int doltliteReportConflicts(
-  sqlite3 *db,
-  sqlite3_context *ctx,
-  int nConflicts,
-  const char *zOp
-){
-  char msg[256];
-  int rc;
-  rc = doltliteRegisterConflictTables(db);
-  if( rc!=SQLITE_OK ) return rc;
-  rc = doltlitePersistOrSaveWorkingSet(db);
-  if( rc!=SQLITE_OK ) return rc;
-  sqlite3_snprintf(sizeof(msg), msg,
-    "%s has %d conflict(s). Resolve and then commit with dolt_commit.",
-    zOp, nConflicts);
-  sqlite3_result_error(ctx, msg, -1);
-  return SQLITE_OK;
-}
-
-static int doltliteReportConstraintViolations(
-  sqlite3 *db,
-  sqlite3_context *ctx,
-  const char *zOp
-){
-  char msg[256];
-  int rc;
-  rc = doltlitePersistOrSaveWorkingSet(db);
-  if( rc!=SQLITE_OK ) return rc;
-  sqlite3_snprintf(sizeof(msg), msg,
-    "%s resulted in constraint violations. Resolve the rows in "
-    "dolt_constraint_violations and then commit with dolt_commit.",
-    zOp);
-  sqlite3_result_error(ctx, msg, -1);
-  return SQLITE_OK;
-}
-
-static int doltliteDetectPostMergeConstraintViolations(
-  sqlite3 *db,
-  const ProllyHash *pAncCatHash,
-  int *pnViolations
-){
-  int nViolations = 0;
-  int nUnique = 0;
-  int nCheck = 0;
-  char *zDetectErrMsg = 0;
-  int rc;
-  sqlite3_stmt *pStmt = 0;
-  int needsDetection = 0;
-
-  rc = sqlite3_prepare_v2(db,
-      "SELECT 1 "
-      "FROM main.sqlite_master "
-      "WHERE (type='table' AND name NOT LIKE 'sqlite_%' AND name NOT LIKE 'dolt_%' "
-      "       AND sql IS NOT NULL "
-      "       AND (instr(upper(sql), 'REFERENCES')>0 "
-      "            OR instr(upper(sql), 'CHECK')>0 "
-      "            OR instr(upper(sql), 'UNIQUE')>0)) "
-      "   OR (type='index' AND sql IS NOT NULL "
-      "       AND instr(upper(sql), 'CREATE UNIQUE INDEX')>0) "
-      "LIMIT 1",
-      -1, &pStmt, 0);
-  if( rc!=SQLITE_OK ) return rc;
-  rc = sqlite3_step(pStmt);
-  if( rc==SQLITE_ROW ){
-    needsDetection = 1;
-    rc = SQLITE_OK;
-  }else if( rc==SQLITE_DONE ){
-    rc = SQLITE_OK;
-  }
-  sqlite3_finalize(pStmt);
-  if( rc!=SQLITE_OK ) return rc;
-  if( !needsDetection ){
-    if( pnViolations ) *pnViolations = 0;
-    return SQLITE_OK;
-  }
-
-  rc = doltliteConstraintViolationBatchBegin(db);
-  if( rc==SQLITE_OK ){
-    rc = doltliteDetectMergeFkViolations(db, pAncCatHash,
-                                         &zDetectErrMsg, &nViolations);
-  }
-  if( rc==SQLITE_OK ){
-    rc = doltliteDetectMergeUniqueViolations(db, pAncCatHash,
-                                             &zDetectErrMsg, &nUnique);
-  }
-  if( rc==SQLITE_OK ){
-    rc = doltliteDetectMergeCheckViolations(db, pAncCatHash,
-                                            &zDetectErrMsg, &nCheck);
-  }
-  {
-    int erc = doltliteConstraintViolationBatchEnd(db, rc==SQLITE_OK);
-    if( rc==SQLITE_OK ) rc = erc;
-  }
-  sqlite3_free(zDetectErrMsg);
-  if( rc!=SQLITE_OK ) return rc;
-
-  if( pnViolations ) *pnViolations = nViolations + nUnique + nCheck;
-  return SQLITE_OK;
-}
-
-static int doltliteSavepointIsTopLevelTxn(sqlite3 *db){
-  return db->pSavepoint!=0 && db->nSavepoint==0;
-}
-
-/* SQLite represents a top-level SAVEPOINT as the transaction boundary.
-** Doltlite treats that like autocommit for VC operations: seal the boundary
-** instead of leaving a savepoint that later ROLLBACK TO can undo. */
-static int doltliteVcSealTopLevelSavepointTxn(sqlite3 *db){
-  if( doltliteSavepointIsTopLevelTxn(db) ){
-    return sqlite3_exec(db, "COMMIT", 0, 0, 0);
-  }
-  return SQLITE_OK;
-}
-
-DoltliteVcTxnMode doltliteVcTxnMode(sqlite3 *db){
-  if( db->autoCommit || doltliteSavepointIsTopLevelTxn(db) ){
-    return DOLTLITE_VC_TXN_AUTOCOMMIT_LIKE;
-  }
-  if( db->pSavepoint ){
-    return DOLTLITE_VC_TXN_NESTED_SAVEPOINT;
-  }
-  return DOLTLITE_VC_TXN_PLAIN;
-}
-
-int doltliteVcSealActiveSavepoints(sqlite3 *db){
-  int rc = SQLITE_OK;
-  while( rc==SQLITE_OK && db->pSavepoint ){
-    char *zSql = sqlite3_mprintf("RELEASE SAVEPOINT \"%w\"", db->pSavepoint->zName);
-    if( !zSql ) return SQLITE_NOMEM;
-    rc = sqlite3_exec(db, zSql, 0, 0, 0);
-    sqlite3_free(zSql);
-  }
-  return rc;
-}
-
-int doltliteVcSealSavepointError(sqlite3 *db){
-  if( db->pSavepoint ){
-    return doltliteVcSealActiveSavepoints(db);
-  }
-  return SQLITE_OK;
-}
-
-void doltliteVcResultError(sqlite3_context *ctx, sqlite3 *db, const char *zMsg){
-  (void)doltliteVcSealSavepointError(db);
-  sqlite3_result_error(ctx, zMsg, -1);
-}
-
-int doltliteVcSealBranchStyleTxn(sqlite3 *db){
-  int rc;
-  if( db->autoCommit ) return SQLITE_OK;
-  if( db->pSavepoint ){
-    return doltliteVcSealActiveSavepoints(db);
-  }
-  rc = sqlite3_exec(db, "COMMIT", 0, 0, 0);
-  if( rc!=SQLITE_OK ) return rc;
-  return sqlite3_exec(db, "BEGIN", 0, 0, 0);
-}
-
-static int doltliteVcSealBranchStyleTxnMaybeKeepTopLevelSavepoint(sqlite3 *db){
-  if( doltliteSavepointIsTopLevelTxn(db) ){
-    return SQLITE_OK;
-  }
-  return doltliteVcSealBranchStyleTxn(db);
-}
-
-static int rebaseRestoreReturnBranchWorkingState(
-  sqlite3 *db,
-  const char *zBranch
-){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  DoltliteCommit c;
-  ProllyHash headHash;
-  int rc;
-
-  if( !cs || !zBranch || !zBranch[0] ) return SQLITE_OK;
-  rc = chunkStoreFindBranch(cs, zBranch, &headHash);
-  if( rc!=SQLITE_OK ) return rc;
-  memset(&c, 0, sizeof(c));
-  rc = doltliteLoadCommit(db, &headHash, &c);
-  if( rc==SQLITE_OK ){
-    rc = doltliteWriteBranchCleanWorkingState(db, zBranch, &c.catalogHash, &headHash);
-  }
-  doltliteCommitClear(&c);
-  return rc;
-}
-
-static int doltlitePrimeSchemaCache(sqlite3 *db){
-  sqlite3_stmt *pStmt = 0;
-  int rc = sqlite3_prepare_v2(
-      db, "SELECT name FROM main.sqlite_master LIMIT 1", -1, &pStmt, 0);
-  if( rc!=SQLITE_OK ) return rc;
-  while( (rc = sqlite3_step(pStmt))==SQLITE_ROW ){}
-  if( rc==SQLITE_DONE ) rc = SQLITE_OK;
-  sqlite3_finalize(pStmt);
-  return rc;
-}
-
-static void doltliteReportAutocommitConflictRollback(sqlite3_context *ctx){
-  sqlite3_result_error(ctx,
-    "Merge conflict detected, @autocommit transaction rolled back. "
-    "@autocommit must be disabled so that merge conflicts can be "
-    "resolved using the dolt_conflicts and dolt_schema_conflicts "
-    "tables before manually committing the transaction. "
-    "Alternatively, to commit transactions with merge conflicts, set "
-    "@@dolt_allow_commit_conflicts = 1",
-    -1);
-}
-
-static int doltliteRollbackAutocommitConflict(
-  sqlite3 *db,
-  sqlite3_context *ctx,
-  DoltliteTxnState *pSaved
-){
-  int rc;
-  int hadTopLevelSavepoint = db->pSavepoint!=0 && db->nSavepoint==0;
-  sqlite3RollbackAll(db, SQLITE_OK);
-  rc = doltliteRestoreTxnState(db, pSaved);
-  if( rc==SQLITE_OK ){
-    doltliteSetSessionMergeState(db, pSaved->sessionIsMerging,
-                                 &pSaved->sessionMergeCommit,
-                                 &pSaved->sessionConflictsCatalog);
-    doltliteSetSessionConstraintViolationsCatalog(
-        db, &pSaved->sessionConstraintViolationsCatalog);
-  }
-  doltliteTxnStateClear(pSaved);
-  if( rc==SQLITE_OK && hadTopLevelSavepoint ){
-    rc = doltliteVcSealTopLevelSavepointTxn(db);
-  }
-  if( rc==SQLITE_OK ){
-    doltliteReportAutocommitConflictRollback(ctx);
-  }
-  return rc;
-}
-
-static void addFreeEntries(
-  struct TableEntry *aWorking, int nWorking,
-  struct TableEntry *aStaged,  int nStaged,
-  struct TableEntry *aNew,     int nNew
-){
-  doltliteFreeCatalog(aWorking, nWorking);
-  doltliteFreeCatalog(aStaged, nStaged);
-  doltliteFreeCatalog(aNew, nNew);
-}
-
-static void addResultIgnoreConflict(sqlite3_context *context, char *zIgnErr){
-  if( zIgnErr ){
-    sqlite3_result_error(context, zIgnErr, -1);
-    sqlite3_free(zIgnErr);
-  }else{
-    sqlite3_result_error(context, "dolt_ignore conflict", -1);
-  }
-}
-
-static int addCheckIgnore(
-  sqlite3 *db,
-  sqlite3_context *context,
-  const char *zName,
-  int *pIgnored
-){
-  char *zIgnErr = 0;
-  int rc;
-  *pIgnored = 0;
-  rc = doltliteCheckIgnore(db, zName, pIgnored, &zIgnErr);
-  if( rc==SQLITE_CONSTRAINT ){
-    addResultIgnoreConflict(context, zIgnErr);
-  }else if( rc!=SQLITE_OK ){
-    sqlite3_free(zIgnErr);
-    sqlite3_result_error_code(context, rc);
-  }
-  return rc;
-}
-
-static int addAppendTableEntry(
-  sqlite3_context *context,
-  struct TableEntry **paEntries,
-  int *pnEntries,
-  const struct TableEntry *pEntry
-){
-  struct TableEntry *aNew;
-  char *zDup = 0;
-  if( pEntry->zName ){
-    zDup = sqlite3_mprintf("%s", pEntry->zName);
-    if( !zDup ){
-      if( context ) sqlite3_result_error_nomem(context);
-      return SQLITE_NOMEM;
-    }
-  }
-  aNew = sqlite3_realloc(
-      *paEntries, (*pnEntries + 1) * (int)sizeof(struct TableEntry));
-  if( !aNew ){
-    sqlite3_free(zDup);
-    if( context ) sqlite3_result_error_nomem(context);
-    return SQLITE_NOMEM;
-  }
-  *paEntries = aNew;
-  (*paEntries)[*pnEntries] = *pEntry;
-  (*paEntries)[*pnEntries].zName = zDup;
-  (*pnEntries)++;
-  return SQLITE_OK;
-}
-
-static struct TableEntry *addFindEntryByName(
-  struct TableEntry *aEntries,
-  int nEntries,
-  const char *zName
-){
-  int i;
-  if( !zName ) return 0;
-  for(i=0; i<nEntries; i++){
-    if( aEntries[i].zName && strcmp(aEntries[i].zName, zName)==0 ){
-      return &aEntries[i];
-    }
-  }
-  return 0;
-}
-
-/* Catalog name->entry index over the shared DoltliteNameIndex, which reads
-** names through the live array so it survives dolt_commit -A rewriting entry
-** names in place. */
-typedef DoltliteNameIndex AddNameIndex;
-
-static int addNameIndexInit(
-  AddNameIndex *pIdx,
-  struct TableEntry *aEntry,
-  int nEntry
-){
-  return doltliteNameIndexInit(pIdx, aEntry, nEntry,
-                               (int)sizeof(struct TableEntry),
-                               (int)offsetof(struct TableEntry, zName));
-}
-
-static void addNameIndexFree(AddNameIndex *pIdx){
-  doltliteNameIndexFree(pIdx);
-}
-
-static struct TableEntry *addNameIndexFind(
-  const AddNameIndex *pIdx,
-  const char *zName
-){
-  int r = doltliteNameIndexFind(pIdx, zName);
-  return r<0 ? 0 : (struct TableEntry*)(pIdx->aBase + (size_t)r*pIdx->stride);
-}
-
-static void addAlignStagedEntriesToWorking(
-  struct TableEntry *aWorking,
-  int nWorking,
-  struct TableEntry *aStaged,
-  int nStaged
-){
-  int i;
-  AddNameIndex workingIdx;
-  if( addNameIndexInit(&workingIdx, aWorking, nWorking)!=SQLITE_OK ){
-    for(i=0; i<nStaged; i++){
-      struct TableEntry *pWorking;
-      if( !aStaged[i].zName ) continue;
-      pWorking = addFindEntryByName(aWorking, nWorking, aStaged[i].zName);
-      if( pWorking ){
-        aStaged[i].iTable = pWorking->iTable;
-      }
-    }
-    return;
-  }
-  for(i=0; i<nStaged; i++){
-    struct TableEntry *pWorking;
-    if( !aStaged[i].zName ) continue;
-    pWorking = addNameIndexFind(&workingIdx, aStaged[i].zName);
-    if( pWorking ){
-      aStaged[i].iTable = pWorking->iTable;
-    }
-  }
-  addNameIndexFree(&workingIdx);
-}
-
-static int addLoadWorkingAndStagedCatalogs(
-  sqlite3 *db,
-  const ProllyHash *pWorkingHash,
-  struct TableEntry **paWorking,
-  int *pnWorking,
-  struct TableEntry **paStaged,
-  int *pnStaged
-){
-  ProllyHash stagedHash;
-  int rc;
-
-  *paWorking = 0;
-  *pnWorking = 0;
-  *paStaged = 0;
-  *pnStaged = 0;
-
-  rc = doltliteLoadCatalog(db, pWorkingHash, paWorking, pnWorking, 0);
-  if( rc!=SQLITE_OK ) return rc;
-
-  doltliteGetSessionStaged(db, &stagedHash);
-  if( prollyHashIsEmpty(&stagedHash) ){
-    ProllyHash headCat;
-    rc = doltliteGetHeadCatalogHash(db, &headCat);
-    if( rc==SQLITE_OK && !prollyHashIsEmpty(&headCat) ){
-      rc = doltliteLoadCatalog(db, &headCat, paStaged, pnStaged, 0);
-    }
-  }else{
-    rc = doltliteLoadCatalog(db, &stagedHash, paStaged, pnStaged, 0);
-  }
-  if( rc!=SQLITE_OK ){
-    doltliteFreeCatalog(*paWorking, *pnWorking);
-    *paWorking = 0;
-    *pnWorking = 0;
-  }
-  return rc;
-}
-
-static int addWriteStagedCatalog(
-  sqlite3 *db,
-  ChunkStore *cs,
-  struct TableEntry *aEntries,
-  int nEntries
-){
-  u8 *buf = 0;
-  int nBuf = 0;
-  ProllyHash newStagedHash;
-  int rc = doltliteSerializeCatalogEntries(db, aEntries, nEntries, &buf, &nBuf);
-  if( rc==SQLITE_OK ){
-    rc = chunkStorePut(cs, buf, nBuf, &newStagedHash);
-  }
-  sqlite3_free(buf);
-  if( rc==SQLITE_OK ){
-    doltliteSetSessionStaged(db, &newStagedHash);
-  }
-  return rc;
-}
-
-/* True when the two catalogs' index schema rows for zTable differ: any
-** index added, dropped, re-targeted, or redefined. Index changes carry no
-** named catalog entry, so every surface that reports per-table change
-** (status, the diff summary) attributes them through this comparison. */
-int doltliteIndexSchemaRowsDifferForTable(
-  SchemaEntry *aA, int nA,
-  SchemaEntry *aB, int nB,
-  const char *zTable
-){
-  int i, j, nMatchA = 0, nBForTable = 0;
-  for(i=0; i<nA; i++){
-    int found = 0;
-    if( !aA[i].zType || strcmp(aA[i].zType, "index")!=0
-     || !aA[i].zTblName || strcmp(aA[i].zTblName, zTable)!=0 ){
-      continue;
-    }
-    for(j=0; j<nB; j++){
-      if( aB[j].zType && strcmp(aB[j].zType, "index")==0
-       && aB[j].zTblName && strcmp(aB[j].zTblName, zTable)==0
-       && aB[j].zName && aA[i].zName
-       && strcmp(aB[j].zName, aA[i].zName)==0
-       && ((aB[j].zSql==0)==(aA[i].zSql==0))
-       && (aB[j].zSql==0 || strcmp(aB[j].zSql, aA[i].zSql)==0) ){
-        found = 1;
-        break;
-      }
-    }
-    if( !found ) return 1;
-    nMatchA++;
-  }
-  for(j=0; j<nB; j++){
-    if( aB[j].zType && strcmp(aB[j].zType, "index")==0
-     && aB[j].zTblName && strcmp(aB[j].zTblName, zTable)==0 ){
-      nBForTable++;
-    }
-  }
-  return nMatchA!=nBForTable;
-}
-
-/* True when the final staged entry list contains a table entry named
-** zTbl. Index entries follow their parent table through -a staging, and
-** the parent's presence decides whether an index entry belongs at all. */
-static int amTableStagedByName(struct TableEntry *aStaged, int nStaged,
-                               const char *zTbl){
-  int i;
-  if( !zTbl ) return 0;
-  for(i=0; i<nStaged; i++){
-    if( aStaged[i].zName && strcmp(aStaged[i].zName, zTbl)==0 ) return 1;
-  }
-  return 0;
-}
-
-static int addStageAllTables(
-  sqlite3 *db,
-  sqlite3_context *context,
-  ChunkStore *cs,
-  const ProllyHash *pWorkingHash
-){
-  struct TableEntry *aWorking = 0;
-  struct TableEntry *aStaged = 0;
-  struct TableEntry *aNew = 0;
-  int nWorking = 0;
-  int nStaged = 0;
-  int nNew = 0;
-  int k;
-  int useWorkingHash = 1;
-  int rc;
-  AddNameIndex stagedIdx;
-  AddNameIndex workingIdx;
-  int stagedIdxInit = 0;
-  int workingIdxInit = 0;
-
-  rc = addLoadWorkingAndStagedCatalogs(db, pWorkingHash,
-                                       &aWorking, &nWorking,
-                                       &aStaged, &nStaged);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error(context, "failed to load staged catalog", -1);
-    return rc;
-  }
-  rc = addNameIndexInit(&stagedIdx, aStaged, nStaged);
-  if( rc!=SQLITE_OK ){
-    addFreeEntries(aWorking, nWorking, aStaged, nStaged, aNew, nNew);
-    sqlite3_result_error_nomem(context);
-    return rc;
-  }
-  stagedIdxInit = 1;
-  rc = addNameIndexInit(&workingIdx, aWorking, nWorking);
-  if( rc!=SQLITE_OK ){
-    addNameIndexFree(&stagedIdx);
-    addFreeEntries(aWorking, nWorking, aStaged, nStaged, aNew, nNew);
-    sqlite3_result_error_nomem(context);
-    return rc;
-  }
-  workingIdxInit = 1;
-
-  for(k=0; k<nWorking; k++){
-    const char *zName = aWorking[k].zName;
-    struct TableEntry *pUse = &aWorking[k];
-    if( aWorking[k].iTable>1 && zName ){
-      int ignored = 0;
-      rc = addCheckIgnore(db, context, zName, &ignored);
-      if( rc!=SQLITE_OK ){
-        if( workingIdxInit ) addNameIndexFree(&workingIdx);
-        if( stagedIdxInit ) addNameIndexFree(&stagedIdx);
-        addFreeEntries(aWorking, nWorking, aStaged, nStaged, aNew, nNew);
-        return rc;
-      }
-      if( ignored ){
-        useWorkingHash = 0;
-        pUse = addNameIndexFind(&stagedIdx, zName);
-        if( !pUse ) continue;
-      }
-    }
-    rc = addAppendTableEntry(context, &aNew, &nNew, pUse);
-    if( rc!=SQLITE_OK ){
-      if( workingIdxInit ) addNameIndexFree(&workingIdx);
-      if( stagedIdxInit ) addNameIndexFree(&stagedIdx);
-      addFreeEntries(aWorking, nWorking, aStaged, nStaged, aNew, nNew);
-      return rc;
-    }
-  }
-
-  for(k=0; k<nStaged; k++){
-    const char *zName = aStaged[k].zName;
-    if( aStaged[k].iTable<=1 || !zName ) continue;
-    if( addNameIndexFind(&workingIdx, zName) ) continue;
-    {
-      int ignored = 0;
-      rc = addCheckIgnore(db, context, zName, &ignored);
-      if( rc!=SQLITE_OK ){
-        if( workingIdxInit ) addNameIndexFree(&workingIdx);
-        if( stagedIdxInit ) addNameIndexFree(&stagedIdx);
-        addFreeEntries(aWorking, nWorking, aStaged, nStaged, aNew, nNew);
-        return rc;
-      }
-      if( !ignored ) continue;
-    }
-    useWorkingHash = 0;
-    rc = addAppendTableEntry(context, &aNew, &nNew, &aStaged[k]);
-    if( rc!=SQLITE_OK ){
-      if( workingIdxInit ) addNameIndexFree(&workingIdx);
-      if( stagedIdxInit ) addNameIndexFree(&stagedIdx);
-      addFreeEntries(aWorking, nWorking, aStaged, nStaged, aNew, nNew);
-      return rc;
-    }
-  }
-
-  if( useWorkingHash ){
-    doltliteSetSessionStaged(db, pWorkingHash);
-    rc = SQLITE_OK;
-  }else{
-    addAlignStagedEntriesToWorking(aWorking, nWorking, aNew, nNew);
-    rc = addWriteStagedCatalog(db, cs, aNew, nNew);
-  }
-  if( workingIdxInit ) addNameIndexFree(&workingIdx);
-  if( stagedIdxInit ) addNameIndexFree(&stagedIdx);
-  addFreeEntries(aWorking, nWorking, aStaged, nStaged, aNew, nNew);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-  }
-  return rc;
-}
-
-/* Remove unnamed (index) staged entries whose parent table is zTable,
-** resolving parents through the staged catalog's own schema rows — entry
-** numbers do not carry across catalogs. */
-static void addRemoveIndexEntriesOfTable(
-  struct TableEntry *aStaged,
-  int *pnStaged,
-  SchemaEntry *aStagedSchema,
-  int nStagedSchema,
-  const char *zTable
-){
-  int j, i;
-  for(j=0; j<*pnStaged; ){
-    const char *zParent = 0;
-    if( aStaged[j].iTable<=1 || aStaged[j].zName ){ j++; continue; }
-    for(i=0; i<nStagedSchema; i++){
-      if( aStagedSchema[i].zType
-       && strcmp(aStagedSchema[i].zType, "index")==0
-       && aStagedSchema[i].iRootpage==aStaged[j].iTable ){
-        zParent = aStagedSchema[i].zTblName;
-        break;
-      }
-    }
-    if( !zParent || strcmp(zParent, zTable)!=0 ){ j++; continue; }
-    sqlite3_free(aStaged[j].zName);
-    if( j+1<*pnStaged ){
-      memmove(&aStaged[j], &aStaged[j+1],
-              (*pnStaged-j-1)*(int)sizeof(struct TableEntry));
-    }
-    (*pnStaged)--;
-  }
-}
-
-/* Stage zTable's index entries from the working catalog alongside its
-** table entry: an index travels with its table through staging. */
-static int addAppendIndexEntriesOfTable(
-  sqlite3_context *context,
-  struct TableEntry **paStaged,
-  int *pnStaged,
-  struct TableEntry *aWorking,
-  int nWorking,
-  SchemaEntry *aWorkSchema,
-  int nWorkSchema,
-  const char *zTable
-){
-  int i, j, rc;
-  for(i=0; i<nWorkSchema; i++){
-    if( !aWorkSchema[i].zType
-     || strcmp(aWorkSchema[i].zType, "index")!=0
-     || aWorkSchema[i].iRootpage<=1
-     || !aWorkSchema[i].zTblName
-     || strcmp(aWorkSchema[i].zTblName, zTable)!=0 ){
-      continue;
-    }
-    for(j=0; j<nWorking; j++){
-      if( aWorking[j].iTable==aWorkSchema[i].iRootpage
-       && aWorking[j].zName==0 ){
-        rc = addAppendTableEntry(context, paStaged, pnStaged, &aWorking[j]);
-        if( rc!=SQLITE_OK ) return rc;
-        break;
-      }
-    }
-  }
-  return SQLITE_OK;
-}
-
-/* Virtual tables persist through their shadow tables: staging a vtab by
-** name must carry the shadows (and their indexes) along, exactly as index
-** entries travel with ordinary tables. */
-static int addStageShadowTablesOf(
-  sqlite3 *db,
-  sqlite3_context *context,
-  struct TableEntry **paStaged,
-  int *pnStaged,
-  struct TableEntry *aWorking,
-  int nWorking,
-  SchemaEntry *aWorkSchema,
-  int nWorkSchema,
-  SchemaEntry *aStagedSchema,
-  int nStagedSchema,
-  const char *zTable
-){
-  Table *pTab;
-  int i, k, rc;
-
-  pTab = sqlite3FindTable(db, zTable, "main");
-  if( !pTab || !IsVirtual(pTab) ) return SQLITE_OK;
-
-  for(i=0; i<nWorkSchema; i++){
-    struct TableEntry *pWork;
-    int updated = 0;
-    if( !aWorkSchema[i].zType
-     || strcmp(aWorkSchema[i].zType, "table")!=0
-     || !aWorkSchema[i].zName
-     || strcmp(aWorkSchema[i].zName, zTable)==0
-     || !sqlite3IsShadowTableOf(db, pTab, aWorkSchema[i].zName) ){
-      continue;
-    }
-    pWork = doltliteFindTableByName(aWorking, nWorking, aWorkSchema[i].zName);
-    if( !pWork ) continue;
-    for(k=0; k<*pnStaged; k++){
-      if( (*paStaged)[k].zName
-       && strcmp((*paStaged)[k].zName, pWork->zName)==0 ){
-        char *zDup = sqlite3_mprintf("%s", pWork->zName);
-        if( !zDup ) return SQLITE_NOMEM;
-        sqlite3_free((*paStaged)[k].zName);
-        (*paStaged)[k] = *pWork;
-        (*paStaged)[k].zName = zDup;
-        updated = 1;
-        break;
-      }
-    }
-    if( !updated ){
-      rc = addAppendTableEntry(context, paStaged, pnStaged, pWork);
-      if( rc!=SQLITE_OK ) return rc;
-    }
-    addRemoveIndexEntriesOfTable(*paStaged, pnStaged,
-                                 aStagedSchema, nStagedSchema,
-                                 pWork->zName);
-    rc = addAppendIndexEntriesOfTable(context, paStaged, pnStaged,
-                                      aWorking, nWorking,
-                                      aWorkSchema, nWorkSchema,
-                                      pWork->zName);
-    if( rc!=SQLITE_OK ) return rc;
-  }
-  return SQLITE_OK;
-}
-
-/* Remove staged entries for shadow tables of a dropped virtual table. The
-** vtab is gone from the live schema, so shadows are identified through the
-** staged catalog's own rows: table rows prefixed "<zTable>_" whose parent
-** row is a CREATE VIRTUAL TABLE statement and which no longer exist in
-** working. */
-static void addRemoveShadowEntriesOfDroppedVtab(
-  struct TableEntry *aStaged,
-  int *pnStaged,
-  SchemaEntry *aStagedSchema,
-  int nStagedSchema,
-  struct TableEntry *aWorking,
-  int nWorking,
-  const char *zTable
-){
-  SchemaEntry *pParent;
-  size_t nPrefix;
-  int i, k;
-
-  pParent = findSchemaEntry(aStagedSchema, nStagedSchema, zTable);
-  if( !pParent || !pParent->zSql
-   || sqlite3_strnicmp(pParent->zSql, "CREATE VIRTUAL", 14)!=0 ){
-    return;
-  }
-  nPrefix = strlen(zTable);
-  for(i=0; i<nStagedSchema; i++){
-    if( !aStagedSchema[i].zType
-     || strcmp(aStagedSchema[i].zType, "table")!=0
-     || !aStagedSchema[i].zName
-     || sqlite3_strnicmp(aStagedSchema[i].zName, zTable, (int)nPrefix)!=0
-     || aStagedSchema[i].zName[nPrefix]!='_'
-     || doltliteFindTableByName(aWorking, nWorking, aStagedSchema[i].zName) ){
-      continue;
-    }
-    for(k=0; k<*pnStaged; ){
-      if( aStaged[k].zName
-       && strcmp(aStaged[k].zName, aStagedSchema[i].zName)==0 ){
-        sqlite3_free(aStaged[k].zName);
-        if( k+1<*pnStaged ){
-          memmove(&aStaged[k], &aStaged[k+1],
-                  (*pnStaged-k-1)*(int)sizeof(struct TableEntry));
-        }
-        (*pnStaged)--;
-        continue;
-      }
-      k++;
-    }
-  }
-}
-
-static int addStageNamedTables(
-  sqlite3 *db,
-  sqlite3_context *context,
-  ChunkStore *cs,
-  const ProllyHash *pWorkingHash,
-  int argc,
-  sqlite3_value **argv
-){
-  struct TableEntry *aWorking = 0;
-  struct TableEntry *aStaged = 0;
-  int nWorking = 0;
-  int nStaged = 0;
-  int i;
-  int updateMaster = 0;
-  int rc;
-
-  SchemaEntry *aWorkSchema = 0;
-  SchemaEntry *aStagedSchema = 0;
-  int nWorkSchema = 0;
-  int nStagedSchema = 0;
-  /* Objects whose master rows follow WORKING in this operation: the named
-  ** tables (adds and staged drops) plus a staged vtab's shadows. Names are
-  ** borrowed from the argv values and schema arrays, which outlive use. */
-  char **azTouched = 0;
-  int nTouched = 0;
-
-  #define ADDNAMED_FREE_ALL() do { \
-    int ft_; \
-    for(ft_=0; ft_<nTouched; ft_++) sqlite3_free(azTouched[ft_]); \
-    sqlite3_free((void*)azTouched); \
-    freeSchemaEntries(aWorkSchema, nWorkSchema); \
-    freeSchemaEntries(aStagedSchema, nStagedSchema); \
-    doltliteFreeCatalog(aWorking, nWorking); \
-    doltliteFreeCatalog(aStaged, nStaged); \
-  } while(0)
-
-  #define ADDNAMED_TOUCH(zN) do { \
-    char **azNew = sqlite3_realloc((void*)azTouched, \
-        (nTouched+1)*(int)sizeof(char*)); \
-    char *zOwn_ = azNew ? sqlite3_mprintf("%s", (zN)) : 0; \
-    if( azNew ) azTouched = azNew; \
-    if( !azNew || !zOwn_ ){ \
-      ADDNAMED_FREE_ALL(); \
-      sqlite3_result_error_nomem(context); \
-      return SQLITE_NOMEM; \
-    } \
-    azTouched[nTouched++] = zOwn_; \
-  } while(0)
-
-  rc = addLoadWorkingAndStagedCatalogs(db, pWorkingHash,
-                                       &aWorking, &nWorking,
-                                       &aStaged, &nStaged);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error(context, "failed to load staged catalog", -1);
-    return rc;
-  }
-
-  {
-    ProllyHash stagedSrc;
-    doltliteGetSessionStaged(db, &stagedSrc);
-    if( prollyHashIsEmpty(&stagedSrc) ){
-      (void)doltliteGetHeadCatalogHash(db, &stagedSrc);
-    }
-    rc = loadSchemaFromCatalog(db, cs, doltliteGetCache(db), pWorkingHash,
-                               &aWorkSchema, &nWorkSchema);
-    if( rc==SQLITE_OK && !prollyHashIsEmpty(&stagedSrc) ){
-      rc = loadSchemaFromCatalog(db, cs, doltliteGetCache(db), &stagedSrc,
-                                 &aStagedSchema, &nStagedSchema);
-    }
-    if( rc!=SQLITE_OK ){
-      ADDNAMED_FREE_ALL();
-      sqlite3_result_error(context, "failed to load schema for staging", -1);
-      return rc;
-    }
-  }
-
-  for(i=0; i<argc; i++){
-    const char *zTable = (const char*)sqlite3_value_text(argv[i]);
-    Pgno iTable = 0;
-    int j;
-    if( !zTable || zTable[0]=='-' || strcmp(zTable, ".")==0 ) continue;
-
-    {
-      int ignored = 0;
-      rc = addCheckIgnore(db, context, zTable, &ignored);
-      if( rc!=SQLITE_OK ){
-        ADDNAMED_FREE_ALL();
-        return rc;
-      }
-      if( ignored ) continue;
-    }
-
-    /* Virtual tables carry no catalog entry of their own (schema row
-    ** only), so the entry-based flow below cannot see them: their commit
-    ** content IS the shadow tables. Stage those, and adopt the working
-    ** master so the vtab's schema row travels too. */
-    {
-      Table *pLive = sqlite3FindTable(db, zTable, "main");
-      if( pLive && IsVirtual(pLive) ){
-        int w;
-        rc = addStageShadowTablesOf(db, context, &aStaged, &nStaged,
-                                    aWorking, nWorking,
-                                    aWorkSchema, nWorkSchema,
-                                    aStagedSchema, nStagedSchema, zTable);
-        if( rc!=SQLITE_OK ){
-          ADDNAMED_FREE_ALL();
-          return rc;
-        }
-        ADDNAMED_TOUCH(zTable);
-        for(w=0; w<nWorkSchema; w++){
-          if( aWorkSchema[w].zType
-           && strcmp(aWorkSchema[w].zType, "table")==0
-           && aWorkSchema[w].zName
-           && strcmp(aWorkSchema[w].zName, zTable)!=0
-           && sqlite3IsShadowTableOf(db, pLive, aWorkSchema[w].zName) ){
-            ADDNAMED_TOUCH(aWorkSchema[w].zName);
-          }
-        }
-        updateMaster = 1;
-        continue;
-      }
-    }
-
-    rc = doltliteResolveTableName(db, zTable, &iTable);
-    if( rc!=SQLITE_OK ){
-      int found = 0;
-      Pgno iDroppedTable = 0;
-      for(j=0; j<nStaged; j++){
-        if( aStaged[j].zName && strcmp(aStaged[j].zName, zTable)==0 ){
-          iDroppedTable = aStaged[j].iTable;
-          found = 1;
-          break;
-        }
-      }
-      if( !found && findSchemaEntry(aStagedSchema, nStagedSchema, zTable) ){
-        /* A dropped virtual table has a staged schema row but no entry;
-        ** the master adoption below stages the row's removal. */
-        found = 1;
-      }
-      if( found ){
-        for(j=0; j<nStaged; ){
-          int removeEntry = 0;
-          if( iDroppedTable!=0 && aStaged[j].iTable==iDroppedTable ){
-            removeEntry = 1;
-          }else if( aStaged[j].zName && strcmp(aStaged[j].zName, zTable)==0 ){
-            removeEntry = 1;
-          }
-          if( removeEntry ){
-            sqlite3_free(aStaged[j].zName);
-            if( j+1 < nStaged ){
-              memmove(&aStaged[j], &aStaged[j+1],
-                      (nStaged-j-1) * (int)sizeof(struct TableEntry));
-            }
-            nStaged--;
-            continue;
-          }
-          j++;
-        }
-        addRemoveIndexEntriesOfTable(aStaged, &nStaged,
-                                     aStagedSchema, nStagedSchema, zTable);
-        addRemoveShadowEntriesOfDroppedVtab(aStaged, &nStaged,
-                                            aStagedSchema, nStagedSchema,
-                                            aWorking, nWorking, zTable);
-        ADDNAMED_TOUCH(zTable);
-        updateMaster = 1;
-      }
-      if( !found ){
-        char *zErr = sqlite3_mprintf("table not found: %s", zTable);
-        ADDNAMED_FREE_ALL();
-        if( zErr ){
-          sqlite3_result_error(context, zErr, -1);
-          sqlite3_free(zErr);
-        }else{
-          sqlite3_result_error_nomem(context);
-        }
-        return SQLITE_ERROR;
-      }
-      continue;
-    }
-
-    for(j=0; j<nWorking; j++){
-      if( aWorking[j].iTable==iTable ){
-        int k;
-        int updated = 0;
-        for(k=0; k<nStaged; k++){
-          int nameMatch = aStaged[k].zName && aWorking[j].zName
-            && strcmp(aStaged[k].zName, aWorking[j].zName)==0;
-          int rootMatch = aStaged[k].iTable==iTable;
-          /* An unnamed staged entry on a bare number match can be a
-          ** cross-domain collision with an index entry; pair only when
-          ** the staged catalog's own schema rows say the entry at this
-          ** number is this very table. */
-          int unnamedRootMatch = 0;
-          int renameRootMatch = rootMatch
-            && aStaged[k].zName && aWorking[j].zName
-            && strcmp(aStaged[k].zName, aWorking[j].zName)!=0
-            && !addFindEntryByName(aWorking, nWorking, aStaged[k].zName)
-            && !addFindEntryByName(aStaged, nStaged, aWorking[j].zName);
-          if( rootMatch && !aStaged[k].zName && aWorking[j].zName ){
-            int r;
-            for(r=0; r<nStagedSchema; r++){
-              if( aStagedSchema[r].iRootpage==aStaged[k].iTable
-               && aStagedSchema[r].zType
-               && strcmp(aStagedSchema[r].zType, "table")==0
-               && aStagedSchema[r].zName
-               && strcmp(aStagedSchema[r].zName, aWorking[j].zName)==0 ){
-                unnamedRootMatch = 1;
-                break;
-              }
-            }
-          }
-          if( nameMatch || unnamedRootMatch || renameRootMatch ){
-            int schemaChanged =
-              prollyHashCompare(&aStaged[k].schemaHash, &aWorking[j].schemaHash)!=0;
-            int nameChanged =
-              (!aStaged[k].zName) != (!aWorking[j].zName)
-              || (aStaged[k].zName && aWorking[j].zName
-                  && strcmp(aStaged[k].zName, aWorking[j].zName)!=0);
-            char *zDup;
-            /* A rename retires the old name: its rows (and its indexes'
-            ** rows, keyed by the old tbl_name) must follow WORKING too,
-            ** where they no longer exist. */
-            if( nameChanged && aStaged[k].zName ){
-              ADDNAMED_TOUCH(aStaged[k].zName);
-            }
-            zDup = aWorking[j].zName
-                           ? sqlite3_mprintf("%s", aWorking[j].zName) : 0;
-            if( aWorking[j].zName && !zDup ){
-              ADDNAMED_FREE_ALL();
-              sqlite3_result_error_nomem(context);
-              return SQLITE_NOMEM;
-            }
-            sqlite3_free(aStaged[k].zName);
-            aStaged[k] = aWorking[j];
-            aStaged[k].zName = zDup;
-            if( schemaChanged || nameChanged ){
-              updateMaster = 1;
-            }
-            updated = 1;
-            break;
-          }
-        }
-        if( !updated ){
-          updateMaster = 1;
-          rc = addAppendTableEntry(context, &aStaged, &nStaged, &aWorking[j]);
-          if( rc!=SQLITE_OK ){
-            ADDNAMED_FREE_ALL();
-            return rc;
-          }
-        }
-        /* An index travels with its table: replace whatever index entries
-        ** the staged catalog carried for this table with the working ones. */
-        addRemoveIndexEntriesOfTable(aStaged, &nStaged,
-                                     aStagedSchema, nStagedSchema, zTable);
-        rc = addAppendIndexEntriesOfTable(context, &aStaged, &nStaged,
-                                          aWorking, nWorking,
-                                          aWorkSchema, nWorkSchema, zTable);
-        if( rc!=SQLITE_OK ){
-          ADDNAMED_FREE_ALL();
-          return rc;
-        }
-        rc = addStageShadowTablesOf(db, context, &aStaged, &nStaged,
-                                    aWorking, nWorking,
-                                    aWorkSchema, nWorkSchema,
-                                    aStagedSchema, nStagedSchema, zTable);
-        if( rc!=SQLITE_OK ){
-          ADDNAMED_FREE_ALL();
-          return rc;
-        }
-        ADDNAMED_TOUCH(zTable);
-        break;
-      }
-    }
-  }
-
-  if( updateMaster ){
-    struct TableEntry *pWorkingMaster = doltliteFindTableByNumber(aWorking, nWorking, 1);
-    struct TableEntry *pStagedMaster = doltliteFindTableByNumber(aStaged, nStaged, 1);
-    if( pWorkingMaster ){
-      /* The staged master must share the working numbering domain, but a
-      ** wholesale adoption would carry unstaged view and trigger changes
-      ** into the commit (Dolt keeps them out of a named add). Compose the
-      ** master: table and index rows from working, view and trigger rows
-      ** from the previously staged state. */
-      ProllyHash composedRoot;
-      rc = doltliteBuildNamedStageMasterRoot(db,
-              &pWorkingMaster->root, pWorkingMaster->flags,
-              pStagedMaster ? &pStagedMaster->root : 0,
-              pStagedMaster ? pStagedMaster->flags : 0,
-              (const char**)azTouched, nTouched,
-              &composedRoot);
-      if( rc!=SQLITE_OK ){
-        ADDNAMED_FREE_ALL();
-        sqlite3_result_error_code(context, rc);
-        return rc;
-      }
-      if( pStagedMaster ){
-        pStagedMaster->root = composedRoot;
-        pStagedMaster->schemaHash = pWorkingMaster->schemaHash;
-        pStagedMaster->flags = pWorkingMaster->flags;
-      }else{
-        struct TableEntry composed = *pWorkingMaster;
-        composed.zName = 0;
-        composed.root = composedRoot;
-        rc = addAppendTableEntry(context, &aStaged, &nStaged, &composed);
-        if( rc!=SQLITE_OK ){
-          ADDNAMED_FREE_ALL();
-          return rc;
-        }
-      }
-    }
-  }
-
-  addAlignStagedEntriesToWorking(aWorking, nWorking, aStaged, nStaged);
-  rc = addWriteStagedCatalog(db, cs, aStaged, nStaged);
-  ADDNAMED_FREE_ALL();
-  #undef ADDNAMED_TOUCH
-  #undef ADDNAMED_FREE_ALL
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-  }
-  return rc;
-}
-
-static int doltliteStageArgsAndPersist(
-  sqlite3 *db,
-  sqlite3_context *context,
-  ChunkStore *cs,
-  int argc,
-  sqlite3_value **argv,
-  int stageAll
-){
-  ProllyHash workingHash;
-  ProllyHash savedStaged;
-  int rc;
-
-  doltliteGetSessionStaged(db, &savedStaged);
-
-  rc = doltlitePrepareCatalogForPersistence(db);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error(context, "failed to prepare catalog", -1);
-    return rc;
-  }
-
-  rc = doltliteFlushCatalogToHash(db, &workingHash);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error(context, "failed to flush", -1);
-    return rc;
-  }
-
-  if( stageAll ){
-    rc = addStageAllTables(db, context, cs, &workingHash);
-  }else{
-    rc = addStageNamedTables(db, context, cs, &workingHash, argc, argv);
-  }
-  if( rc!=SQLITE_OK ) return rc;
-
-  rc = doltlitePersistWorkingSet(db);
-  if( rc!=SQLITE_OK ){
-    doltliteSetSessionStaged(db, &savedStaged);
-    sqlite3_result_error_code(context, rc);
-    return rc;
-  }
-  return SQLITE_OK;
-}
-
-static void doltliteAddFunc(
-  sqlite3_context *context,
-  int argc,
-  sqlite3_value **argv
-){
-  sqlite3 *db = sqlite3_context_db_handle(context);
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  int sealTopLevel = doltliteSavepointIsTopLevelTxn(db);
-  int rc;
-  int i;
-  int stageAll = 0;
-  int opRc = SQLITE_OK;
-
-  if( !cs ){
-    sqlite3_result_error(context, doltliteVcUnavailableMessage(db), -1);
-    goto add_cleanup;
-  }
-  if( argc==0 ){
-    sqlite3_result_error(context, "dolt_add requires table name or '-A'", -1);
-    goto add_cleanup;
-  }
-
-  for(i=0; i<argc; i++){
-    const char *arg = (const char*)sqlite3_value_text(argv[i]);
-    if( !arg ) continue;
-    if( strcmp(arg, "-A")==0
-     || strcmp(arg, "--all")==0
-     || strcmp(arg, ".")==0 ){
-      stageAll = 1;
-    }else if( arg[0]=='-' ){
-      char *zErr = sqlite3_mprintf("unknown option `%s`", arg);
-      if( zErr ){
-        sqlite3_result_error(context, zErr, -1);
-        sqlite3_free(zErr);
-      }else{
-        sqlite3_result_error_nomem(context);
-      }
-      goto add_cleanup;
-    }
-  }
-
-  opRc = doltliteStageArgsAndPersist(db, context, cs, argc, argv, stageAll);
-  if( opRc!=SQLITE_OK ) goto add_cleanup;
-
-  sqlite3_result_int(context, 0);
-
-add_cleanup:
-  if( sealTopLevel ){
-    rc = doltliteVcSealTopLevelSavepointTxn(db);
-    if( rc==SQLITE_OK && opRc==SQLITE_OK ){
-      rc = doltliteStageArgsAndPersist(db, context, cs, argc, argv, stageAll);
-      if( rc!=SQLITE_OK ){
-        opRc = rc;
-      }
-    }
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error_code(context, rc);
-    }else if( opRc!=SQLITE_OK ){
-      sqlite3_result_error_code(context, opRc);
-    }
-  }else if( opRc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, opRc);
-  }
-}
-
-typedef struct DoltliteCommitOptions {
-  const char *zMessage;
-  const char *zAuthor;
-  const char *zDate;
-  int addAll;
-  int addModifiedOnly;
-  int amend;
-  int allowEmpty;
-  int skipEmpty;
-  int force;
-  i64 explicitTimestamp;
-} DoltliteCommitOptions;
-
-/* Parse dolt_commit's argv into opts. On a malformed option the context error
-** is set and a non-OK code returned; explicitTimestamp is left zero for the
-** caller to fill from --date after its own validation ordering. */
-static int doltliteCommitParseOptions(
-  sqlite3_context *context,
-  int argc,
-  sqlite3_value **argv,
-  DoltliteCommitOptions *opts
-){
-  int i;
-  memset(opts, 0, sizeof(*opts));
-  for(i=0; i<argc; i++){
-    const char *arg = (const char*)sqlite3_value_text(argv[i]);
-    if( !arg ) continue;
-    if( arg[0]=='-' && arg[1]!='-' && arg[1]!=0 && arg[2]!=0 ){
-      int j;
-      for(j=1; arg[j]; j++){
-        if( arg[j]=='A' ){
-          opts->addAll = 1;
-        }else if( arg[j]=='a' ){
-          opts->addModifiedOnly = 1;
-        }else if( arg[j]=='f' ){
-          opts->force = 1;
-        }else if( arg[j]=='m' ){
-          if( arg[j+1]!=0 ){
-            opts->zMessage = &arg[j+1];
-          }else if( i+1<argc ){
-            opts->zMessage = (const char*)sqlite3_value_text(argv[++i]);
-          }else{
-            sqlite3_result_error(context, "no value for option `message'", -1);
-            return SQLITE_ERROR;
-          }
-          break;
-        }else{
-          char *zErr = sqlite3_mprintf("unknown option `-%c'", arg[j]);
-          if( zErr ){
-            sqlite3_result_error(context, zErr, -1);
-            sqlite3_free(zErr);
-          }else{
-            sqlite3_result_error_nomem(context);
-          }
-          return SQLITE_ERROR;
-        }
-      }
-    }else if( strcmp(arg, "-m")==0 ){
-      if( i+1<argc ){
-        opts->zMessage = (const char*)sqlite3_value_text(argv[++i]);
-      }else{
-        sqlite3_result_error(context, "no value for option `message'", -1);
-        return SQLITE_ERROR;
-      }
-    }else if( strcmp(arg, "--message")==0 ){
-      if( i+1<argc ){
-        opts->zMessage = (const char*)sqlite3_value_text(argv[++i]);
-      }else{
-        sqlite3_result_error(context, "no value for option `message'", -1);
-        return SQLITE_ERROR;
-      }
-    }else if( strcmp(arg, "--author")==0 ){
-      if( i+1<argc ){
-        opts->zAuthor = (const char*)sqlite3_value_text(argv[++i]);
-      }else{
-        sqlite3_result_error(context, "no value for option `author'", -1);
-        return SQLITE_ERROR;
-      }
-    }else if( strcmp(arg, "--date")==0 ){
-      if( i+1<argc ){
-        opts->zDate = (const char*)sqlite3_value_text(argv[++i]);
-      }else{
-        sqlite3_result_error(context, "no value for option `date'", -1);
-        return SQLITE_ERROR;
-      }
-    }else if( strcmp(arg, "--amend")==0 ){
-      opts->amend = 1;
-    }else if( strcmp(arg, "--allow-empty")==0 ){
-      opts->allowEmpty = 1;
-    }else if( strcmp(arg, "--skip-empty")==0 ){
-      opts->skipEmpty = 1;
-    }else if( strcmp(arg, "-f")==0 || strcmp(arg, "--force")==0 ){
-      opts->force = 1;
-    }else if( strcmp(arg, "-A")==0 ){
-      opts->addAll = 1;
-    }else if( strcmp(arg, "-a")==0 || strcmp(arg, "--all")==0 ){
-      opts->addModifiedOnly = 1;
-    }else if( arg[0]=='-' ){
-      char *zErr = sqlite3_mprintf("unknown option `%s`", arg);
-      if( zErr ){
-        sqlite3_result_error(context, zErr, -1);
-        sqlite3_free(zErr);
-      }else{
-        sqlite3_result_error_nomem(context);
-      }
-      return SQLITE_ERROR;
-    }else{
-      char *zErr = sqlite3_mprintf(
-          "commit does not take positional arguments, but found 1: %s", arg);
-      if( zErr ){
-        sqlite3_result_error(context, zErr, -1);
-        sqlite3_free(zErr);
-      }else{
-        sqlite3_result_error_nomem(context);
-      }
-      return SQLITE_ERROR;
-    }
-  }
-  return SQLITE_OK;
-}
-
-/* Rebuild the staged catalog for `dolt_commit -a`, overlaying working-tree
-** changes for tables that already exist in HEAD onto the staged catalog, and
-** publish it as the session's staged catalog. On failure the context error is
-** set and a non-OK code returned. */
-static int doltliteCommitStageModifiedOnly(sqlite3 *db, sqlite3_context *context){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  ProllyHash workingHash, headCatHash, stagedHash;
-  struct TableEntry *aWorking = 0, *aHead = 0, *aStaged = 0;
-  int nWorking = 0, nHead = 0, nStaged = 0;
-  int nStagedAlloc = 0;
-  int j, k;
-  int rc;
-  AddNameIndex workingIdx;
-  AddNameIndex headIdx;
-  AddNameIndex stagedIdx;
-  u8 *aRemoveStaged = 0;
-
-  memset(&workingIdx, 0, sizeof(workingIdx));
-  memset(&headIdx, 0, sizeof(headIdx));
-  memset(&stagedIdx, 0, sizeof(stagedIdx));
-
-  #define FREE_ADD_MODIFIED_CATALOGS() do { \
-    sqlite3_free(aRemoveStaged); \
-    addNameIndexFree(&workingIdx); \
-    addNameIndexFree(&headIdx); \
-    addNameIndexFree(&stagedIdx); \
-    doltliteFreeCatalog(aWorking, nWorking); \
-    doltliteFreeCatalog(aHead, nHead); \
-    doltliteFreeCatalog(aStaged, nStaged); \
-  } while(0)
-
-  rc = doltliteFlushCatalogToHash(db, &workingHash);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error(context, "failed to flush", -1);
-    return rc;
-  }
-  rc = doltliteLoadCatalog(db, &workingHash, &aWorking, &nWorking, 0);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error(context, "failed to load working catalog", -1);
-    FREE_ADD_MODIFIED_CATALOGS();
-    return rc;
-  }
-  rc = doltliteGetHeadCatalogHash(db, &headCatHash);
-  if( rc==SQLITE_OK && !prollyHashIsEmpty(&headCatHash) ){
-    rc = doltliteLoadCatalog(db, &headCatHash, &aHead, &nHead, 0);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error(context, "failed to load HEAD catalog", -1);
-      FREE_ADD_MODIFIED_CATALOGS();
-      return rc;
-    }
-  }
-
-  doltliteGetSessionStaged(db, &stagedHash);
-  if( !prollyHashIsEmpty(&stagedHash) ){
-    rc = doltliteLoadCatalog(db, &stagedHash, &aStaged, &nStaged, 0);
-  }else if( !prollyHashIsEmpty(&headCatHash) ){
-    rc = doltliteLoadCatalog(db, &headCatHash, &aStaged, &nStaged, 0);
-  }
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error(context, "failed to load staged catalog", -1);
-    FREE_ADD_MODIFIED_CATALOGS();
-    return rc;
-  }
-
-  nStagedAlloc = nStaged + nWorking + 1;
-  if( nStagedAlloc>0 ){
-    struct TableEntry *aNewStaged = sqlite3_realloc(
-        aStaged, nStagedAlloc*(int)sizeof(struct TableEntry));
-    if( !aNewStaged ){
-      sqlite3_result_error_nomem(context);
-      FREE_ADD_MODIFIED_CATALOGS();
-      return SQLITE_NOMEM;
-    }
-    aStaged = aNewStaged;
-  }
-  aRemoveStaged = sqlite3_malloc(nStagedAlloc>0 ? nStagedAlloc : 1);
-  if( !aRemoveStaged ){
-    sqlite3_result_error_nomem(context);
-    FREE_ADD_MODIFIED_CATALOGS();
-    return SQLITE_NOMEM;
-  }
-  memset(aRemoveStaged, 0, nStagedAlloc>0 ? nStagedAlloc : 1);
-
-  rc = addNameIndexInit(&workingIdx, aWorking, nWorking);
-  if( rc==SQLITE_OK ) rc = addNameIndexInit(&headIdx, aHead, nHead);
-  if( rc==SQLITE_OK ) rc = addNameIndexInit(&stagedIdx, aStaged, nStaged);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_nomem(context);
-    FREE_ADD_MODIFIED_CATALOGS();
-    return rc;
-  }
-
-  /* The master entry carries no name and pairs by table number 1. The
-  ** staged catalog must use the WORKING master root so the schema rows
-  ** and the overlaid entries share one numbering domain; keeping HEAD's
-  ** master under working-numbered entries makes the serializer pair
-  ** entries with the wrong rows. */
-  for(j=0; j<nWorking; j++){
-    if( aWorking[j].iTable!=1 ) continue;
-    for(k=0; k<nStaged; k++){
-      if( aStaged[k].iTable==1 ){
-        sqlite3_free(aStaged[k].zName);
-        aStaged[k] = aWorking[j];
-        aStaged[k].zName = 0;
-        break;
-      }
-    }
-    break;
-  }
-
-  for(j=0; j<nWorking; j++){
-    const char *zName = aWorking[j].zName;
-    int updated = 0;
-    char *zDup;
-    struct TableEntry *pStaged;
-    if( !addNameIndexFind(&headIdx, zName) ) continue;
-
-    pStaged = addNameIndexFind(&stagedIdx, zName);
-    if( pStaged ){
-        k = (int)(pStaged - aStaged);
-        zDup = zName ? sqlite3_mprintf("%s", zName) : 0;
-        if( zName && !zDup ){
-          sqlite3_result_error_nomem(context);
-          FREE_ADD_MODIFIED_CATALOGS();
-          return SQLITE_NOMEM;
-        }
-        sqlite3_free(aStaged[k].zName);
-        aStaged[k] = aWorking[j];
-        aStaged[k].zName = zDup;
-        updated = 1;
-    }
-    if( !updated ){
-      zDup = zName ? sqlite3_mprintf("%s", zName) : 0;
-      if( zName && !zDup ){
-        sqlite3_result_error_nomem(context);
-        FREE_ADD_MODIFIED_CATALOGS();
-        return SQLITE_NOMEM;
-      }
-      aStaged[nStaged] = aWorking[j];
-      aStaged[nStaged].zName = zDup;
-      nStaged++;
-    }
-  }
-
-  for(k=0; k<nHead; k++){
-    const char *zName = aHead[k].zName;
-    struct TableEntry *pStaged;
-    if( addNameIndexFind(&workingIdx, zName) ) continue;
-    pStaged = addNameIndexFind(&stagedIdx, zName);
-    if( pStaged ){
-      int j2 = (int)(pStaged - aStaged);
-      if( j2>=0 && j2<nStaged ) aRemoveStaged[j2] = 1;
-    }
-  }
-  for(k=0; k<nStaged; ){
-    if( aRemoveStaged[k] ){
-      sqlite3_free(aStaged[k].zName);
-      if( k+1<nStaged ){
-        memmove(&aStaged[k], &aStaged[k+1],
-                (nStaged-k-1)*(int)sizeof(struct TableEntry));
-        memmove(&aRemoveStaged[k], &aRemoveStaged[k+1],
-                (nStaged-k-1)*(int)sizeof(u8));
-      }
-      nStaged--;
-      continue;
-    }
-    k++;
-  }
-
-  /* Index entries carry no name, so the by-name overlays above never
-  ** refresh them: the staged list still holds the old index roots under
-  ** tables whose data was just staged from working, committing a catalog
-  ** whose indexes disagree with their tables. Rebuild the unnamed
-  ** entries so each index follows its table's source: indexes of
-  ** working-sourced tables adopt the working entries, indexes kept for
-  ** staged-only tables stay, and indexes of deleted tables go away.
-  ** Parents resolve through each catalog's own schema rows because entry
-  ** numbers are meaningless across catalogs. */
-  {
-    SchemaEntry *aWorkSchema = 0, *aOldSchema = 0;
-    int nWorkSchema = 0, nOldSchema = 0;
-    const ProllyHash *pOldSrcHash =
-        !prollyHashIsEmpty(&stagedHash) ? &stagedHash : &headCatHash;
-    int i2, k2, nAppend = 0;
-
-    rc = loadSchemaFromCatalog(db, cs, doltliteGetCache(db), &workingHash,
-                               &aWorkSchema, &nWorkSchema);
-    if( rc==SQLITE_OK && !prollyHashIsEmpty(pOldSrcHash) ){
-      rc = loadSchemaFromCatalog(db, cs, doltliteGetCache(db), pOldSrcHash,
-                                 &aOldSchema, &nOldSchema);
-    }
-    if( rc!=SQLITE_OK ){
-      freeSchemaEntries(aWorkSchema, nWorkSchema);
-      freeSchemaEntries(aOldSchema, nOldSchema);
-      sqlite3_result_error(context, "failed to load schema for staging", -1);
-      FREE_ADD_MODIFIED_CATALOGS();
-      return rc;
-    }
-
-    for(k2=0; k2<nStaged; ){
-      const char *zParent = 0;
-      const char *zIdxName = 0;
-      if( aStaged[k2].iTable<=1 || aStaged[k2].zName ){ k2++; continue; }
-      for(i2=0; i2<nOldSchema; i2++){
-        if( aOldSchema[i2].zType
-         && strcmp(aOldSchema[i2].zType, "index")==0
-         && aOldSchema[i2].iRootpage==aStaged[k2].iTable ){
-          zParent = aOldSchema[i2].zTblName;
-          zIdxName = aOldSchema[i2].zName;
-          break;
-        }
-      }
-      if( zParent
-       && amTableStagedByName(aStaged, nStaged, zParent)
-       && !(addNameIndexFind(&workingIdx, zParent)
-            && addNameIndexFind(&headIdx, zParent)) ){
-        /* Staged-sourced index: its table was staged explicitly and is
-        ** not refreshed from working, so the entry keeps the staged data
-        ** root — but its number is from the add-time catalog's domain,
-        ** which the serializer resolves against working-domain schema
-        ** rows. Renumber by index name so the row pairs; an index that
-        ** no longer exists in working falls through to the drop. */
-        int renumbered = 0;
-        for(i2=0; i2<nWorkSchema; i2++){
-          if( aWorkSchema[i2].zType
-           && strcmp(aWorkSchema[i2].zType, "index")==0
-           && aWorkSchema[i2].zName && zIdxName
-           && strcmp(aWorkSchema[i2].zName, zIdxName)==0 ){
-            aStaged[k2].iTable = aWorkSchema[i2].iRootpage;
-            renumbered = 1;
-            break;
-          }
-        }
-        if( renumbered ){
-          k2++;
-          continue;
-        }
-      }
-      sqlite3_free(aStaged[k2].zName);
-      if( k2+1<nStaged ){
-        memmove(&aStaged[k2], &aStaged[k2+1],
-                (nStaged-k2-1)*(int)sizeof(struct TableEntry));
-      }
-      nStaged--;
-    }
-
-    for(i2=0; i2<nWorkSchema; i2++){
-      if( aWorkSchema[i2].zType
-       && strcmp(aWorkSchema[i2].zType, "index")==0
-       && aWorkSchema[i2].iRootpage>1
-       && aWorkSchema[i2].zTblName
-       && addNameIndexFind(&workingIdx, aWorkSchema[i2].zTblName)
-       && addNameIndexFind(&headIdx, aWorkSchema[i2].zTblName)
-       && amTableStagedByName(aStaged, nStaged, aWorkSchema[i2].zTblName) ){
-        nAppend++;
-      }
-    }
-    if( nAppend>0 && nStaged+nAppend>nStagedAlloc ){
-      struct TableEntry *aGrown = sqlite3_realloc(
-          aStaged, (nStaged+nAppend)*(int)sizeof(struct TableEntry));
-      if( !aGrown ){
-        freeSchemaEntries(aWorkSchema, nWorkSchema);
-        freeSchemaEntries(aOldSchema, nOldSchema);
-        sqlite3_result_error_nomem(context);
-        FREE_ADD_MODIFIED_CATALOGS();
-        return SQLITE_NOMEM;
-      }
-      aStaged = aGrown;
-      nStagedAlloc = nStaged + nAppend;
-    }
-    for(i2=0; i2<nWorkSchema; i2++){
-      if( !aWorkSchema[i2].zType
-       || strcmp(aWorkSchema[i2].zType, "index")!=0
-       || aWorkSchema[i2].iRootpage<=1
-       || !aWorkSchema[i2].zTblName
-       || !addNameIndexFind(&workingIdx, aWorkSchema[i2].zTblName)
-       || !addNameIndexFind(&headIdx, aWorkSchema[i2].zTblName)
-       || !amTableStagedByName(aStaged, nStaged, aWorkSchema[i2].zTblName) ){
-        continue;
-      }
-      for(j=0; j<nWorking; j++){
-        if( aWorking[j].iTable==aWorkSchema[i2].iRootpage
-         && aWorking[j].zName==0 ){
-          aStaged[nStaged] = aWorking[j];
-          aStaged[nStaged].zName = 0;
-          nStaged++;
-          break;
-        }
-      }
-    }
-
-    freeSchemaEntries(aWorkSchema, nWorkSchema);
-    freeSchemaEntries(aOldSchema, nOldSchema);
-  }
-
-  if( nStaged==0 ){
-    sqlite3_result_error(context,
-      "nothing to commit, working tree clean (use dolt_add to stage changes)", -1);
-    FREE_ADD_MODIFIED_CATALOGS();
-    return SQLITE_ERROR;
-  }
-
-  {
-    u8 *buf = 0;
-    int nBuf = 0;
-    ProllyHash newStagedHash;
-    rc = doltliteSerializeCatalogEntries(db, aStaged, nStaged, &buf, &nBuf);
-    if( rc==SQLITE_OK ){
-      rc = chunkStorePut(cs, buf, nBuf, &newStagedHash);
-    }
-    sqlite3_free(buf);
-    if( rc==SQLITE_OK ){
-      doltliteSetSessionStaged(db, &newStagedHash);
-    }
-  }
-
-  FREE_ADD_MODIFIED_CATALOGS();
-  #undef FREE_ADD_MODIFIED_CATALOGS
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-    return rc;
-  }
-  return SQLITE_OK;
-}
-
-/* Resolve the parent, author, and message (handling --amend and --author
-** parsing) and store the new commit object, returning its hash via
-** pCommitHashOut. On failure the context error is set and non-OK returned. */
-static int doltliteCommitCreateObject(
-  sqlite3 *db,
-  sqlite3_context *context,
-  const DoltliteCommitOptions *opts,
-  const ProllyHash *pCatalogHash,
-  ProllyHash *pCommitHashOut
-){
-  ProllyHash parentHash;
-  char *zParsedName = 0, *zParsedEmail = 0;
-  const char *zMessage = opts->zMessage;
-  const char *zAuthor = opts->zAuthor;
-  int rc;
-  doltliteGetSessionHead(db, &parentHash);
-
-  if( opts->amend ){
-    DoltliteCommit headCommit;
-    ProllyHash headHash;
-    memset(&headCommit, 0, sizeof(headCommit));
-    doltliteGetSessionHead(db, &headHash);
-    if( prollyHashIsEmpty(&headHash) ){
-      sqlite3_result_error(context,
-        "cannot --amend: branch has no commits", -1);
-      return SQLITE_ERROR;
-    }
-    rc = doltliteLoadCommit(db, &headHash, &headCommit);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error(context,
-        "cannot --amend: failed to load HEAD commit", -1);
-      return rc;
-    }
-    if( doltliteCommitParentCount(&headCommit)==0 ){
-      doltliteCommitClear(&headCommit);
-      sqlite3_result_error(context,
-        "cannot --amend: HEAD has no parent (initial commit)", -1);
-      return SQLITE_ERROR;
-    }
-
-    {
-      const ProllyHash *pParent = doltliteCommitParentHash(&headCommit, 0);
-      if( !pParent || prollyHashIsEmpty(pParent) ){
-        doltliteCommitClear(&headCommit);
-        sqlite3_result_error(context,
-          "cannot --amend: HEAD has no parent (initial commit)", -1);
-        return SQLITE_ERROR;
-      }
-      memcpy(&parentHash, pParent, sizeof(ProllyHash));
-    }
-    if( !zMessage || !*zMessage ){
-      zMessage = sqlite3_mprintf("%s",
-          headCommit.zMessage ? headCommit.zMessage : "");
-    }
-    doltliteCommitClear(&headCommit);
-  }
-
-  if( zAuthor ){
-    const char *lt = strchr(zAuthor, '<');
-    const char *gt = lt ? strchr(lt, '>') : 0;
-    if( lt && gt ){
-      int nameLen = (int)(lt - zAuthor);
-      while( nameLen>0 && zAuthor[nameLen-1]==' ' ) nameLen--;
-      zParsedName = sqlite3_mprintf("%.*s", nameLen, zAuthor);
-      zParsedEmail = sqlite3_mprintf("%.*s", (int)(gt-lt-1), lt+1);
-    }else{
-      zParsedName = sqlite3_mprintf("%s", zAuthor);
-      zParsedEmail = sqlite3_mprintf("");
-    }
-  }
-
-  {
-    const char *p = zMessage;
-    while( *p==' ' || *p=='\t' || *p=='\n' || *p=='\r' ) p++;
-    if( *p==0 ){
-      sqlite3_free(zParsedName);
-      sqlite3_free(zParsedEmail);
-      sqlite3_result_error(context,
-        "dolt_commit requires a non-empty message", -1);
-      return SQLITE_ERROR;
-    }
-  }
-
-  rc = doltliteCreateAndStoreCommitWithTime(db, &parentHash, pCatalogHash,
-      zMessage, zParsedName, zParsedEmail, 0, 0, opts->explicitTimestamp,
-      pCommitHashOut);
-  sqlite3_free(zParsedName);
-  sqlite3_free(zParsedEmail);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-    return rc;
-  }
-  return SQLITE_OK;
-}
-
-static void doltliteCommitFunc(
-  sqlite3_context *context,
-  int argc,
-  sqlite3_value **argv
-){
-  sqlite3 *db = sqlite3_context_db_handle(context);
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  DoltliteCommitOptions opts;
-  const char *zMessage, *zDate;
-  int addAll, addModifiedOnly, amend, allowEmpty, skipEmpty, force;
-  ProllyHash commitHash;
-  ProllyHash catalogHash;
-  ProllyHash sessionHeadBeforeLock;
-  char hexBuf[PROLLY_HASH_SIZE*2+1];
-  int sealTopLevel = doltliteSavepointIsTopLevelTxn(db);
-  int rc;
-
-  if( !cs ){
-    sqlite3_result_error(context, doltliteVcUnavailableMessage(db), -1);
-    return;
-  }
-
-  /* Top-level SAVEPOINT is sealed before option validation because Dolt keeps
-  ** that SQL transaction boundary durable even when dolt_commit later errors. */
-  if( sealTopLevel ){
-    rc = sqlite3_exec(db, "COMMIT", 0, 0, 0);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error(context, sqlite3_errmsg(db), -1);
-      sqlite3_result_error_code(context, rc);
-      return;
-    }
-  }
-
-  if( doltliteCommitParseOptions(context, argc, argv, &opts)!=SQLITE_OK ){
-    return;
-  }
-  zMessage = opts.zMessage;
-  zDate = opts.zDate;
-  addAll = opts.addAll;
-  addModifiedOnly = opts.addModifiedOnly;
-  amend = opts.amend;
-  allowEmpty = opts.allowEmpty;
-  skipEmpty = opts.skipEmpty;
-  force = opts.force;
-
-  if( !force && doltliteSessionHasConstraintViolations(db) ){
-    sqlite3_result_error(context,
-      "cannot commit: unresolved entries in dolt_constraint_violations. "
-      "Resolve them (DELETE from the per-table vtable) then retry, or "
-      "pass --force to commit anyway.",
-      -1);
-    return;
-  }
-
-  if( zDate ){
-    struct tm tm;
-    const char *p;
-    memset(&tm, 0, sizeof(tm));
-    p = strptime(zDate, "%Y-%m-%dT%H:%M:%S", &tm);
-    if( !p ){
-      memset(&tm, 0, sizeof(tm));
-      p = strptime(zDate, "%Y-%m-%d %H:%M:%S", &tm);
-    }
-    if( !p ){
-      char *zErr = sqlite3_mprintf(
-          "could not parse --date `%s` (expected YYYY-MM-DDTHH:MM:SS)", zDate);
-      sqlite3_result_error(context, zErr ? zErr : "bad --date", -1);
-      sqlite3_free(zErr);
-      return;
-    }
-    opts.explicitTimestamp = (i64)timegm(&tm);
-  }
-
-  if( !zMessage || zMessage[0]==0 ){
-    sqlite3_result_error(context,
-      "dolt_commit requires a message: SELECT dolt_commit('-m', 'msg')", -1);
-    return;
-  }
-
-  if( !sealTopLevel
-   && (!db->autoCommit
-       || sqlite3_txn_state(db, "main")!=SQLITE_TXN_NONE
-       || db->pSavepoint) ){
-    /* Plain BEGIN and nested SAVEPOINT cases stay rollbackable until argument
-    ** validation and basic commit guards have succeeded. */
-    rc = sqlite3_exec(db, "COMMIT", 0, 0, 0);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error(context, sqlite3_errmsg(db), -1);
-      sqlite3_result_error_code(context, rc);
-      return;
-    }
-  }
-
-  rc = doltlitePrepareCatalogForPersistence(db);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error(context, "failed to prepare catalog", -1);
-    return;
-  }
-
-  if( addAll ){
-
-    rc = doltliteFlushCatalogToHash(db, &catalogHash);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error(context, "failed to flush", -1);
-      return;
-    }
-    doltliteSetSessionStaged(db, &catalogHash);
-  }else if( addModifiedOnly ){
-    rc = doltliteCommitStageModifiedOnly(db, context);
-    if( rc!=SQLITE_OK ) return;
-  }
-
-  if( doltliteSessionHasSchemaConflicts(db) ){
-    sqlite3_result_error(context,
-      "cannot commit: unresolved schema conflicts. Abort the merge, align "
-      "the schemas on one side, then rerun the merge.", -1);
-    return;
-  }
-
-  {
-    ProllyHash cfHash;
-    doltliteGetSessionConflictsCatalog(db, &cfHash);
-    if( !prollyHashIsEmpty(&cfHash) ){
-      sqlite3_result_error(context,
-        "cannot commit: unresolved merge conflicts. Use dolt_conflicts_resolve() first.", -1);
-      return;
-    }
-  }
-
-  /* Final merge commits refresh derived sqlite_stat rows. */
-  {
-    u8 isMergingCommit = 0;
-    doltliteGetSessionMergeState(db, &isMergingCommit, 0, 0);
-    if( isMergingCommit ){
-      sqlite3_stmt *pProbe = 0;
-      int hasStat1 = 0;
-      if( sqlite3_prepare_v2(db,
-          "SELECT 1 FROM main.sqlite_master "
-          "WHERE type='table' AND name='sqlite_stat1' LIMIT 1",
-          -1, &pProbe, 0)==SQLITE_OK ){
-        if( sqlite3_step(pProbe)==SQLITE_ROW ) hasStat1 = 1;
-        sqlite3_finalize(pProbe);
-      }
-      if( hasStat1 ){
-        ProllyHash refreshedCat;
-        (void)sqlite3_exec(db, "ANALYZE", 0, 0, 0);
-        if( doltliteFlushCatalogToHash(db, &refreshedCat)==SQLITE_OK ){
-          doltliteSetSessionStaged(db, &refreshedCat);
-        }
-      }
-    }
-  }
-
-  doltliteGetSessionStaged(db, &catalogHash);
-  if( prollyHashIsEmpty(&catalogHash) ){
-    if( allowEmpty ){
-      ProllyHash headCatHash;
-      rc = doltliteGetHeadCatalogHash(db, &headCatHash);
-      if( rc==SQLITE_OK && !prollyHashIsEmpty(&headCatHash) ){
-        memcpy(&catalogHash, &headCatHash, sizeof(ProllyHash));
-      }else{
-        sqlite3_result_error(context,
-          "nothing to commit (use dolt_add first, or dolt_commit('-A', '-m', 'msg'))", -1);
-        return;
-      }
-    }else{
-      sqlite3_result_error(context,
-        "nothing to commit (use dolt_add first, or dolt_commit('-A', '-m', 'msg'))", -1);
-      return;
-    }
-  }
-
-  {
-    u8 isMerging = 0;
-    doltliteGetSessionMergeState(db, &isMerging, 0, 0);
-    if( !isMerging && !amend ){
-      ProllyHash headCatHash;
-      rc = doltliteGetHeadCatalogHash(db, &headCatHash);
-      if( rc==SQLITE_OK && !prollyHashIsEmpty(&headCatHash)
-       && prollyHashCompare(&catalogHash, &headCatHash)==0 ){
-        if( allowEmpty ){
-
-        }else if( skipEmpty ){
-          sqlite3_result_int(context, 0);
-          return;
-        }else{
-          sqlite3_result_error(context,
-            "nothing to commit, working tree clean (use dolt_add to stage changes)", -1);
-          return;
-        }
-      }
-    }
-  }
-
-  rc = doltliteCommitCreateObject(db, context, &opts, &catalogHash, &commitHash);
-  if( rc!=SQLITE_OK ) return;
-
-  doltliteGetSessionHead(db, &sessionHeadBeforeLock);
-
-  rc = doltliteRefreshAndConfirmHead(db, cs, &sessionHeadBeforeLock);
-  if( rc==SQLITE_BUSY ){
-    sqlite3_result_error(context,
-      "commit conflict: another connection committed to this branch. "
-      "Please retry your transaction.", -1);
-    return;
-  }
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-    return;
-  }
-
-  {
-    u8 wasMerging = 0;
-    doltliteGetSessionMergeState(db, &wasMerging, 0, 0);
-    if( wasMerging ){
-      doltliteClearSessionMergeState(db);
-    }
-  }
-
-  {
-    extern int doltliteClearAllConstraintViolations(sqlite3*);
-    if( doltliteSessionHasConstraintViolations(db) ){
-      doltliteClearAllConstraintViolations(db);
-    }
-  }
-
-  {
-    ProllyHash workingCatHash;
-    rc = doltliteFlushCatalogToHash(db, &workingCatHash);
-    if( rc==SQLITE_OK ){
-      rc = doltliteAdvanceBranch(db, &commitHash, &catalogHash, &workingCatHash);
-    }
-  }
-  chunkStoreUnlock(cs);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-    return;
-  }
-
-  doltliteHashToHex(&commitHash, hexBuf);
-
-  rc = doltliteRegisterDiffTables(db);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-    return;
-  }
-  rc = doltliteRegisterWorkspaceTables(db);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-    return;
-  }
-  rc = doltliteRegisterHistoryTables(db);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-    return;
-  }
-  rc = doltliteRegisterAtTablesForCatalog(db, &catalogHash);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-    return;
-  }
-  rc = doltliteRegisterBlameTables(db);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-    return;
-  }
-  rc = doltliteRefreshConstraintViolationTables(db);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-    return;
-  }
-
-  sqlite3_result_text(context, hexBuf, -1, SQLITE_TRANSIENT);
-}
-
-static int resetPathMatchesName(const struct TableEntry *pEntry, const char *zName){
-  return pEntry->zName && strcmp(pEntry->zName, zName)==0;
-}
-
-static int resetFindTableIndex(struct TableEntry *aTables, int nTables,
-                               const char *zTable){
-  int i;
-  for(i=0; i<nTables; i++){
-    if( resetPathMatchesName(&aTables[i], zTable) ) return i;
-  }
-  return -1;
-}
-
-static int resetStageNamedPaths(
-  sqlite3 *db,
-  ChunkStore *cs,
-  const char **azPaths,
-  int nPaths
-){
-  struct TableEntry *aHead = 0, *aStaged = 0;
-  SchemaEntry *aHeadSchema = 0;
-  int nHead = 0, nStaged = 0, nHeadSchema = 0;
-  ProllyHash headCatHash, stagedHash;
-  Pgno iNextFree = 2;
-  int p, k;
-  int rc;
-
-  rc = doltliteGetHeadCatalogHash(db, &headCatHash);
-  if( rc!=SQLITE_OK ) return rc;
-  if( !prollyHashIsEmpty(&headCatHash) ){
-    rc = doltliteLoadCatalog(db, &headCatHash, &aHead, &nHead, 0);
-    if( rc!=SQLITE_OK ) return rc;
-    rc = loadSchemaFromCatalog(db, cs, doltliteGetCache(db), &headCatHash,
-                               &aHeadSchema, &nHeadSchema);
-    if( rc!=SQLITE_OK ){
-      doltliteFreeCatalog(aHead, nHead);
-      return rc;
-    }
-  }
-
-  doltliteGetSessionStaged(db, &stagedHash);
-  if( !prollyHashIsEmpty(&stagedHash) ){
-    rc = doltliteLoadCatalog(db, &stagedHash, &aStaged, &nStaged, 0);
-    if( rc!=SQLITE_OK ) goto done;
-  }
-
-  for(k=0; k<nStaged; k++){
-    if( aStaged[k].iTable >= iNextFree ) iNextFree = aStaged[k].iTable + 1;
-  }
-
-  for(p=0; p<nPaths; p++){
-    const char *zTable = azPaths[p];
-    int iH = resetFindTableIndex(aHead, nHead, zTable);
-    int iS = resetFindTableIndex(aStaged, nStaged, zTable);
-    char *zDup;
-    if( iH<0 && iS<0 ){
-      rc = SQLITE_NOTFOUND;
-      goto done;
-    }
-    if( iH<0 ){
-      sqlite3_free(aStaged[iS].zName);
-      if( iS+1<nStaged ){
-        memmove(&aStaged[iS], &aStaged[iS+1],
-                (nStaged-iS-1)*(int)sizeof(struct TableEntry));
-      }
-      nStaged--;
-    }else if( iS<0 ){
-      /* Restoring a staged-dropped table: the HEAD entry is numbered in
-      ** HEAD's domain and its schema row is absent from the staged master
-      ** (and from the live schema -- the table is dropped there), so it
-      ** gets a fresh number and its row comes from the HEAD fallback. */
-      struct TableEntry *aNew = sqlite3_realloc(aStaged,
-          (nStaged+1)*(int)sizeof(struct TableEntry));
-      if( !aNew ){
-        rc = SQLITE_NOMEM;
-        goto done;
-      }
-      aStaged = aNew;
-      addRemoveIndexEntriesOfTable(aStaged, &nStaged,
-                                   aHeadSchema, nHeadSchema, zTable);
-      zDup = aHead[iH].zName ? sqlite3_mprintf("%s", aHead[iH].zName) : 0;
-      if( aHead[iH].zName && !zDup ){
-        rc = SQLITE_NOMEM;
-        goto done;
-      }
-      aStaged[nStaged] = aHead[iH];
-      aStaged[nStaged].zName = zDup;
-      aStaged[nStaged].iTable = iNextFree++;
-      nStaged++;
-      rc = addAppendIndexEntriesOfTable(0, &aStaged, &nStaged,
-                                        aHead, nHead,
-                                        aHeadSchema, nHeadSchema,
-                                        zTable);
-      if( rc!=SQLITE_OK ) goto done;
-    }else{
-      /* Take HEAD's content under the STAGED entry's number so the entry
-      ** keeps pairing with the staged catalog's schema row. */
-      Pgno iKeep = aStaged[iS].iTable;
-      zDup = aHead[iH].zName ? sqlite3_mprintf("%s", aHead[iH].zName) : 0;
-      if( aHead[iH].zName && !zDup ){
-        rc = SQLITE_NOMEM;
-        goto done;
-      }
-      sqlite3_free(aStaged[iS].zName);
-      aStaged[iS] = aHead[iH];
-      aStaged[iS].zName = zDup;
-      aStaged[iS].iTable = iKeep;
-    }
-  }
-
-  {
-    u8 *buf = 0;
-    int nBuf = 0;
-    ProllyHash newStagedHash;
-    rc = doltliteSerializeCatalogEntriesWithFallbackSchema(
-        db, aStaged, nStaged, aHeadSchema, nHeadSchema, &buf, &nBuf);
-    if( rc==SQLITE_OK ){
-      rc = chunkStorePut(cs, buf, nBuf, &newStagedHash);
-    }
-    sqlite3_free(buf);
-    if( rc==SQLITE_OK ){
-      doltliteSetSessionStaged(db, &newStagedHash);
-    }
-  }
-
-done:
-  doltliteFreeCatalog(aHead, nHead);
-  doltliteFreeCatalog(aStaged, nStaged);
-  if( aHeadSchema ){
-    for(k=0; k<nHeadSchema; k++) clearSchemaEntry(&aHeadSchema[k]);
-    sqlite3_free(aHeadSchema);
-  }
-  return rc;
-}
-
-static int mergeAbortInPlace(sqlite3 *db){
-  ProllyHash headCatHash;
-  int rc = doltliteGetHeadCatalogHash(db, &headCatHash);
-  if( rc!=SQLITE_OK ) return rc;
-  rc = doltliteHardReset(db, &headCatHash);
-  if( rc!=SQLITE_OK ) return rc;
-  doltliteSetSessionStaged(db, &headCatHash);
-  doltliteClearSessionMergeState(db);
-  {
-    extern int doltliteClearAllConstraintViolations(sqlite3*);
-  if( doltliteSessionHasConstraintViolations(db) ){
-      doltliteClearAllConstraintViolations(db);
-    }
-  }
-  rc = doltlitePersistWorkingSet(db);
-  if( rc!=SQLITE_OK ) return rc;
-  return doltliteVcSealActiveSavepoints(db);
-}
-
-static int mergeFastForward(
-  sqlite3 *db,
-  sqlite3_context *context,
-  ChunkStore *cs,
-  const ProllyHash *pOurHead,
-  const ProllyHash *pTheirHead
-){
-  DoltliteCommit theirCommit;
-  DoltliteTxnState savedState;
-  int graphLocked = 0;
-  int rc;
-  char hx[PROLLY_HASH_SIZE*2+1];
-
-  memset(&theirCommit, 0, sizeof(theirCommit));
-  memset(&savedState, 0, sizeof(savedState));
-
-  rc = doltliteLoadCommit(db, pTheirHead, &theirCommit);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error(context, "failed to load commit", -1);
-    return rc;
-  }
-  rc = doltliteSaveTxnState(db, &savedState);
-  if( rc!=SQLITE_OK ){
-    doltliteCommitClear(&theirCommit);
-    sqlite3_result_error_code(context, rc);
-    return rc;
-  }
-  rc = doltliteRefreshAndConfirmHead(db, cs, pOurHead);
-  if( rc==SQLITE_BUSY ){
-    doltliteTxnStateClear(&savedState);
-    doltliteCommitClear(&theirCommit);
-    sqlite3_result_error(context,
-      "merge conflict: another connection committed to this branch. Please retry your transaction.",
-      -1);
-    return rc;
-  }
-  if( rc!=SQLITE_OK ){
-    doltliteTxnStateClear(&savedState);
-    doltliteCommitClear(&theirCommit);
-    sqlite3_result_error_code(context, rc);
-    return rc;
-  }
-  graphLocked = 1;
-  rc = doltliteSwitchCatalog(db, &theirCommit.catalogHash);
-  if( rc==SQLITE_OK ){
-    rc = doltliteUpdateBranchWorkingState(db, doltliteGetSessionBranch(db),
-                                          &theirCommit.catalogHash, NULL);
-  }
-  if( rc==SQLITE_OK ){
-    rc = doltliteAdvanceBranch(db, pTheirHead, &theirCommit.catalogHash, 0);
-  }
-  if( graphLocked ) chunkStoreUnlock(cs);
-  if( rc!=SQLITE_OK ){
-    doltliteCommitClear(&theirCommit);
-    sqlite3_result_error_code(context,
-        doltliteRestoreTxnStateOnFailure(db, &savedState, rc));
-    return rc;
-  }
-  rc = doltliteVcSealActiveSavepoints(db);
-  if( rc!=SQLITE_OK ){
-    doltliteCommitClear(&theirCommit);
-    sqlite3_result_error_code(context, rc);
-    return rc;
-  }
-  doltliteTxnStateClear(&savedState);
-  doltliteCommitClear(&theirCommit);
-  doltliteHashToHex(pTheirHead, hx);
-  sqlite3_result_text(context, hx, -1, SQLITE_TRANSIENT);
-  return SQLITE_OK;
-}
-
-static int doltlitePreserveUntrackedTablesOnHardReset(
-  sqlite3 *db,
-  ChunkStore *cs,
-  const ProllyHash *pPreResetHeadCatHash,
-  ProllyHash *pTargetCatHash
-){
-  struct TableEntry *aHead = 0;
-  int nHead = 0;
-  int nUntracked = 0;
-  char **azUntracked = 0;
-  sqlite3_stmt *pStmt = 0;
-  int j, k;
-  int rc;
-
-  rc = doltliteLoadCatalog(db, pPreResetHeadCatHash, &aHead, &nHead, 0);
-  if( rc==SQLITE_OK ){
-    rc = sqlite3_prepare_v2(db,
-        "SELECT name FROM sqlite_master WHERE type='table' "
-        "AND name NOT LIKE 'sqlite_%' AND name NOT LIKE 'dolt_%'",
-        -1, &pStmt, 0);
-  }
-  if( rc==SQLITE_OK ){
-    while( sqlite3_step(pStmt)==SQLITE_ROW ){
-      const char *zName = (const char*)sqlite3_column_text(pStmt, 0);
-      int inHead = 0;
-      if( !zName ) continue;
-      for(k=0; k<nHead; k++){
-        if( aHead[k].zName && strcmp(aHead[k].zName, zName)==0 ){
-          inHead = 1;
-          break;
-        }
-      }
-      if( !inHead ){
-        char **aNew = sqlite3_realloc(azUntracked,
-            (nUntracked+1)*(int)sizeof(char*));
-        if( !aNew ){ rc = SQLITE_NOMEM; break; }
-        azUntracked = aNew;
-        azUntracked[nUntracked++] = sqlite3_mprintf("%s", zName);
-      }
-    }
-    sqlite3_finalize(pStmt);
-    pStmt = 0;
-  }
-
-  /* An untracked table's indexes carry their own catalog entries. */
-  if( rc==SQLITE_OK && nUntracked>0 ){
-    rc = sqlite3_prepare_v2(db,
-        "SELECT name FROM sqlite_master WHERE type='index' AND tbl_name=?",
-        -1, &pStmt, 0);
-    for(j=0; rc==SQLITE_OK && j<nUntracked; j++){
-      sqlite3_bind_text(pStmt, 1, azUntracked[j], -1, SQLITE_STATIC);
-      while( sqlite3_step(pStmt)==SQLITE_ROW ){
-        const char *zIdx = (const char*)sqlite3_column_text(pStmt, 0);
-        char **aNew;
-        if( !zIdx ) continue;
-        aNew = sqlite3_realloc(azUntracked,
-            (nUntracked+1)*(int)sizeof(char*));
-        if( !aNew ){ rc = SQLITE_NOMEM; break; }
-        azUntracked = aNew;
-        azUntracked[nUntracked++] = sqlite3_mprintf("%s", zIdx);
-      }
-      sqlite3_reset(pStmt);
-    }
-    sqlite3_finalize(pStmt);
-    pStmt = 0;
-  }
-
-  if( rc==SQLITE_OK && nUntracked>0 ){
-    ProllyHash workingHash;
-    struct TableEntry *aWorking = 0, *aTarget = 0;
-    SchemaEntry *aWorkSchema = 0;
-    int nWorking = 0, nTarget = 0, nWorkSchema = 0;
-    Pgno iNextFree = 2;
-
-    rc = doltliteFlushCatalogToHash(db, &workingHash);
-    if( rc==SQLITE_OK ){
-      rc = doltliteLoadCatalog(db, &workingHash, &aWorking, &nWorking, 0);
-    }
-    if( rc==SQLITE_OK ){
-      rc = doltliteLoadCatalog(db, pTargetCatHash, &aTarget, &nTarget, 0);
-    }
-    if( rc==SQLITE_OK ){
-      rc = loadSchemaFromCatalog(db, cs, doltliteGetCache(db), &workingHash,
-                                 &aWorkSchema, &nWorkSchema);
-    }
-    /* Build FROM the target so every tracked table -- including ones
-    ** dropped in the working tree -- is restored, then append the
-    ** untracked entries renumbered past the target range (working numbers
-    ** can collide with target numbers for different tables). Loaded index
-    ** entries carry no name, so each untracked object is located through
-    ** the working schema rows (name -> working number -> entry); its
-    ** schema row reaches the blob via the fallback list, which pairs by
-    ** name and stamps the appended number. */
-    if( rc==SQLITE_OK ){
-      for(k=0; k<nTarget; k++){
-        if( aTarget[k].iTable >= iNextFree ) iNextFree = aTarget[k].iTable + 1;
-      }
-      for(j=0; j<nUntracked && rc==SQLITE_OK; j++){
-        SchemaEntry *pSe = findSchemaEntry(aWorkSchema, nWorkSchema,
-                                           azUntracked[j]);
-        struct TableEntry *pWork = 0;
-        struct TableEntry *aNew;
-        char *zDup;
-        if( !pSe || pSe->iRootpage<=1 ) continue;
-        for(k=0; k<nWorking; k++){
-          if( aWorking[k].iTable==pSe->iRootpage ){
-            pWork = &aWorking[k];
-            break;
-          }
-        }
-        if( !pWork ) continue;
-        zDup = sqlite3_mprintf("%s", azUntracked[j]);
-        aNew = zDup ? sqlite3_realloc(aTarget,
-                        (nTarget+1)*(int)sizeof(struct TableEntry)) : 0;
-        if( !aNew ){
-          sqlite3_free(zDup);
-          rc = SQLITE_NOMEM;
-          break;
-        }
-        aTarget = aNew;
-        aTarget[nTarget] = *pWork;
-        aTarget[nTarget].zName = zDup;
-        aTarget[nTarget].iTable = iNextFree++;
-        nTarget++;
-      }
-    }
-    if( rc==SQLITE_OK ){
-      u8 *buf = 0;
-      int nBuf = 0;
-      ProllyHash mergedHash;
-      rc = doltliteSerializeCatalogEntriesForeignDomain(
-          db, aTarget, nTarget, aWorkSchema, nWorkSchema, &buf, &nBuf);
-      if( rc==SQLITE_OK ){
-        rc = chunkStorePut(cs, buf, nBuf, &mergedHash);
-      }
-      sqlite3_free(buf);
-      if( rc==SQLITE_OK ){
-        memcpy(pTargetCatHash, &mergedHash, sizeof(ProllyHash));
-      }
-    }
-    doltliteFreeCatalog(aWorking, nWorking);
-    doltliteFreeCatalog(aTarget, nTarget);
-    if( aWorkSchema ){
-      for(j=0; j<nWorkSchema; j++) clearSchemaEntry(&aWorkSchema[j]);
-      sqlite3_free(aWorkSchema);
-    }
-  }
-
-  if( pStmt ) sqlite3_finalize(pStmt);
-  for(j=0; j<nUntracked; j++) sqlite3_free(azUntracked[j]);
-  sqlite3_free(azUntracked);
-  doltliteFreeCatalog(aHead, nHead);
-  return rc;
-}
-
-static void doltliteResetFunc(
-  sqlite3_context *context,
-  int argc,
-  sqlite3_value **argv
-){
-  sqlite3 *db = sqlite3_context_db_handle(context);
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  ProllyHash targetCatHash;
-  ProllyHash targetCommit;
-  ProllyHash preResetHeadCatHash;
-  ProllyHash sessionHeadBeforeLock;
-  int havePreResetHead = 0;
-  int isHard = 0;
-  int isSoft = 0;
-  const char *zRef = 0;
-  const char **azPaths = 0;
-  int nPaths = 0;
-  int rc;
-  int i;
-  int graphLocked = 0;
-  u8 isMerging = 0;
-  int bSucceeded = 0;
-
-  if( !cs ){
-    sqlite3_result_error(context, doltliteVcUnavailableMessage(db), -1);
-    goto reset_cleanup;
-  }
-
-  if( doltliteGetHeadCatalogHash(db, &preResetHeadCatHash)==SQLITE_OK
-   && !prollyHashIsEmpty(&preResetHeadCatHash) ){
-    havePreResetHead = 1;
-  }
-
-  azPaths = (const char**)sqlite3_malloc(sizeof(char*) * (argc>0?argc:1));
-  if( !azPaths ){ sqlite3_result_error_nomem(context); goto reset_cleanup; }
-  for(i=0; i<argc; i++){
-    const char *arg = (const char*)sqlite3_value_text(argv[i]);
-    if( !arg ) continue;
-    if( strcmp(arg, "--hard")==0 ){ isHard = 1; }
-    else if( strcmp(arg, "--soft")==0 ){ isSoft = 1; }
-    else if( arg[0]=='-' ){
-      char *zErr = sqlite3_mprintf("unknown option `%s`", arg);
-      sqlite3_result_error(context, zErr ? zErr : "unknown option", -1);
-      sqlite3_free(zErr);
-      sqlite3_free(azPaths);
-      azPaths = 0;
-      goto reset_cleanup;
-    }
-    else if( !zRef ){
-
-      if( isHard || isSoft ){
-        zRef = arg;
-      }else{
-        ProllyHash probe;
-        if( doltliteResolveRef(db, arg, &probe)==SQLITE_OK ){
-          zRef = arg;
-        }else{
-          azPaths[nPaths++] = arg;
-        }
-      }
-    }
-    else{
-
-      azPaths[nPaths++] = arg;
-    }
-  }
-
-  if( isHard && isSoft ){
-    sqlite3_result_error(context,
-      "--hard and --soft are mutually exclusive options.", -1);
-    sqlite3_free(azPaths);
-    azPaths = 0;
-    goto reset_cleanup;
-  }
-
-  doltliteGetSessionMergeState(db, &isMerging, 0, 0);
-  if( isMerging && !isHard ){
-    sqlite3_free(azPaths);
-    azPaths = 0;
-    sqlite3_result_error(context,
-      "Merge conflict detected, transaction rolled back. "
-      "Merge conflicts must be resolved using the dolt_conflicts and "
-      "dolt_schema_conflicts tables before committing a transaction. "
-      "To commit transactions with merge conflicts, set "
-      "@@dolt_allow_commit_conflicts = 1", -1);
-    goto reset_cleanup;
-  }
-
-  if( nPaths>1 && !isHard && !isSoft && !zRef ){
-    nPaths = 0;
-  }
-
-  if( nPaths>0 ){
-    if( isHard || isSoft || zRef ){
-      sqlite3_result_error(context,
-        "table paths cannot be combined with --hard / --soft or a target ref", -1);
-      sqlite3_free(azPaths);
-      azPaths = 0;
-      goto reset_cleanup;
-    }
-    rc = resetStageNamedPaths(db, cs, azPaths, nPaths);
-    sqlite3_free(azPaths);
-    azPaths = 0;
-    if( rc==SQLITE_NOTFOUND ){
-      sqlite3_result_error(context, "table not found", -1);
-      goto reset_cleanup;
-    }
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error_code(context, rc);
-      goto reset_cleanup;
-    }
-    rc = doltlitePersistWorkingSet(db);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error_code(context, rc);
-      goto reset_cleanup;
-    }
-    sqlite3_result_int(context, 0);
-    goto reset_cleanup;
-  }
-  sqlite3_free(azPaths);
-  azPaths = 0;
-
-  if( isSoft && !zRef ){
-    sqlite3_result_int(context, 0);
-    goto reset_cleanup;
-  }
-
-  if( zRef ){
-    DoltliteCommit commit;
-
-    rc = doltliteResolveRef(db,zRef, &targetCommit);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error(context, "commit not found", -1);
-      goto reset_cleanup;
-    }
-
-    rc = doltliteLoadCommit(db, &targetCommit, &commit);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error(context, "failed to load commit", -1);
-      goto reset_cleanup;
-    }
-    memcpy(&targetCatHash, &commit.catalogHash, sizeof(ProllyHash));
-    doltliteCommitClear(&commit);
-
-    doltliteGetSessionHead(db, &sessionHeadBeforeLock);
-    rc = doltliteRefreshAndConfirmHead(db, cs, &sessionHeadBeforeLock);
-    if( rc==SQLITE_BUSY ){
-      sqlite3_result_error(context,
-        "reset conflict: another connection moved this branch. "
-        "Please retry your transaction.", -1);
-      goto reset_cleanup;
-    }
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error_code(context, rc);
-      goto reset_cleanup;
-    }
-    graphLocked = 1;
-
-    doltliteSetSessionHead(db, &targetCommit);
-    rc = chunkStoreUpdateBranch(cs, doltliteGetSessionBranch(db), &targetCommit);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error_code(context, rc);
-      goto reset_cleanup;
-    }
-
-    doltliteClearSessionMergeState(db);
-  }else{
-    rc = doltliteGetHeadCatalogHash(db, &targetCatHash);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error(context, "failed to read HEAD", -1);
-      goto reset_cleanup;
-    }
-  }
-
-  if( !isSoft ){
-    doltliteSetSessionStaged(db, &targetCatHash);
-  }
-
-  if( isHard ){
-
-    ProllyHash origStagedAfterReset;
-    memcpy(&origStagedAfterReset, &targetCatHash, sizeof(ProllyHash));
-
-    if( prollyHashIsEmpty(&targetCatHash) ){
-      sqlite3_result_error(context, "no commit to reset to", -1);
-      goto reset_cleanup;
-    }
-
-    if( havePreResetHead ){
-      rc = doltlitePreserveUntrackedTablesOnHardReset(
-        db, cs, &preResetHeadCatHash, &targetCatHash
-      );
-      if( rc!=SQLITE_OK ){
-        sqlite3_result_error_code(context, rc);
-        goto reset_cleanup;
-      }
-    }
-
-    rc = doltliteSaveWorkingSet(db);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error_code(context, rc);
-      goto reset_cleanup;
-    }
-    rc = doltliteHardReset(db, &targetCatHash);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error(context, "hard reset failed", -1);
-      goto reset_cleanup;
-    }
-
-    doltliteClearSessionMergeState(db);
-
-    {
-      extern int doltliteClearAllConstraintViolations(sqlite3*);
-      if( doltliteSessionHasConstraintViolations(db) ){
-        doltliteClearAllConstraintViolations(db);
-      }
-    }
-
-    doltliteSetSessionStaged(db, &origStagedAfterReset);
-    rc = doltlitePersistWorkingSet(db);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error_code(context, rc);
-      goto reset_cleanup;
-    }
-  }else{
-    rc = doltlitePersistWorkingSet(db);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error_code(context, rc);
-      goto reset_cleanup;
-    }
-  }
-
-  sqlite3_result_int(context, 0);
-  bSucceeded = 1;
-reset_cleanup:
-  sqlite3_free(azPaths);
-  if( graphLocked ){
-    chunkStoreUnlock(cs);
-  }
-  if( bSucceeded ){
-    rc = doltliteVcSealActiveSavepoints(db);
-  }else{
-    rc = doltliteVcSealTopLevelSavepointTxn(db);
-  }
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-  }
-}
-
-static int doltliteApplyMergeSchemaActions(
-  sqlite3 *db,
-  const ProllyHash *pAncCatHash,
-  const ProllyHash *pTheirCatHash,
-  SchemaMergeAction *aSchemaActions,
-  int nSchemaActions,
-  ProllyHash *pMergedCatHash
-){
-  int rc = SQLITE_OK;
-  int si;
-
-  for(si=0; si<nSchemaActions && rc==SQLITE_OK; si++){
-    int sj;
-    for(sj=0; sj<aSchemaActions[si].nAddColumns; sj++){
-      char *zAlter = sqlite3_mprintf("ALTER TABLE \"%w\" ADD COLUMN %s",
-                                      aSchemaActions[si].zTableName,
-                                      aSchemaActions[si].azAddColumns[sj]);
-      if( !zAlter ) return SQLITE_NOMEM;
-      rc = sqlite3_exec(db, zAlter, 0, 0, 0);
-      sqlite3_free(zAlter);
-      if( rc!=SQLITE_OK ) break;
-    }
-  }
-
-  /* Row data (theirs' adds, edits to shared columns, deletes, and added-column
-  ** values) is merged three-way in doltliteMergeCatalogs against a normalized
-  ** theirs root, so only the schema evolution (the ALTERs above) happens here. */
-  if( rc==SQLITE_OK ){
-    rc = doltliteFlushCatalogToHash(db, pMergedCatHash);
-  }
-  return rc;
-}
-
-static void doltliteMergeFunc(
-  sqlite3_context *context,
-  int argc,
-  sqlite3_value **argv
-){
-  sqlite3 *db = sqlite3_context_db_handle(context);
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  const char *zBranch = 0;
-  const char *zMessage = 0;
-  int isAbort = 0;
-  int noFastForward = 0;
-  u8 isMerging = 0;
-  ProllyHash ourHead, theirHead, ancestorHash;
-  ProllyHash ourCatHash, theirCatHash, ancCatHash, mergedCatHash;
-  DoltliteTxnState savedState;
-  int nMergeConflicts = 0;
-  DoltliteCommit ourCommit, theirCommit, ancCommit;
-  int graphLocked = 0;
-  int rc, i;
-
-  memset(&ourCommit, 0, sizeof(ourCommit));
-  memset(&theirCommit, 0, sizeof(theirCommit));
-  memset(&ancCommit, 0, sizeof(ancCommit));
-  memset(&savedState, 0, sizeof(savedState));
-
-  if( !cs ){ sqlite3_result_error(context, doltliteVcUnavailableMessage(db), -1); return; }
-  if( argc<1 ){ sqlite3_result_error(context, "usage: dolt_merge('branch')", -1); return; }
-
-  for(i=0; i<argc; i++){
-    const char *arg = (const char*)sqlite3_value_text(argv[i]);
-    if( !arg ) continue;
-    if( strcmp(arg, "--abort")==0 ){
-      isAbort = 1;
-    }else if( strcmp(arg, "--no-ff")==0 ){
-      noFastForward = 1;
-    }else if( strcmp(arg, "-m")==0 || strcmp(arg, "--message")==0 ){
-      if( i+1<argc ){
-        zMessage = (const char*)sqlite3_value_text(argv[++i]);
-      }else{
-        sqlite3_result_error(context, "-m requires a message", -1);
-        return;
-      }
-    }else if( arg[0]=='-' ){
-      char *zErr = sqlite3_mprintf("unknown option `%s`", arg);
-      if( zErr ){
-        sqlite3_result_error(context, zErr, -1);
-        sqlite3_free(zErr);
-      }else{
-        sqlite3_result_error_nomem(context);
-      }
-      return;
-    }else if( !zBranch ){
-      zBranch = arg;
-    }else{
-      sqlite3_result_error(context, "too many positional arguments to dolt_merge", -1);
-      return;
-    }
-  }
-
-  if( isAbort ){
-    if( zBranch || zMessage || noFastForward ){
-      sqlite3_result_error(context,
-        "--abort does not take other arguments", -1);
-      return;
-    }
-    doltliteGetSessionMergeState(db, &isMerging, 0, 0);
-    if( !isMerging ){
-      sqlite3_result_error(context, "no merge in progress", -1);
-      return;
-    }
-    rc = mergeAbortInPlace(db);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error_code(context, rc);
-      return;
-    }
-
-    sqlite3_result_int(context, 0);
-    return;
-  }
-
-  if( !zBranch ){
-    sqlite3_result_error(context, "branch name required", -1);
-    return;
-  }
-
-  doltliteGetSessionHead(db, &ourHead);
-  if( prollyHashIsEmpty(&ourHead) ){
-    sqlite3_result_error(context, "no commits on current branch", -1);
-    return;
-  }
-
-  rc = doltliteResolveRef(db, zBranch, &theirHead);
-  if( rc!=SQLITE_OK || prollyHashIsEmpty(&theirHead) ){
-    sqlite3_result_error(context, "merge source not found", -1);
-    return;
-  }
-
-  if( prollyHashCompare(&ourHead, &theirHead)==0 ){
-    sqlite3_result_text(context, "Already up to date", -1, SQLITE_STATIC);
-    return;
-  }
-
-  if( doltliteHasUncommittedChanges(db) ){
-    sqlite3_result_error(context,
-      "uncommitted changes \xe2\x80\x94 commit or reset before merging", -1);
-    return;
-  }
-
-  rc = doltliteFindAncestor(db, &ourHead, &theirHead, &ancestorHash);
-  if( rc!=SQLITE_OK || prollyHashIsEmpty(&ancestorHash) ){
-    sqlite3_result_error(context, "no common ancestor found", -1);
-    return;
-  }
-
-  if( prollyHashCompare(&ancestorHash, &theirHead)==0 ){
-    sqlite3_result_text(context, "Already up to date", -1, SQLITE_STATIC);
-    return;
-  }
-
-  if( prollyHashCompare(&ancestorHash, &ourHead)==0 && !noFastForward ){
-    rc = mergeFastForward(db, context, cs, &ourHead, &theirHead);
-    return;
-  }
-
-  rc = doltliteLoadCommit(db, &ourHead, &ourCommit);
-  if( rc!=SQLITE_OK ){ sqlite3_result_error(context, "failed to load our commit", -1); return; }
-  memcpy(&ourCatHash, &ourCommit.catalogHash, sizeof(ProllyHash));
-
-  rc = doltliteLoadCommit(db, &theirHead, &theirCommit);
-  if( rc!=SQLITE_OK ){ doltliteCommitClear(&ourCommit); sqlite3_result_error(context, "failed to load their commit", -1); return; }
-  memcpy(&theirCatHash, &theirCommit.catalogHash, sizeof(ProllyHash));
-
-  rc = doltliteLoadCommit(db, &ancestorHash, &ancCommit);
-  if( rc!=SQLITE_OK ){ doltliteCommitClear(&ourCommit); doltliteCommitClear(&theirCommit); sqlite3_result_error(context, "failed to load ancestor", -1); return; }
-  memcpy(&ancCatHash, &ancCommit.catalogHash, sizeof(ProllyHash));
-  doltliteCommitClear(&ancCommit);
-
-  rc = doltliteEnsureWriteTxnAndSavepoints(db);
-  if( rc!=SQLITE_OK ){
-    doltliteCommitClear(&ourCommit);
-    doltliteCommitClear(&theirCommit);
-    sqlite3_result_error_code(context, rc);
-    return;
-  }
-
-  rc = doltliteSaveTxnState(db, &savedState);
-  if( rc!=SQLITE_OK ){
-    doltliteCommitClear(&ourCommit);
-    doltliteCommitClear(&theirCommit);
-    sqlite3_result_error_code(context, rc);
-    return;
-  }
-
-  {
-    char *zMergeErr = 0;
-    SchemaMergeAction *aSchemaActions = 0;
-    int nSchemaActions = 0;
-    char **azReindex = 0;
-    int nReindex = 0;
-    rc = doltliteMergeCatalogs(db, &ancCatHash, &ourCatHash, &theirCatHash,
-                                &mergedCatHash, &nMergeConflicts, &zMergeErr,
-                                &aSchemaActions, &nSchemaActions, 0,
-                                &azReindex, &nReindex);
-    if( rc!=SQLITE_OK ){
-      doltliteCommitClear(&ourCommit);
-      doltliteCommitClear(&theirCommit);
-      if( zMergeErr ){
-        sqlite3_result_error(context, zMergeErr, -1);
-        sqlite3_free(zMergeErr);
-      }else{
-        sqlite3_result_error(context, "merge failed", -1);
-      }
-      doltliteTxnStateClear(&savedState);
-      freeSchemaMergeActions(aSchemaActions, nSchemaActions);
-      doltliteFreeNameList(azReindex, nReindex);
-      return;
-    }
-    sqlite3_free(zMergeErr);
-
-    if( nMergeConflicts>0 ){
-      ProllyHash conflictsHash;
-      doltliteGetSessionConflictsCatalog(db, &conflictsHash);
-      doltliteSetSessionMergeState(db, 1, &theirHead, &conflictsHash);
-    }
-
-    rc = doltliteRefreshAndConfirmHead(db, cs, &ourHead);
-    if( rc==SQLITE_BUSY ){
-      doltliteTxnStateClear(&savedState);
-      doltliteCommitClear(&ourCommit);
-      doltliteCommitClear(&theirCommit);
-      freeSchemaMergeActions(aSchemaActions, nSchemaActions);
-      doltliteFreeNameList(azReindex, nReindex);
-      sqlite3_result_error(context,
-        "merge conflict: another connection committed to this branch. Please retry your transaction.",
-        -1);
-      return;
-    }
-    if( rc!=SQLITE_OK ){
-      doltliteTxnStateClear(&savedState);
-      doltliteCommitClear(&ourCommit);
-      doltliteCommitClear(&theirCommit);
-      freeSchemaMergeActions(aSchemaActions, nSchemaActions);
-      doltliteFreeNameList(azReindex, nReindex);
-      sqlite3_result_error_code(context, rc);
-      return;
-    }
-    graphLocked = 1;
-
-    rc = doltliteSwitchCatalog(db, &mergedCatHash);
-    doltliteCommitClear(&ourCommit);
-    doltliteCommitClear(&theirCommit);
-    if( rc!=SQLITE_OK ){
-      if( graphLocked ){
-        chunkStoreUnlock(cs);
-        graphLocked = 0;
-      }
-      freeSchemaMergeActions(aSchemaActions, nSchemaActions);
-      doltliteFreeNameList(azReindex, nReindex);
-      sqlite3_result_error_code(context,
-          doltliteRestoreTxnStateOnFailure(db, &savedState, rc));
-      return;
-    }
-
-    /* Indexes adopted from the other branch carry only that branch's
-    ** rows; rebuild them over the merged tables while the merged catalog
-    ** is live so the flush below captures correct roots. */
-    if( nReindex>0 ){
-      rc = doltliteReindexNamedIndexes(db, azReindex, nReindex);
-    }
-    doltliteFreeNameList(azReindex, nReindex);
-    azReindex = 0;
-    nReindex = 0;
-    if( rc!=SQLITE_OK ){
-      if( graphLocked ){
-        chunkStoreUnlock(cs);
-        graphLocked = 0;
-      }
-      freeSchemaMergeActions(aSchemaActions, nSchemaActions);
-      sqlite3_result_error_code(context,
-          doltliteRestoreTxnStateOnFailure(db, &savedState, rc));
-      return;
-    }
-
-    if( nSchemaActions > 0 ){
-      rc = doltliteApplyMergeSchemaActions(db, &ancCatHash, &theirCatHash,
-                                           aSchemaActions, nSchemaActions,
-                                           &mergedCatHash);
-      freeSchemaMergeActions(aSchemaActions, nSchemaActions);
-      if( rc!=SQLITE_OK ){
-        if( graphLocked ){
-          chunkStoreUnlock(cs);
-          graphLocked = 0;
-        }
-        sqlite3_result_error_code(context,
-            doltliteRestoreTxnStateOnFailure(db, &savedState, rc));
-        return;
-      }
-    }else{
-      freeSchemaMergeActions(aSchemaActions, nSchemaActions);
-    }
-
-    /* Regenerate stats after a clean merge; stat rows are derived data. */
-    if( nMergeConflicts==0 ){
-      sqlite3_stmt *pProbe = 0;
-      int hasStat1 = 0;
-      if( sqlite3_prepare_v2(db,
-          "SELECT 1 FROM main.sqlite_master "
-          "WHERE type='table' AND name='sqlite_stat1' LIMIT 1",
-          -1, &pProbe, 0)==SQLITE_OK ){
-        if( sqlite3_step(pProbe)==SQLITE_ROW ) hasStat1 = 1;
-        sqlite3_finalize(pProbe);
-      }
-      if( hasStat1 ){
-        (void)sqlite3_exec(db, "ANALYZE", 0, 0, 0);
-      }
-    }
-
-    rc = doltliteFlushCatalogToHash(db, &mergedCatHash);
-    if( rc==SQLITE_OK ){
-      rc = doltliteSwitchCatalog(db, &mergedCatHash);
-    }
-    if( rc==SQLITE_OK ){
-      rc = doltlitePrimeSchemaCache(db);
-    }
-    if( rc==SQLITE_OK ){
-      doltliteSetSessionStaged(db, &mergedCatHash);
-      rc = doltliteUpdateBranchWorkingState(db,
-          doltliteGetSessionBranch(db), &mergedCatHash, NULL);
-    }
-    if( rc!=SQLITE_OK ){
-      if( graphLocked ){
-        chunkStoreUnlock(cs);
-        graphLocked = 0;
-      }
-      sqlite3_result_error_code(context,
-          doltliteRestoreTxnStateOnFailure(db, &savedState, rc));
-      return;
-    }
-  }
-
-  if( graphLocked ){
-    chunkStoreUnlock(cs);
-    graphLocked = 0;
-  }
-  {
-    int nViolations = 0;
-    int nUnique = 0;
-    int nCheck = 0;
-    char *zDetectErrMsg = 0;
-    int vrc = doltliteConstraintViolationBatchBegin(db);
-    if( vrc == SQLITE_OK ){
-      vrc = doltliteDetectMergeFkViolations(db, &ancCatHash,
-                                            &zDetectErrMsg, &nViolations);
-    }
-    if( vrc == SQLITE_OK ){
-      vrc = doltliteDetectMergeUniqueViolations(db, &ancCatHash,
-                                                &zDetectErrMsg, &nUnique);
-    }
-    if( vrc == SQLITE_OK ){
-      vrc = doltliteDetectMergeCheckViolations(db, &ancCatHash,
-                                               &zDetectErrMsg, &nCheck);
-    }
-    {
-      int erc = doltliteConstraintViolationBatchEnd(db, vrc==SQLITE_OK);
-      if( vrc==SQLITE_OK ) vrc = erc;
-    }
-    if( vrc != SQLITE_OK ){
-      if( zDetectErrMsg ){
-        sqlite3_result_error(context, zDetectErrMsg, -1);
-        sqlite3_free(zDetectErrMsg);
-        doltliteRestoreTxnStateOnFailure(db, &savedState, vrc);
-      }else{
-        sqlite3_result_error_code(context,
-            doltliteRestoreTxnStateOnFailure(db, &savedState, vrc));
-      }
-      return;
-    }
-    sqlite3_free(zDetectErrMsg);
-    if( nViolations + nUnique + nCheck > 0 ){
-      switch( doltliteVcTxnMode(db) ){
-      case DOLTLITE_VC_TXN_AUTOCOMMIT_LIKE:
-        rc = doltliteHardReset(db, &savedState.sessionCatalogHash);
-        if( rc==SQLITE_OK ){
-          doltliteSetSessionBranch(db, savedState.zSessionBranch);
-          doltliteSetSessionHead(db, &savedState.sessionHead);
-          doltliteSetSessionStaged(db, &savedState.sessionStaged);
-          doltliteSetSessionMergeState(db, savedState.sessionIsMerging,
-                                       &savedState.sessionMergeCommit,
-                                       &savedState.sessionConflictsCatalog);
-          doltliteSetSessionConstraintViolationsCatalog(
-              db, &savedState.sessionConstraintViolationsCatalog);
-          rc = doltlitePersistWorkingSet(db);
-        }
-        doltliteTxnStateClear(&savedState);
-        if( rc!=SQLITE_OK ){
-          sqlite3_result_error_code(context, rc);
-        }else{
-          sqlite3_result_error(context,
-            "Committing this transaction resulted in a working set with "
-            "constraint violations, transaction rolled back.", -1);
-        }
-        break;
-      case DOLTLITE_VC_TXN_NESTED_SAVEPOINT:
-        rc = doltliteRestoreTxnState(db, &savedState);
-        if( rc==SQLITE_OK ){
-          doltliteSetSessionMergeState(db, savedState.sessionIsMerging,
-                                       &savedState.sessionMergeCommit,
-                                       &savedState.sessionConflictsCatalog);
-          doltliteSetSessionConstraintViolationsCatalog(
-              db, &savedState.sessionConstraintViolationsCatalog);
-        }
-        doltliteTxnStateClear(&savedState);
-        if( rc!=SQLITE_OK ){
-          sqlite3_result_error_code(context, rc);
-        }else{
-          sqlite3_result_error(context,
-            "Merge aborted: would have introduced constraint violations. "
-            "The merge and the would-be violations have been rolled back "
-            "with the enclosing savepoint, so dolt_constraint_violations "
-            "is empty. To inspect the violations, re-run the merge inside "
-            "a plain BEGIN/COMMIT transaction (no SAVEPOINT) so the "
-            "violations are preserved instead of rolled back.",
-            -1);
-        }
-        break;
-      case DOLTLITE_VC_TXN_PLAIN:
-        rc = doltliteReportConstraintViolations(db, context, "Merge");
-        if( rc!=SQLITE_OK ){
-          sqlite3_result_error_code(context,
-              doltliteRestoreTxnStateOnFailure(db, &savedState, rc));
-          return;
-        }
-        doltliteTxnStateClear(&savedState);
-        break;
-      }
-      return;
-    }
-  }
-
-  if( nMergeConflicts > 0 ){
-    if( graphLocked ){
-      chunkStoreUnlock(cs);
-      graphLocked = 0;
-    }
-    switch( doltliteVcTxnMode(db) ){
-    case DOLTLITE_VC_TXN_AUTOCOMMIT_LIKE:
-      rc = doltliteRollbackAutocommitConflict(db, context, &savedState);
-      if( rc!=SQLITE_OK ){
-        sqlite3_result_error_code(context, rc);
-      }
-      return;
-    case DOLTLITE_VC_TXN_NESTED_SAVEPOINT:
-      rc = doltliteRestoreTxnState(db, &savedState);
-      if( rc==SQLITE_OK ){
-        doltliteSetSessionMergeState(db, savedState.sessionIsMerging,
-                                     &savedState.sessionMergeCommit,
-                                     &savedState.sessionConflictsCatalog);
-        doltliteSetSessionConstraintViolationsCatalog(
-            db, &savedState.sessionConstraintViolationsCatalog);
-      }
-      doltliteTxnStateClear(&savedState);
-      if( rc!=SQLITE_OK ){
-        sqlite3_result_error_code(context, rc);
-      }else{
-        char msg[256];
-        sqlite3_snprintf(sizeof(msg), msg,
-          "Merge has %d conflict(s). Resolve and then commit with dolt_commit.",
-          nMergeConflicts);
-        sqlite3_result_error(context, msg, -1);
-      }
-      return;
-    case DOLTLITE_VC_TXN_PLAIN:
-      rc = doltliteReportConflicts(db, context, nMergeConflicts, "Merge");
-      if( rc!=SQLITE_OK ){
-        sqlite3_result_error_code(context,
-            doltliteRestoreTxnStateOnFailure(db, &savedState, rc));
-        return;
-      }
-      doltliteTxnStateClear(&savedState);
-      break;
-    }
-  }else{
-    ProllyHash commitHash;
-    char hexBuf[PROLLY_HASH_SIZE*2+1];
-    char msg[256];
-
-    doltliteSetSessionStaged(db, &mergedCatHash);
-
-    if( zMessage && zMessage[0] ){
-      sqlite3_snprintf(sizeof(msg), msg, "%s", zMessage);
-    }else{
-      snprintf(msg, sizeof(msg), "Merge branch '%s' into %s",
-               zBranch, doltliteGetSessionBranch(db));
-    }
-    rc = doltliteCreateAndStoreCommit(db, &ourHead, &mergedCatHash,
-        msg, NULL, NULL, &theirHead, 1, &commitHash);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error(context, "failed to create merge commit", -1);
-      return;
-    }
-
-    /* Re-confirm under the lock right before advancing; the merge's first
-    ** confirm is staled by intervening lock-cycling SQL (ANALYZE, etc.). */
-    rc = doltliteRefreshAndConfirmHead(db, cs, &ourHead);
-    if( rc==SQLITE_BUSY ){
-      sqlite3_result_error(context,
-        "merge conflict: another connection committed to this branch. Please retry your transaction.",
-        -1);
-      doltliteRestoreTxnStateOnFailure(db, &savedState, rc);
-      return;
-    }
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error_code(context,
-          doltliteRestoreTxnStateOnFailure(db, &savedState, rc));
-      return;
-    }
-    graphLocked = 1;
-
-    rc = doltliteAdvanceBranch(db, &commitHash, &mergedCatHash, 0);
-    if( graphLocked ){
-      chunkStoreUnlock(cs);
-      graphLocked = 0;
-    }
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error_code(context,
-          doltliteRestoreTxnStateOnFailure(db, &savedState, rc));
-      return;
-    }
-    rc = doltliteVcSealActiveSavepoints(db);
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error_code(context, rc);
-      return;
-    }
-    doltliteTxnStateClear(&savedState);
-
-    doltliteHashToHex(&commitHash, hexBuf);
-    sqlite3_result_text(context, hexBuf, -1, SQLITE_TRANSIENT);
-  }
-}
-
-static int doltliteLoadFirstParentCommit(
-  sqlite3 *db,
-  const DoltliteCommit *pCommit,
-  DoltliteCommit *pParentCommit
-){
-  const ProllyHash *pParent = doltliteCommitParentHash(pCommit, 0);
-  if( !pParent || prollyHashIsEmpty(pParent) ){
-    return SQLITE_EMPTY;
-  }
-  return doltliteLoadCommit(db, pParent, pParentCommit);
-}
-
-static int doltliteLoadHeadAndParentedCommit(
-  sqlite3 *db,
-  const ProllyHash *pTargetHash,
-  ProllyHash *pOurHead,
-  DoltliteCommit *pTargetCommit,
-  DoltliteCommit *pParentCommit,
-  DoltliteCommit *pOurCommit
-){
-  int rc = doltliteLoadCommit(db, pTargetHash, pTargetCommit);
-  if( rc!=SQLITE_OK ) return SQLITE_NOTFOUND;
-
-  if( doltliteCommitParentCount(pTargetCommit)==0 ){
-    return SQLITE_EMPTY;
-  }
-
-  rc = doltliteLoadFirstParentCommit(db, pTargetCommit, pParentCommit);
-  if( rc!=SQLITE_OK ) return SQLITE_NOTFOUND;
-
-  doltliteGetSessionHead(db, pOurHead);
-  if( prollyHashIsEmpty(pOurHead) ){
-    return SQLITE_DONE;
-  }
-
-  rc = doltliteLoadCommit(db, pOurHead, pOurCommit);
-  if( rc!=SQLITE_OK ) return SQLITE_ABORT;
-  return SQLITE_OK;
-}
-
-static int applyMergedCatalogAndCommit(
-  sqlite3 *db,
-  sqlite3_context *context,
-  const ProllyHash *ancCatHash,
-  const ProllyHash *ourCatHash,
-  const ProllyHash *theirCatHash,
-  const ProllyHash *ourHead,
-  const char *zMessage,
-  int *pnConflicts,
-  char *hexBuf
-){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  DoltliteTxnState savedState;
-  ProllyHash mergedCatHash;
-  ProllyHash liveMergedCatHash;
-  ProllyHash commitHash;
-  char *zMergeErr = 0;
-  int graphLocked = 0;
-  int bPreferOurMaster;
-  const char *zOpLabel;
-  int rc;
-
-  memset(&savedState, 0, sizeof(savedState));
-  if( hexBuf ) hexBuf[0] = '\0';
-  bPreferOurMaster = (sqlite3_strnicmp(zMessage, "Revert", 6)==0);
-  zOpLabel = bPreferOurMaster ? "Revert" : "Cherry-pick";
-
-  rc = doltliteEnsureWriteTxnAndSavepoints(db);
-  if( rc!=SQLITE_OK ) return rc;
-
-  rc = doltliteSaveTxnState(db, &savedState);
-  if( rc!=SQLITE_OK ) return rc;
-
-  {
-    char **azReindex = 0;
-    int nReindex = 0;
-    rc = doltliteMergeCatalogs(db, ancCatHash, ourCatHash, theirCatHash,
-                                &mergedCatHash, pnConflicts, &zMergeErr, 0, 0,
-                                bPreferOurMaster, &azReindex, &nReindex);
-    if( rc!=SQLITE_OK ){
-      sqlite3_free(zMergeErr);
-      doltliteTxnStateClear(&savedState);
-      doltliteFreeNameList(azReindex, nReindex);
-      return rc;
-    }
-    sqlite3_free(zMergeErr);
-
-    rc = doltliteRefreshAndConfirmHead(db, cs, ourHead);
-    if( rc!=SQLITE_OK ){
-      doltliteTxnStateClear(&savedState);
-      doltliteFreeNameList(azReindex, nReindex);
-      return rc;
-    }
-    graphLocked = 1;
-
-    rc = doltliteSwitchCatalog(db, &mergedCatHash);
-    if( rc==SQLITE_OK ){
-      rc = doltlitePrimeSchemaCache(db);
-    }
-    if( rc==SQLITE_OK && nReindex>0 ){
-      rc = doltliteReindexNamedIndexes(db, azReindex, nReindex);
-    }
-    doltliteFreeNameList(azReindex, nReindex);
-    if( rc!=SQLITE_OK ) goto apply_rollback;
-  }
-
-  rc = doltliteFlushCatalogToHash(db, &liveMergedCatHash);
-  if( rc==SQLITE_OK ){
-    rc = doltliteSwitchCatalog(db, &liveMergedCatHash);
-  }
-  if( rc!=SQLITE_OK ) goto apply_rollback;
-
-  doltliteSetSessionStaged(db, &liveMergedCatHash);
-  rc = doltliteUpdateBranchWorkingState(db,
-      doltliteGetSessionBranch(db), &liveMergedCatHash, NULL);
-  if( rc!=SQLITE_OK ) goto apply_rollback;
-
-  if( graphLocked ){
-    chunkStoreUnlock(cs);
-    graphLocked = 0;
-  }
-
-  {
-    int nViolations = 0;
-    int nUnique = 0;
-    int nCheck = 0;
-    char *zDetectErrMsg = 0;
-
-    rc = doltliteConstraintViolationBatchBegin(db);
-    if( rc==SQLITE_OK ){
-      rc = doltliteDetectMergeFkViolations(db, ancCatHash,
-                                           &zDetectErrMsg, &nViolations);
-    }
-    if( rc==SQLITE_OK ){
-      rc = doltliteDetectMergeUniqueViolations(db, ancCatHash,
-                                               &zDetectErrMsg, &nUnique);
-    }
-    if( rc==SQLITE_OK ){
-      rc = doltliteDetectMergeCheckViolations(db, ancCatHash,
-                                              &zDetectErrMsg, &nCheck);
-    }
-    {
-      int erc = doltliteConstraintViolationBatchEnd(db, rc==SQLITE_OK);
-      if( rc==SQLITE_OK ) rc = erc;
-    }
-    if( rc!=SQLITE_OK ){
-      if( zDetectErrMsg ){
-        sqlite3_result_error(context, zDetectErrMsg, -1);
-        sqlite3_free(zDetectErrMsg);
-      }
-      goto apply_rollback;
-    }
-    sqlite3_free(zDetectErrMsg);
-
-    if( nViolations + nUnique + nCheck > 0 ){
-      switch( doltliteVcTxnMode(db) ){
-      case DOLTLITE_VC_TXN_AUTOCOMMIT_LIKE:
-        rc = doltliteHardReset(db, &savedState.sessionCatalogHash);
-        if( rc==SQLITE_OK ){
-          doltliteSetSessionBranch(db, savedState.zSessionBranch);
-          doltliteSetSessionHead(db, &savedState.sessionHead);
-          doltliteSetSessionStaged(db, &savedState.sessionStaged);
-          doltliteSetSessionMergeState(db, savedState.sessionIsMerging,
-                                       &savedState.sessionMergeCommit,
-                                       &savedState.sessionConflictsCatalog);
-          {
-            extern int doltliteClearAllConstraintViolations(sqlite3*);
-            doltliteClearAllConstraintViolations(db);
-          }
-          rc = doltlitePersistWorkingSet(db);
-        }
-        doltliteTxnStateClear(&savedState);
-        if( rc!=SQLITE_OK ){
-          return rc;
-        }
-        sqlite3_result_error(context,
-          "Committing this transaction resulted in a working set with "
-          "constraint violations, transaction rolled back.", -1);
-        break;
-      case DOLTLITE_VC_TXN_PLAIN:
-        rc = doltliteReportConstraintViolations(db, context, zOpLabel);
-        if( rc!=SQLITE_OK ) goto apply_rollback;
-        rc = doltliteVcSealActiveSavepoints(db);
-        if( rc!=SQLITE_OK ) goto apply_rollback;
-        doltliteTxnStateClear(&savedState);
-        break;
-      case DOLTLITE_VC_TXN_NESTED_SAVEPOINT:
-        rc = doltliteRestoreTxnStateOnFailure(db, &savedState, SQLITE_OK);
-        if( rc!=SQLITE_OK ) return rc;
-        sqlite3_result_error(context,
-          "Merge aborted: would have introduced constraint violations. "
-          "The merge and the would-be violations have been rolled back "
-          "with the enclosing savepoint, so dolt_constraint_violations "
-          "is empty. Re-run the merge in autocommit mode (outside a "
-          "transaction) to inspect the violations in "
-          "dolt_constraint_violations.",
-          -1);
-        break;
-      }
-      return SQLITE_OK;
-    }
-  }
-
-  if( *pnConflicts > 0 ){
-    if( graphLocked ){
-      chunkStoreUnlock(cs);
-      graphLocked = 0;
-    }
-    switch( doltliteVcTxnMode(db) ){
-    case DOLTLITE_VC_TXN_AUTOCOMMIT_LIKE:
-      rc = doltliteRollbackAutocommitConflict(db, context, &savedState);
-      if( rc!=SQLITE_OK ) return rc;
-      return SQLITE_OK;
-    case DOLTLITE_VC_TXN_PLAIN:
-      rc = doltliteReportConflicts(db, context, *pnConflicts, zOpLabel);
-      if( rc!=SQLITE_OK ) goto apply_rollback;
-      rc = doltliteVcSealActiveSavepoints(db);
-      if( rc!=SQLITE_OK ) goto apply_rollback;
-      doltliteTxnStateClear(&savedState);
-      return SQLITE_OK;
-    case DOLTLITE_VC_TXN_NESTED_SAVEPOINT:
-      rc = doltliteRestoreTxnStateOnFailure(db, &savedState, SQLITE_OK);
-      if( rc!=SQLITE_OK ) return rc;
-      {
-        char msg[256];
-        sqlite3_snprintf(sizeof(msg), msg,
-          "%s has %d conflict(s). Resolve and then commit with dolt_commit.",
-          zOpLabel, *pnConflicts);
-        sqlite3_result_error(context, msg, -1);
-      }
-      return SQLITE_OK;
-    }
-  }
-
-  rc = doltliteCreateAndStoreCommit(db, ourHead, &liveMergedCatHash,
-      zMessage, NULL, NULL, NULL, 0, &commitHash);
-  if( rc!=SQLITE_OK ) goto apply_rollback;
-
-  /* Re-confirm under the lock right before advancing; the first confirm is
-  ** staled by intervening lock-cycling SQL. */
-  rc = doltliteRefreshAndConfirmHead(db, cs, ourHead);
-  if( rc!=SQLITE_OK ) goto apply_rollback;
-  graphLocked = 1;
-
-  rc = doltliteAdvanceBranch(db, &commitHash, &liveMergedCatHash, 0);
-  if( rc!=SQLITE_OK ) goto apply_rollback;
-
-  rc = doltliteVcSealActiveSavepoints(db);
-  if( rc!=SQLITE_OK ) goto apply_rollback;
-
-  if( graphLocked ){
-    chunkStoreUnlock(cs);
-  }
-  doltliteTxnStateClear(&savedState);
-  doltliteHashToHex(&commitHash, hexBuf);
-  return SQLITE_OK;
-
-apply_rollback:
-  if( graphLocked ){
-    chunkStoreUnlock(cs);
-  }
-  {
-    return doltliteRestoreTxnStateOnFailure(db, &savedState, rc);
-  }
-}
-
-static void doltliteCherryPickFunc(
-  sqlite3_context *context,
-  int argc,
-  sqlite3_value **argv
-){
-  sqlite3 *db = sqlite3_context_db_handle(context);
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  const char *zRef;
-  ProllyHash pickHash, ourHead;
-  DoltliteCommit pickCommit, parentCommit, ourCommit;
-  int nConflicts = 0;
-  int rc;
-  char hexBuf[PROLLY_HASH_SIZE*2+1];
-
-  memset(&pickCommit, 0, sizeof(pickCommit));
-  memset(&parentCommit, 0, sizeof(parentCommit));
-  memset(&ourCommit, 0, sizeof(ourCommit));
-
-  if( !cs ){ sqlite3_result_error(context, "no database", -1); return; }
-  if( argc<1 ){
-    sqlite3_result_error(context, "usage: dolt_cherry_pick('commit_hash')", -1);
-    return;
-  }
-  if( argc>1 ){
-    sqlite3_result_error(context,
-      "cherry-picking multiple commits is not supported yet.", -1);
-    return;
-  }
-
-  zRef = (const char*)sqlite3_value_text(argv[0]);
-  if( !zRef ){
-    sqlite3_result_error(context, "commit hash required", -1);
-    return;
-  }
-
-  if( doltliteHasUncommittedChanges(db) ){
-    sqlite3_result_error(context,
-      "cannot cherry-pick with uncommitted changes", -1);
-    return;
-  }
-
-  rc = doltliteResolveRef(db,zRef, &pickHash);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error(context, "invalid commit hash", -1);
-    return;
-  }
-  rc = doltliteLoadHeadAndParentedCommit(
-    db, &pickHash,
-    &ourHead, &pickCommit, &parentCommit, &ourCommit
-  );
-  if( rc==SQLITE_NOTFOUND ){
-    doltliteCommitClear(&pickCommit);
-    doltliteCommitClear(&parentCommit);
-    sqlite3_result_error(context, "commit not found", -1);
-    return;
-  }
-  if( rc==SQLITE_EMPTY ){
-    doltliteCommitClear(&pickCommit);
-    sqlite3_result_error(context, "cannot cherry-pick the initial commit", -1);
-    return;
-  }
-  if( rc==SQLITE_DONE ){
-    doltliteCommitClear(&pickCommit);
-    doltliteCommitClear(&parentCommit);
-    sqlite3_result_error(context, "no commits on current branch", -1);
-    return;
-  }
-  if( rc==SQLITE_ABORT ){
-    doltliteCommitClear(&pickCommit);
-    doltliteCommitClear(&parentCommit);
-    sqlite3_result_error(context, "failed to load HEAD commit", -1);
-    return;
-  }
-
-  {
-    const char *zMsg = pickCommit.zMessage;
-    char fallback[256];
-    if( !zMsg || !*zMsg ){
-      sqlite3_snprintf(sizeof(fallback), fallback, "cherry-pick of %s", zRef);
-      zMsg = fallback;
-    }
-
-    rc = applyMergedCatalogAndCommit(db, context,
-        &parentCommit.catalogHash, &ourCommit.catalogHash,
-        &pickCommit.catalogHash, &ourHead, zMsg, &nConflicts, hexBuf);
-  }
-
-  doltliteCommitClear(&pickCommit);
-  doltliteCommitClear(&parentCommit);
-  doltliteCommitClear(&ourCommit);
-
-  if( rc==SQLITE_BUSY ){
-    sqlite3_result_error(context,
-      "cherry-pick conflict: another connection committed to this branch. Please retry your transaction.",
-      -1);
-    return;
-  }
-  if( rc!=SQLITE_OK ){
-    char *zMsg = sqlite3_mprintf("cherry-pick of %s failed", zRef);
-    sqlite3_result_error(context, zMsg ? zMsg : "cherry-pick failed", -1);
-    sqlite3_free(zMsg);
-    return;
-  }
-
-  if( nConflicts > 0 ){
-    char *zMsg = sqlite3_mprintf(
-        "cherry-pick of %s resulted in %d conflict(s); "
-        "resolve in dolt_conflicts then dolt_commit to finish", zRef, nConflicts);
-    if( zMsg ){
-      sqlite3_result_error(context, zMsg, -1);
-      sqlite3_free(zMsg);
-    }else{
-      sqlite3_result_error_nomem(context);
-    }
-    return;
-  }else if( hexBuf[0] ){
-    sqlite3_result_text(context, hexBuf, -1, SQLITE_TRANSIENT);
-  }
-}
-
-static int doltliteTableEntryDiffers(
-  const struct TableEntry *a, const struct TableEntry *b
-){
-  if( !a && !b ) return 0;
-  if( !a || !b ) return 1;
-  if( prollyHashCompare(&a->root, &b->root)!=0 ) return 1;
-  if( prollyHashCompare(&a->schemaHash, &b->schemaHash)!=0 ) return 1;
-  return 0;
-}
-
-static int doltliteRevertCheckDirty(
-  sqlite3 *db,
-  const ProllyHash *pCommitCat,
-  const ProllyHash *pParentCat,
-  int *pConflict
-){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  ProllyHash headCatHash, stagedHash, workingCatHash;
-  u8 *wBuf = 0; int nWBuf = 0;
-  struct TableEntry *aCommit = 0, *aParent = 0;
-  struct TableEntry *aWorking = 0, *aCompare = 0;
-  int nCommit = 0, nParent = 0;
-  int nWorking = 0, nCompare = 0;
-  int i, rc;
-
-  *pConflict = 0;
-
-  if( !cs ) return SQLITE_OK;
-
-  rc = doltliteGetHeadCatalogHash(db, &headCatHash);
-  if( rc!=SQLITE_OK ) return rc;
-
-  doltliteGetSessionStaged(db, &stagedHash);
-  if( !prollyHashIsEmpty(&stagedHash)
-   && prollyHashCompare(&headCatHash, &stagedHash)!=0 ){
-    *pConflict = 1;
-    return SQLITE_OK;
-  }
-
-  rc = doltliteFlushAndSerializeCatalog(db, &wBuf, &nWBuf);
-  if( rc!=SQLITE_OK ) return rc;
-  rc = chunkStorePut(cs, wBuf, nWBuf, &workingCatHash);
-  sqlite3_free(wBuf);
-  if( rc!=SQLITE_OK ) return rc;
-
-  if( prollyHashCompare(&headCatHash, &workingCatHash)==0 ) return SQLITE_OK;
-
-  rc = doltliteLoadCatalog(db, pCommitCat, &aCommit, &nCommit, 0);
-  if( rc!=SQLITE_OK ) goto done;
-  rc = doltliteLoadCatalog(db, pParentCat, &aParent, &nParent, 0);
-  if( rc!=SQLITE_OK ) goto done;
-  rc = doltliteLoadCatalog(db, &workingCatHash, &aWorking, &nWorking, 0);
-  if( rc!=SQLITE_OK ) goto done;
-  if( !prollyHashIsEmpty(&headCatHash) ){
-    rc = doltliteLoadCatalog(db, &headCatHash, &aCompare, &nCompare, 0);
-    if( rc!=SQLITE_OK ) goto done;
-  }
-
-  for(i=0; i<nWorking; i++){
-    struct TableEntry *pCmp;
-    struct TableEntry *pInCommit;
-    struct TableEntry *pInParent;
-    pCmp = doltliteFindTableByName(aCompare, nCompare, aWorking[i].zName);
-    if( !doltliteTableEntryDiffers(pCmp, &aWorking[i]) ) continue;
-    pInCommit = doltliteFindTableByName(aCommit, nCommit, aWorking[i].zName);
-    pInParent = doltliteFindTableByName(aParent, nParent, aWorking[i].zName);
-    if( doltliteTableEntryDiffers(pInCommit, pInParent) ){
-      *pConflict = 1;
-      goto done;
-    }
-  }
-  for(i=0; i<nCompare; i++){
-    struct TableEntry *pInCommit;
-    struct TableEntry *pInParent;
-    if( doltliteFindTableByName(aWorking, nWorking, aCompare[i].zName) ) continue;
-    pInCommit = doltliteFindTableByName(aCommit, nCommit, aCompare[i].zName);
-    pInParent = doltliteFindTableByName(aParent, nParent, aCompare[i].zName);
-    if( doltliteTableEntryDiffers(pInCommit, pInParent) ){
-      *pConflict = 1;
-      goto done;
-    }
-  }
-
-done:
-  doltliteFreeCatalog(aCommit, nCommit);
-  doltliteFreeCatalog(aParent, nParent);
-  doltliteFreeCatalog(aWorking, nWorking);
-  doltliteFreeCatalog(aCompare, nCompare);
-  return rc;
-}
-
-static void doltliteRevertFunc(
-  sqlite3_context *context,
-  int argc,
-  sqlite3_value **argv
-){
-  sqlite3 *db = sqlite3_context_db_handle(context);
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  const char *zRef;
-  ProllyHash revertHash, ourHead;
-  ProllyHash liveOurCatalog;
-  DoltliteCommit revertCommit, parentCommit, ourCommit;
-  int nConflicts = 0;
-  int rc;
-  char hexBuf[PROLLY_HASH_SIZE*2+1];
-
-  memset(&revertCommit, 0, sizeof(revertCommit));
-  memset(&parentCommit, 0, sizeof(parentCommit));
-  memset(&ourCommit, 0, sizeof(ourCommit));
-
-  if( !cs ){ sqlite3_result_error(context, "no database", -1); return; }
-
-  if( argc<1 ){
-    sqlite3_result_int(context, 0);
-    return;
-  }
-  if( argc>1 ){
-    char *zErr = sqlite3_mprintf("branch not found: %s",
-        (const char*)sqlite3_value_text(argv[1]));
-    if( zErr ){
-      sqlite3_result_error(context, zErr, -1);
-      sqlite3_free(zErr);
-    }else{
-      sqlite3_result_error_nomem(context);
-    }
-    return;
-  }
-
-  zRef = (const char*)sqlite3_value_text(argv[0]);
-  if( !zRef ){
-    sqlite3_result_int(context, 0);
-    return;
-  }
-
-  rc = doltliteResolveRef(db,zRef, &revertHash);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error(context, "invalid commit hash", -1);
-    return;
-  }
-  rc = doltliteLoadHeadAndParentedCommit(
-    db, &revertHash,
-    &ourHead, &revertCommit, &parentCommit, &ourCommit
-  );
-  if( rc==SQLITE_NOTFOUND ){
-    doltliteCommitClear(&revertCommit);
-    doltliteCommitClear(&parentCommit);
-    sqlite3_result_error(context, "commit not found", -1);
-    return;
-  }
-  if( rc==SQLITE_EMPTY ){
-    doltliteCommitClear(&revertCommit);
-    sqlite3_result_error(context, "cannot revert the initial commit", -1);
-    return;
-  }
-  if( rc==SQLITE_DONE ){
-    doltliteCommitClear(&revertCommit);
-    doltliteCommitClear(&parentCommit);
-    sqlite3_result_error(context, "no commits on current branch", -1);
-    return;
-  }
-  if( rc==SQLITE_ABORT ){
-    doltliteCommitClear(&revertCommit);
-    doltliteCommitClear(&parentCommit);
-    sqlite3_result_error(context, "failed to load HEAD commit", -1);
-    return;
-  }
-
-  {
-    int dirtyConflict = 0;
-    rc = doltliteRevertCheckDirty(db,
-        &revertCommit.catalogHash, &parentCommit.catalogHash,
-        &dirtyConflict);
-    if( rc!=SQLITE_OK ) goto revert_error;
-    if( dirtyConflict ){
-      doltliteCommitClear(&revertCommit);
-      doltliteCommitClear(&parentCommit);
-      doltliteCommitClear(&ourCommit);
-      sqlite3_result_error(context,
-        "Your local changes would be overwritten by revert.\n"
-        "hint: Please commit your changes before you revert.", -1);
-      return;
-    }
-  }
-  rc = doltliteFlushCatalogToHash(db, &liveOurCatalog);
-  if( rc!=SQLITE_OK ) goto revert_error;
-
-  {
-    char msg[512];
-    sqlite3_snprintf(sizeof(msg), msg, "Revert \"%s\"",
-                     revertCommit.zMessage ? revertCommit.zMessage : zRef);
-
-    rc = applyMergedCatalogAndCommit(db, context,
-        &revertCommit.catalogHash, &liveOurCatalog,
-        &parentCommit.catalogHash, &ourHead, msg, &nConflicts, hexBuf);
-  }
-
-  doltliteCommitClear(&revertCommit);
-  doltliteCommitClear(&parentCommit);
-  doltliteCommitClear(&ourCommit);
-
-  if( rc==SQLITE_BUSY ){
-    sqlite3_result_error(context,
-      "revert conflict: another connection committed to this branch. Please retry your transaction.",
-      -1);
-    return;
-  }
-  if( rc!=SQLITE_OK ){
-    char *zMsg = sqlite3_mprintf("revert of \"%s\" failed", zRef);
-    sqlite3_result_error(context, zMsg ? zMsg : "revert failed", -1);
-    sqlite3_free(zMsg);
-    return;
-  }
-
-  if( nConflicts > 0 ){
-    char *zMsg = sqlite3_mprintf(
-        "revert of \"%s\" resulted in %d conflict(s); "
-        "resolve in dolt_conflicts then dolt_commit to finish", zRef, nConflicts);
-    if( zMsg ){
-      sqlite3_result_error(context, zMsg, -1);
-      sqlite3_free(zMsg);
-    }else{
-      sqlite3_result_error_nomem(context);
-    }
-    return;
-  }else if( hexBuf[0] ){
-    sqlite3_result_text(context, hexBuf, -1, SQLITE_TRANSIENT);
-  }
-  return;
-
-revert_error:
-  doltliteCommitClear(&revertCommit);
-  doltliteCommitClear(&parentCommit);
-  doltliteCommitClear(&ourCommit);
-  {
-    char *zMsg = sqlite3_mprintf("revert of \"%s\" failed", zRef);
-    sqlite3_result_error(context, zMsg ? zMsg : "revert failed", -1);
-    sqlite3_free(zMsg);
-  }
-}
-
-static int doltliteRebaseCollectReplaySet(
-  sqlite3 *db,
-  const ProllyHash *pHeadHash,
-  const ProllyHash *pUpstreamHash,
-  ProllyHash **paReplay,
-  int *pnReplay
-){
-  ProllyHashSet upstreamAncestors;
-  ProllyHashSet visited;
-  ProllyHash *queue = 0;
-  int qHead = 0, qTail = 0, qAlloc = 0;
-  ProllyHash *aReplay = 0;
-  int nReplay = 0, nAllocReplay = 0;
-  int rc;
-  int upstreamInit = 0;
-  int visitedInit = 0;
-  int i, j;
-
-  *paReplay = 0;
-  *pnReplay = 0;
-
-  rc = prollyHashSetInit(&upstreamAncestors, 256);
-  if( rc!=SQLITE_OK ) return rc;
-  upstreamInit = 1;
-
-  qAlloc = 64;
-  queue = sqlite3_malloc(qAlloc * (int)sizeof(ProllyHash));
-  if( !queue ){ rc = SQLITE_NOMEM; goto cleanup; }
-  queue[qTail++] = *pUpstreamHash;
-
-  while( qHead < qTail ){
-    ProllyHash cur = queue[qHead++];
-    DoltliteCommit c;
-
-    if( prollyHashIsEmpty(&cur) ) continue;
-    if( prollyHashSetContains(&upstreamAncestors, &cur) ) continue;
-    rc = prollyHashSetAdd(&upstreamAncestors, &cur);
-    if( rc!=SQLITE_OK ) goto cleanup;
-
-    memset(&c, 0, sizeof(c));
-    if( doltliteLoadCommit(db, &cur, &c)!=SQLITE_OK ) continue;
-    for(i=0; i<doltliteCommitParentCount(&c); i++){
-      const ProllyHash *pp = doltliteCommitParentHash(&c, i);
-      if( !pp || prollyHashIsEmpty(pp) ) continue;
-      if( prollyHashSetContains(&upstreamAncestors, pp) ) continue;
-      if( qTail >= qAlloc ){
-        int nNew = qAlloc * 2;
-        ProllyHash *tmp = sqlite3_realloc(queue, nNew*(int)sizeof(ProllyHash));
-        if( !tmp ){ doltliteCommitClear(&c); rc = SQLITE_NOMEM; goto cleanup; }
-        queue = tmp;
-        qAlloc = nNew;
-      }
-      queue[qTail++] = *pp;
-    }
-    doltliteCommitClear(&c);
-  }
-
-  /* BFS from HEAD dissolves merge commits: traverse all parents but only
-  ** include non-merge commits in the replay set, linearizing the history. */
-  qHead = qTail = 0;
-  queue[qTail++] = *pHeadHash;
-
-  rc = prollyHashSetInit(&visited, 256);
-  if( rc!=SQLITE_OK ) goto cleanup;
-  visitedInit = 1;
-
-  while( qHead < qTail ){
-    ProllyHash cur = queue[qHead++];
-    DoltliteCommit c;
-    int nParents;
-
-    if( prollyHashIsEmpty(&cur) ) continue;
-    if( prollyHashSetContains(&upstreamAncestors, &cur) ) continue;
-    if( prollyHashSetContains(&visited, &cur) ) continue;
-    rc = prollyHashSetAdd(&visited, &cur);
-    if( rc!=SQLITE_OK ) goto cleanup;
-
-    memset(&c, 0, sizeof(c));
-    rc = doltliteLoadCommit(db, &cur, &c);
-    if( rc!=SQLITE_OK ) goto cleanup;
-    nParents = doltliteCommitParentCount(&c);
-
-    if( nParents<=1 ){
-      if( nReplay >= nAllocReplay ){
-        int nNew = nAllocReplay ? nAllocReplay*2 : 16;
-        ProllyHash *tmp = sqlite3_realloc(aReplay, nNew*(int)sizeof(ProllyHash));
-        if( !tmp ){ doltliteCommitClear(&c); rc = SQLITE_NOMEM; goto cleanup; }
-        aReplay = tmp;
-        nAllocReplay = nNew;
-      }
-      aReplay[nReplay++] = cur;
-    }
-
-    for( j=0; j<nParents; j++ ){
-      const ProllyHash *pp = doltliteCommitParentHash(&c, j);
-      if( !pp || prollyHashIsEmpty(pp) ) continue;
-      if( prollyHashSetContains(&upstreamAncestors, pp) ) continue;
-      if( prollyHashSetContains(&visited, pp) ) continue;
-      if( qTail >= qAlloc ){
-        int nNew = qAlloc * 2;
-        ProllyHash *tmp = sqlite3_realloc(queue, nNew*(int)sizeof(ProllyHash));
-        if( !tmp ){ doltliteCommitClear(&c); rc = SQLITE_NOMEM; goto cleanup; }
-        queue = tmp;
-        qAlloc = nNew;
-      }
-      queue[qTail++] = *pp;
-    }
-    doltliteCommitClear(&c);
-  }
-
-  for(i=0; i<nReplay/2; i++){
-    ProllyHash tmp = aReplay[i];
-    aReplay[i] = aReplay[nReplay-1-i];
-    aReplay[nReplay-1-i] = tmp;
-  }
-
-  *paReplay = aReplay;
-  *pnReplay = nReplay;
-  aReplay = 0;
-  rc = SQLITE_OK;
-
-cleanup:
-  sqlite3_free(queue);
-  sqlite3_free(aReplay);
-  if( upstreamInit ) prollyHashSetFree(&upstreamAncestors);
-  if( visitedInit ) prollyHashSetFree(&visited);
-  return rc;
-}
-
-static int doltliteRebaseLinearReplay(
-  sqlite3 *db,
-  sqlite3_context *context,
-  const char *zUpstream,
-  char **pzFinalMessage
-){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  int sealTopLevel = db->pSavepoint!=0 && db->nSavepoint==0;
-  ProllyHash upstreamHash, headHash;
-  ProllyHash *aReplay = 0;
-  int nReplay = 0;
-  DoltliteTxnState saved;
-  DoltliteCommit upstreamCommit;
-  int savedInit = 0;
-  int rc;
-  int i;
-  char *zFailedMsg = 0;
-  int bConflict = 0;
-
-  *pzFinalMessage = 0;
-  memset(&saved, 0, sizeof(saved));
-  memset(&upstreamCommit, 0, sizeof(upstreamCommit));
-
-  if( doltliteHasUncommittedChanges(db) ){
-    sqlite3_result_error(context,
-      "cannot start a rebase with uncommitted changes", -1);
-    if( sealTopLevel ) (void)doltliteVcSealTopLevelSavepointTxn(db);
-    return SQLITE_ERROR;
-  }
-
-  rc = doltliteResolveRef(db, zUpstream, &upstreamHash);
-  if( rc!=SQLITE_OK ){
-    char *zErr = sqlite3_mprintf("branch not found: %s", zUpstream);
-    sqlite3_result_error(context, zErr ? zErr : "branch not found", -1);
-    sqlite3_free(zErr);
-    if( sealTopLevel ) (void)doltliteVcSealTopLevelSavepointTxn(db);
-    return SQLITE_ERROR;
-  }
-
-  doltliteGetSessionHead(db, &headHash);
-  if( prollyHashIsEmpty(&headHash) ){
-    sqlite3_result_error(context, "no commits on current branch", -1);
-    if( sealTopLevel ) (void)doltliteVcSealTopLevelSavepointTxn(db);
-    return SQLITE_ERROR;
-  }
-
-  rc = doltliteRebaseCollectReplaySet(db, &headHash, &upstreamHash,
-                                      &aReplay, &nReplay);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-    if( sealTopLevel ) (void)doltliteVcSealTopLevelSavepointTxn(db);
-    return rc;
-  }
-  if( nReplay==0 ){
-    sqlite3_free(aReplay);
-    sqlite3_result_error(context, "didn't identify any commits!", -1);
-    if( sealTopLevel ) (void)doltliteVcSealTopLevelSavepointTxn(db);
-    return SQLITE_ERROR;
-  }
-
-  rc = doltliteSaveTxnState(db, &saved);
-  if( rc!=SQLITE_OK ){
-    sqlite3_free(aReplay);
-    sqlite3_result_error_code(context, rc);
-    if( sealTopLevel ) (void)doltliteVcSealTopLevelSavepointTxn(db);
-    return rc;
-  }
-  savedInit = 1;
-
-  rc = doltliteLoadCommit(db, &upstreamHash, &upstreamCommit);
-  if( rc!=SQLITE_OK ) goto rollback;
-
-  rc = doltliteSwitchCatalog(db, &upstreamCommit.catalogHash);
-  if( rc!=SQLITE_OK ){
-    doltliteCommitClear(&upstreamCommit);
-    goto rollback;
-  }
-  doltliteSetSessionHead(db, &upstreamHash);
-  doltliteSetSessionStaged(db, &upstreamCommit.catalogHash);
-  rc = doltliteAdvanceBranch(db, &upstreamHash, &upstreamCommit.catalogHash, 0);
-  doltliteCommitClear(&upstreamCommit);
-  memset(&upstreamCommit, 0, sizeof(upstreamCommit));
-  if( rc!=SQLITE_OK ) goto rollback;
-
-  for(i=0; i<nReplay; i++){
-    DoltliteCommit replayCommit, parentCommit, curHeadCommit;
-    ProllyHash curHead;
-    int nConflicts = 0;
-    char hexBuf[PROLLY_HASH_SIZE*2+1];
-
-    memset(&replayCommit, 0, sizeof(replayCommit));
-    memset(&parentCommit, 0, sizeof(parentCommit));
-    memset(&curHeadCommit, 0, sizeof(curHeadCommit));
-
-    rc = doltliteLoadCommit(db, &aReplay[i], &replayCommit);
-    if( rc!=SQLITE_OK ) goto rollback;
-
-    sqlite3_free(zFailedMsg);
-    zFailedMsg = sqlite3_mprintf("%.200s",
-        replayCommit.zMessage ? replayCommit.zMessage : "");
-
-    if( doltliteCommitParentCount(&replayCommit)==0 ){
-      doltliteCommitClear(&replayCommit);
-      rc = SQLITE_ERROR;
-      goto rollback;
-    }
-    rc = doltliteLoadFirstParentCommit(db, &replayCommit, &parentCommit);
-    if( rc!=SQLITE_OK ){
-      doltliteCommitClear(&replayCommit);
-      goto rollback;
-    }
-
-    doltliteGetSessionHead(db, &curHead);
-    rc = doltliteLoadCommit(db, &curHead, &curHeadCommit);
-    if( rc!=SQLITE_OK ){
-      doltliteCommitClear(&replayCommit);
-      doltliteCommitClear(&parentCommit);
-      goto rollback;
-    }
-
-    rc = applyMergedCatalogAndCommit(db, context,
-        &parentCommit.catalogHash,
-        &curHeadCommit.catalogHash,
-        &replayCommit.catalogHash,
-        &curHead,
-        replayCommit.zMessage ? replayCommit.zMessage : "",
-        &nConflicts, hexBuf);
-
-    doltliteCommitClear(&replayCommit);
-    doltliteCommitClear(&parentCommit);
-    doltliteCommitClear(&curHeadCommit);
-
-    if( rc!=SQLITE_OK ) goto rollback;
-    if( nConflicts>0 ){ bConflict = 1; rc = SQLITE_ERROR; goto rollback; }
-  }
-
-  doltliteTxnStateClear(&saved);
-  sqlite3_free(aReplay);
-  sqlite3_free(zFailedMsg);
-  *pzFinalMessage = sqlite3_mprintf(
-    "Successfully rebased and updated refs/heads/%s",
-    doltliteGetSessionBranch(db));
-  return SQLITE_OK;
-
-rollback:
-  if( savedInit ){
-    (void)doltliteRestoreTxnStateOnFailure(db, &saved, rc);
-  }
-  (void)cs;
-  sqlite3_free(aReplay);
-  {
-    char *zErr;
-    if( bConflict && zFailedMsg && zFailedMsg[0] ){
-      zErr = sqlite3_mprintf(
-          "conflict rebasing \"%s\"; rebase aborted, branch restored to pre-rebase state",
-          zFailedMsg);
-    }else if( zFailedMsg && zFailedMsg[0] ){
-      zErr = sqlite3_mprintf(
-          "error rebasing \"%s\"; rebase aborted, branch restored to pre-rebase state",
-          zFailedMsg);
-    }else{
-      zErr = sqlite3_mprintf(
-          "rebase aborted, branch restored to pre-rebase state");
-    }
-    if( zErr ){
-      sqlite3_result_error(context, zErr, -1);
-      sqlite3_free(zErr);
-    }else{
-      sqlite3_result_error_code(context, rc);
-    }
-  }
-  sqlite3_free(zFailedMsg);
-  if( sealTopLevel ) (void)doltliteVcSealTopLevelSavepointTxn(db);
-  return SQLITE_ERROR;
-}
-
-typedef struct RebasePlanRow RebasePlanRow;
-struct RebasePlanRow {
-  double order;
-  char *zAction;
-  ProllyHash commitHash;
-  char *zCommitMessage;
-};
-
-enum DoltliteRebaseSchemaState {
-  DOLTLITE_REBASE_SCHEMA_ABSENT = 0,
-  DOLTLITE_REBASE_SCHEMA_OK = 1,
-  DOLTLITE_REBASE_SCHEMA_BAD = 2
-};
-
-static int doltliteRebaseSchemaState(sqlite3 *db, int *pState){
-  sqlite3_stmt *pStmt = 0;
-  int rc;
-  int nCol = 0;
-
-  *pState = DOLTLITE_REBASE_SCHEMA_ABSENT;
-  rc = sqlite3_prepare_v2(db, "PRAGMA main.table_info(\"dolt_rebase\")",
-                          -1, &pStmt, 0);
-  if( rc!=SQLITE_OK ) return rc;
-
-  while( (rc = sqlite3_step(pStmt))==SQLITE_ROW ){
-    const char *zName = (const char*)sqlite3_column_text(pStmt, 1);
-    const char *zType = (const char*)sqlite3_column_text(pStmt, 2);
-    int notNull = sqlite3_column_int(pStmt, 3);
-    int pkPos = sqlite3_column_int(pStmt, 5);
-    char aff;
-    if( !zName ) goto bad;
-    aff = sqlite3AffinityType(zType ? zType : "", 0);
-    if( nCol==0 ){
-      if( sqlite3_stricmp(zName, "rebase_order")!=0 ) goto bad;
-      if( aff!=SQLITE_AFF_REAL && aff!=SQLITE_AFF_NUMERIC ) goto bad;
-      if( !notNull ) goto bad;
-      if( pkPos!=1 ) goto bad;
-    }else if( nCol==1 ){
-      if( sqlite3_stricmp(zName, "action")!=0 ) goto bad;
-      if( aff!=SQLITE_AFF_TEXT ) goto bad;
-      if( pkPos!=0 ) goto bad;
-    }else if( nCol==2 ){
-      if( sqlite3_stricmp(zName, "commit_hash")!=0 ) goto bad;
-      if( aff!=SQLITE_AFF_TEXT ) goto bad;
-      if( pkPos!=0 ) goto bad;
-    }else if( nCol==3 ){
-      if( sqlite3_stricmp(zName, "commit_message")!=0 ) goto bad;
-      if( aff!=SQLITE_AFF_TEXT ) goto bad;
-      if( pkPos!=0 ) goto bad;
-    }else{
-      goto bad;
-    }
-    nCol++;
-  }
-  if( rc!=SQLITE_DONE ){
-    sqlite3_finalize(pStmt);
-    return rc;
-  }
-  sqlite3_finalize(pStmt);
-
-  if( nCol==0 ){
-    *pState = DOLTLITE_REBASE_SCHEMA_ABSENT;
-  }else if( nCol==4 ){
-    *pState = DOLTLITE_REBASE_SCHEMA_OK;
-  }else{
-    *pState = DOLTLITE_REBASE_SCHEMA_BAD;
-  }
-  return SQLITE_OK;
-
-bad:
-  *pState = DOLTLITE_REBASE_SCHEMA_BAD;
-  sqlite3_finalize(pStmt);
-  return SQLITE_OK;
-}
-
-static int doltliteValidateRebasePlanTable(sqlite3 *db, char **pzErr){
-  int rc;
-  int state;
-  if( pzErr ) *pzErr = 0;
-  rc = doltliteRebaseSchemaState(db, &state);
-  if( rc!=SQLITE_OK ) return rc;
-  if( state==DOLTLITE_REBASE_SCHEMA_OK ) return SQLITE_OK;
-  if( state==DOLTLITE_REBASE_SCHEMA_ABSENT ){
-    if( pzErr ) *pzErr = sqlite3_mprintf("no rebase plan found");
-    return SQLITE_NOTFOUND;
-  }
-  if( pzErr ){
-    *pzErr = sqlite3_mprintf(
-      "dolt_rebase has an unexpected schema; expected: "
-      "CREATE TABLE dolt_rebase("
-      "rebase_order REAL PRIMARY KEY, "
-      "action TEXT, "
-      "commit_hash TEXT, "
-      "commit_message TEXT)");
-  }
-  return SQLITE_CONSTRAINT;
-}
-
-static void rebaseFreePlan(RebasePlanRow *aPlan, int nPlan){
-  int i;
-  for(i=0; i<nPlan; i++){
-    sqlite3_free(aPlan[i].zAction);
-    sqlite3_free(aPlan[i].zCommitMessage);
-  }
-  sqlite3_free(aPlan);
-}
-
-static int rebaseReadPlan(sqlite3 *db, RebasePlanRow **paPlan, int *pnPlan){
-  sqlite3_stmt *pStmt = 0;
-  RebasePlanRow *aPlan = 0;
-  int nPlan = 0, nAlloc = 0;
-  int rc;
-  char *zErr = 0;
-
-  *paPlan = 0;
-  *pnPlan = 0;
-
-  rc = doltliteValidateRebasePlanTable(db, &zErr);
-  if( rc!=SQLITE_OK ){
-    sqlite3_free(zErr);
-    return rc;
-  }
-
-  rc = sqlite3_prepare_v2(db,
-    "SELECT rebase_order, action, commit_hash, commit_message "
-    "FROM main.dolt_rebase ORDER BY rebase_order", -1, &pStmt, 0);
-  if( rc!=SQLITE_OK ) return rc;
-
-  while( (rc = sqlite3_step(pStmt))==SQLITE_ROW ){
-    RebasePlanRow *r;
-    const char *zHex;
-
-    if( nPlan >= nAlloc ){
-      int nNew = nAlloc ? nAlloc*2 : 16;
-      RebasePlanRow *tmp = sqlite3_realloc(aPlan, nNew*(int)sizeof(RebasePlanRow));
-      if( !tmp ){ rc = SQLITE_NOMEM; goto fail; }
-      aPlan = tmp;
-      nAlloc = nNew;
-    }
-    r = &aPlan[nPlan];
-    memset(r, 0, sizeof(*r));
-    r->order = sqlite3_column_double(pStmt, 0);
-    r->zAction = sqlite3_mprintf("%s",
-        (const char*)sqlite3_column_text(pStmt, 1));
-    zHex = (const char*)sqlite3_column_text(pStmt, 2);
-    if( !zHex || doltliteHexToHash(zHex, &r->commitHash)!=SQLITE_OK ){
-      rc = SQLITE_CORRUPT;
-      sqlite3_free(r->zAction);
-      goto fail;
-    }
-    r->zCommitMessage = sqlite3_mprintf("%s",
-        (const char*)sqlite3_column_text(pStmt, 3));
-    if( !r->zAction || !r->zCommitMessage ){
-      rc = SQLITE_NOMEM;
-      goto fail;
-    }
-    nPlan++;
-  }
-  sqlite3_finalize(pStmt);
-  if( rc==SQLITE_DONE ) rc = SQLITE_OK;
-
-  *paPlan = aPlan;
-  *pnPlan = nPlan;
-  return SQLITE_OK;
-
-fail:
-  sqlite3_finalize(pStmt);
-  rebaseFreePlan(aPlan, nPlan);
-  return rc;
-}
-
-static char *rebaseBuildWorkingBranchName(const char *zOrigBranch){
-  if( !zOrigBranch ) return 0;
-  return sqlite3_mprintf("dolt_rebase_%s", zOrigBranch);
-}
-
-static int rebaseRestoreBranchState(sqlite3 *db, const char *zBranch){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  ProllyHash headHash;
-  ProllyHash emptyHash;
-  DoltliteCommit headCommit;
-  int rc;
-
-  if( !cs || !zBranch || !zBranch[0] ) return SQLITE_ERROR;
-  rc = chunkStoreFindBranch(cs, zBranch, &headHash);
-  if( rc!=SQLITE_OK ) return rc;
-  rc = doltliteLoadCommit(db, &headHash, &headCommit);
-  if( rc!=SQLITE_OK ) return rc;
-  rc = doltliteSwitchCatalog(db, &headCommit.catalogHash);
-  if( rc!=SQLITE_OK ){
-    doltliteCommitClear(&headCommit);
-    return rc;
-  }
-  doltliteSetSessionBranch(db, zBranch);
-  doltliteSetSessionHead(db, &headHash);
-  doltliteSetSessionStaged(db, &headCommit.catalogHash);
-  doltliteClearSessionMergeState(db);
-  memset(&emptyHash, 0, sizeof(emptyHash));
-  doltliteSetSessionConstraintViolationsCatalog(db, &emptyHash);
-  doltliteCommitClear(&headCommit);
-  return SQLITE_OK;
-}
-
-static int rebaseCreateAndPopulatePlanTable(
-  sqlite3 *db,
-  const ProllyHash *aReplay,
-  int nReplay
-){
-  int rc;
-  int i;
-
-  rc = sqlite3_exec(db,
-    "CREATE TABLE main.dolt_rebase("
-    "  rebase_order REAL PRIMARY KEY,"
-    "  action TEXT,"
-    "  commit_hash TEXT,"
-    "  commit_message TEXT"
-    ")", 0, 0, 0);
-  if( rc!=SQLITE_OK ) return rc;
-
-  for(i=0; i<nReplay; i++){
-    DoltliteCommit c;
-    char zHex[PROLLY_HASH_SIZE*2+1];
-    char zOrder[64];
-    char *zSql;
-    memset(&c, 0, sizeof(c));
-    rc = doltliteLoadCommit(db, &aReplay[i], &c);
-    if( rc!=SQLITE_OK ) break;
-    doltliteHashToHex(&aReplay[i], zHex);
-    sqlite3_snprintf(sizeof(zOrder), zOrder, "%!.17g", (double)(i + 1));
-    zSql = sqlite3_mprintf(
-      "INSERT INTO main.dolt_rebase VALUES (%s, 'pick', %Q, %Q)",
-      zOrder,
-      zHex,
-      c.zMessage ? c.zMessage : ""
-    );
-    if( !zSql ){
-      doltliteCommitClear(&c);
-      return SQLITE_NOMEM;
-    }
-    rc = sqlite3_exec(db, zSql, 0, 0, 0);
-    sqlite3_free(zSql);
-    doltliteCommitClear(&c);
-    if( rc!=SQLITE_OK ) break;
-  }
-  return rc;
-}
-
-static int rebaseApplyPlanRowCatalog(
-  sqlite3 *db,
-  const RebasePlanRow *pRow,
-  const ProllyHash *pCurCat,
-  ProllyHash *pMergedCat,
-  int bSkipConstraintDetect
-){
-  DoltliteCommit parentC, replayC;
-  int nConflicts = 0;
-  int nViolations = 0;
-  int rc;
-
-  memset(&parentC, 0, sizeof(parentC));
-  memset(&replayC, 0, sizeof(replayC));
-
-  rc = doltliteLoadCommit(db, &pRow->commitHash, &replayC);
-  if( rc!=SQLITE_OK ) return rc;
-  if( doltliteCommitParentCount(&replayC)==0 ){
-    doltliteCommitClear(&replayC);
-    return SQLITE_ERROR;
-  }
-  rc = doltliteLoadFirstParentCommit(db, &replayC, &parentC);
-  if( rc!=SQLITE_OK ){
-    doltliteCommitClear(&replayC);
-    return rc;
-  }
-
-  {
-    char **azReindex = 0;
-    int nReindex = 0;
-    rc = doltliteMergeCatalogs(db, &parentC.catalogHash, pCurCat,
-                               &replayC.catalogHash, pMergedCat,
-                               &nConflicts, 0, 0, 0, 0,
-                               &azReindex, &nReindex);
-    if( rc==SQLITE_OK && nConflicts==0 ){
-      rc = doltliteSwitchCatalog(db, pMergedCat);
-    }
-    if( rc==SQLITE_OK && nConflicts==0 && nReindex>0 ){
-      rc = doltliteReindexNamedIndexes(db, azReindex, nReindex);
-      if( rc==SQLITE_OK ) rc = doltliteFlushCatalogToHash(db, pMergedCat);
-      if( rc==SQLITE_OK ) rc = doltliteSwitchCatalog(db, pMergedCat);
-    }
-    doltliteFreeNameList(azReindex, nReindex);
-  }
-  if( rc==SQLITE_OK && nConflicts==0 && !bSkipConstraintDetect ){
-    rc = doltliteDetectPostMergeConstraintViolations(db,
-                                                     &parentC.catalogHash,
-                                                     &nViolations);
-  }
-  doltliteCommitClear(&parentC);
-  doltliteCommitClear(&replayC);
-  if( rc!=SQLITE_OK ) return rc;
-  if( nConflicts>0 || nViolations>0 ) return SQLITE_CONSTRAINT;
-  return SQLITE_OK;
-}
-
-static int rebaseAdvanceWorkingBranch(
-  sqlite3 *db,
-  const ProllyHash *pNewHead,
-  const ProllyHash *pCatalogHash
-){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  const char *zBranch = doltliteGetSessionBranch(db);
-  int rc;
-
-  if( !cs ) return SQLITE_ERROR;
-  rc = chunkStoreUpdateBranch(cs, zBranch, pNewHead);
-  if( rc!=SQLITE_OK ) return rc;
-
-  doltliteSetSessionHead(db, pNewHead);
-  doltliteSetSessionStaged(db, pCatalogHash);
-  rc = doltliteSwitchCatalog(db, pCatalogHash);
-  if( rc!=SQLITE_OK ) return rc;
-
-  return doltlitePersistWorkingSetWithHash(db, 0);
-}
-
-static int rebaseReplayPlanGroup(
-  sqlite3 *db,
-  RebasePlanRow *aPlan,
-  int nPlan,
-  int iStart,
-  ProllyHash *pCurCat,
-  ProllyHash *pCurHead,
-  int *piNext,
-  int bSkipConstraintDetect
-){
-  char *combinedMsg = 0;
-  int rc;
-  int j;
-
-  rc = rebaseApplyPlanRowCatalog(db, &aPlan[iStart], pCurCat, pCurCat,
-                                 bSkipConstraintDetect);
-  if( rc!=SQLITE_OK ) return rc;
-
-  combinedMsg = sqlite3_mprintf("%s",
-      aPlan[iStart].zCommitMessage ? aPlan[iStart].zCommitMessage : "");
-  if( !combinedMsg ) return SQLITE_NOMEM;
-
-  j = iStart + 1;
-  while( j < nPlan
-      && (strcmp(aPlan[j].zAction, "squash")==0
-       || strcmp(aPlan[j].zAction, "fixup")==0
-       || strcmp(aPlan[j].zAction, "drop")==0) ){
-    if( strcmp(aPlan[j].zAction, "drop")==0 ){
-      j++;
-      continue;
-    }
-
-    rc = rebaseApplyPlanRowCatalog(db, &aPlan[j], pCurCat, pCurCat,
-                                   bSkipConstraintDetect);
-    if( rc!=SQLITE_OK ){
-      sqlite3_free(combinedMsg);
-      return rc;
-    }
-
-    if( strcmp(aPlan[j].zAction, "squash")==0 ){
-      char *zNew = sqlite3_mprintf("%s\n\n%s", combinedMsg,
-                                   aPlan[j].zCommitMessage ? aPlan[j].zCommitMessage : "");
-      sqlite3_free(combinedMsg);
-      combinedMsg = zNew;
-      if( !combinedMsg ) return SQLITE_NOMEM;
-    }
-    j++;
-  }
-
-  {
-    ProllyHash newCommit;
-    rc = doltliteCreateAndStoreCommit(db, pCurHead, pCurCat, combinedMsg,
-                                      NULL, NULL, NULL, 0, &newCommit);
-    if( rc==SQLITE_OK ){
-      rc = doltliteSwitchCatalog(db, pCurCat);
-    }
-    if( rc==SQLITE_OK ){
-      doltliteSetSessionStaged(db, pCurCat);
-    }
-    if( rc==SQLITE_OK ){
-      rc = rebaseAdvanceWorkingBranch(db, &newCommit, pCurCat);
-    }
-    if( rc==SQLITE_OK ) *pCurHead = newCommit;
-  }
-  sqlite3_free(combinedMsg);
-  if( rc!=SQLITE_OK ) return rc;
-
-  *piNext = j;
-  return SQLITE_OK;
-}
-
-typedef struct RebaseFinalizeRefsCtx RebaseFinalizeRefsCtx;
-struct RebaseFinalizeRefsCtx {
-  const char *zOrigBranch;
-  const char *zWorkingBranch;
-  const ProllyHash *pCurHead;
-  const ProllyHash *pCurCat;
-};
-
-static int rebaseFinalizeContinueRefs(sqlite3 *db, ChunkStore *cs, void *pArg){
-  RebaseFinalizeRefsCtx *p = (RebaseFinalizeRefsCtx*)pArg;
-  int rc;
-  UNUSED_PARAMETER(db);
-  rc = chunkStoreUpdateBranch(cs, p->zOrigBranch, p->pCurHead);
-  if( rc!=SQLITE_OK ) return rc;
-  rc = chunkStoreWriteBranchWorkingCatalog(cs, p->zOrigBranch,
-                                           p->pCurCat, p->pCurHead);
-  if( rc!=SQLITE_OK ) return rc;
-  return SQLITE_OK;
-}
-
-static int rebaseDeleteWorkingBranchRefs(sqlite3 *db, ChunkStore *cs, void *pArg){
-  const char *zWorkingBranch = (const char*)pArg;
-  UNUSED_PARAMETER(db);
-  return chunkStoreDeleteBranch(cs, zWorkingBranch);
-}
-
-static void rebaseDiscardWorkingBranch(
-  sqlite3 *db,
-  const char *zOrigBranch,
-  const char *zWorkingBranch
-){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-
-  (void)sqlite3_exec(db, "DROP TABLE IF EXISTS main.dolt_rebase", 0, 0, 0);
-  doltliteClearSessionRebaseState(db);
-  (void)doltlitePersistWorkingSet(db);
-
-  if( zOrigBranch && zOrigBranch[0] ){
-    if( doltliteCheckoutBranchForRebase(db, zOrigBranch)!=SQLITE_OK ){
-      (void)rebaseRestoreBranchState(db, zOrigBranch);
-    }
-  }
-  if( cs && zWorkingBranch && zWorkingBranch[0] ){
-    (void)chunkStoreDeleteBranch(cs, zWorkingBranch);
-    (void)chunkStoreSerializeRefs(cs);
-    (void)chunkStoreCommit(cs);
-    (void)doltlitePersistWorkingSet(db);
-  }
-}
-
-static void rebaseAbortConflictedContinue(
-  sqlite3 *db,
-  const char *zOrigBranch,
-  const char *zReturnBranch,
-  const char *zWorkingBranch
-){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-
-  (void)sqlite3_exec(db, "DROP TABLE IF EXISTS main.dolt_rebase", 0, 0, 0);
-  doltliteClearSessionRebaseState(db);
-  doltliteClearSessionMergeState(db);
-  if( zOrigBranch && zOrigBranch[0] ){
-    (void)rebaseRestoreBranchState(db, zOrigBranch);
-    doltliteClearSessionRebaseState(db);
-  }
-  if( cs && zReturnBranch && zReturnBranch[0] ){
-    (void)rebaseRestoreReturnBranchWorkingState(db, zReturnBranch);
-  }
-  if( cs && zWorkingBranch && zWorkingBranch[0] ){
-    (void)chunkStoreDeleteBranch(cs, zWorkingBranch);
-    (void)chunkStoreSerializeRefs(cs);
-    (void)chunkStoreCommit(cs);
-    (void)doltlitePersistWorkingSet(db);
-  }
-}
-
-static void doltliteRebaseInteractiveStart(
-  sqlite3_context *context,
-  sqlite3 *db,
-  const char *zUpstream
-){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  char *zOrig = 0;
-  char *zReturnBranch = 0;
-  char *zWorking = 0;
-  ProllyHash upstreamHash, headHash;
-  ProllyHash preRebaseCat;
-  ProllyHash *aReplay = 0;
-  int nReplay = 0;
-  int rc;
-  u8 curIsRebasing = 0;
-  int bWorkingBranchCreated = 0;
-  const char *zFailMsg = 0;
-
-  memset(&preRebaseCat, 0, sizeof(preRebaseCat));
-
-  doltliteGetSessionRebaseState(db, &curIsRebasing, 0, 0, 0, 0);
-  if( curIsRebasing ){
-    sqlite3_result_error(context,
-      "rebase already in progress; use --continue or --abort", -1);
-    return;
-  }
-
-  rc = doltliteEnsureWriteTxnAndSavepoints(db);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-    return;
-  }
-
-  if( doltliteHasUncommittedChanges(db) ){
-    sqlite3_result_error(context,
-      "cannot start a rebase with uncommitted changes", -1);
-    return;
-  }
-
-  rc = doltliteResolveRef(db, zUpstream, &upstreamHash);
-  if( rc!=SQLITE_OK ){
-    char *zErr = sqlite3_mprintf("branch not found: %s", zUpstream);
-    sqlite3_result_error(context, zErr ? zErr : "branch not found", -1);
-    sqlite3_free(zErr);
-    return;
-  }
-
-  doltliteGetSessionHead(db, &headHash);
-  if( prollyHashIsEmpty(&headHash) ){
-    sqlite3_result_error(context, "no commits on current branch", -1);
-    return;
-  }
-
-  rc = doltliteRebaseCollectReplaySet(db, &headHash, &upstreamHash,
-                                      &aReplay, &nReplay);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(context, rc);
-    return;
-  }
-  if( nReplay==0 ){
-    sqlite3_free(aReplay);
-    sqlite3_result_error(context, "didn't identify any commits!", -1);
-    return;
-  }
-
-  zOrig = sqlite3_mprintf("%s", doltliteGetSessionBranch(db));
-  zReturnBranch = sqlite3_mprintf("%s", chunkStoreGetDefaultBranch(cs));
-  zWorking = sqlite3_mprintf("dolt_rebase_%s", zOrig ? zOrig : "");
-  if( !zOrig || !zWorking || !zReturnBranch ){
-    sqlite3_free(zOrig);
-    sqlite3_free(zReturnBranch);
-    sqlite3_free(zWorking);
-    sqlite3_free(aReplay);
-    sqlite3_result_error_code(context, SQLITE_NOMEM);
-    return;
-  }
-  {
-    ProllyHash probe;
-    if( chunkStoreFindBranch(cs, zWorking, &probe)==SQLITE_OK ){
-      sqlite3_free(zOrig);
-      sqlite3_free(zReturnBranch);
-      sqlite3_free(zWorking);
-      sqlite3_free(aReplay);
-      sqlite3_result_error(context,
-        "rebase working branch already exists", -1);
-      return;
-    }
-  }
-
-  rc = doltliteFlushCatalogToHash(db, &preRebaseCat);
-  if( rc!=SQLITE_OK ) goto fail;
-
-  rc = chunkStoreAddBranch(cs, zWorking, &upstreamHash);
-  if( rc!=SQLITE_OK ){
-    zFailMsg = "rebase working branch already exists";
-    goto fail;
-  }
-  bWorkingBranchCreated = 1;
-  /* Persist the working branch now. The sub-ops below (checkout, plan table)
-  ** each reload persisted refs at their lock — the VC fresh-read that closes
-  ** the concurrency window — which would discard this still-in-memory branch.
-  ** Committing it makes the reload re-read it instead. */
-  rc = chunkStoreSerializeRefs(cs);
-  if( rc==SQLITE_OK ) rc = chunkStoreCommit(cs);
-  if( rc!=SQLITE_OK ) goto fail;
-  rc = doltliteCheckoutBranchForRebase(db, zWorking);
-  if( rc!=SQLITE_OK ) goto fail;
-
-  rc = rebaseCreateAndPopulatePlanTable(db, aReplay, nReplay);
-  if( rc!=SQLITE_OK ){
-    zFailMsg = "failed to create dolt_rebase table";
-    goto fail;
-  }
-
-  doltliteSetSessionRebaseState(db, 1, &preRebaseCat, &upstreamHash,
-                                zOrig, zReturnBranch);
-  rc = doltlitePersistWorkingSet(db);
-  if( rc!=SQLITE_OK ) goto fail;
-  rc = doltliteVcSealBranchStyleTxn(db);
-  if( rc!=SQLITE_OK ) goto fail;
-  sqlite3ExpirePreparedStatements(db, 0);
-  sqlite3ResetAllSchemasOfConnection(db);
-
-  sqlite3_free(zOrig);
-  sqlite3_free(zReturnBranch);
-  sqlite3_free(aReplay);
-  {
-    char *zMsg = sqlite3_mprintf(
-      "interactive rebase started on branch %s; adjust the rebase plan "
-      "in the dolt_rebase table, then continue rebasing by calling "
-      "dolt_rebase('--continue')", zWorking);
-    sqlite3_free(zWorking);
-    if( zMsg ) sqlite3_result_text(context, zMsg, -1, sqlite3_free);
-    else sqlite3_result_text(context, "interactive rebase started", -1, SQLITE_STATIC);
-  }
-  return;
-
-fail:
-  if( bWorkingBranchCreated ){
-    rebaseDiscardWorkingBranch(db, zOrig, zWorking);
-  }
-  sqlite3_free(zOrig);
-  sqlite3_free(zReturnBranch);
-  sqlite3_free(zWorking);
-  sqlite3_free(aReplay);
-  if( zFailMsg ){
-    sqlite3_result_error(context, zFailMsg, -1);
-  }else{
-    sqlite3_result_error_code(context, rc);
-  }
-}
-
-static void doltliteRebaseInteractiveAbort(
-  sqlite3_context *context,
-  sqlite3 *db
-){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  u8 isRebasing = 0;
-  const char *zOrigBranchConst = 0;
-  const char *zReturnBranchConst = 0;
-  char *zReturnBranch = 0;
-  char *zOrigBranch = 0;
-  char *zWorking = 0;
-  int rc;
-
-  doltliteGetSessionRebaseState(db, &isRebasing, 0, 0,
-                                &zOrigBranchConst, &zReturnBranchConst);
-  if( !isRebasing || !zOrigBranchConst || !zReturnBranchConst ){
-    sqlite3_result_error(context, "no rebase in progress", -1);
-    return;
-  }
-  zOrigBranch = sqlite3_mprintf("%s", zOrigBranchConst);
-  zReturnBranch = sqlite3_mprintf("%s", zReturnBranchConst);
-  zWorking = rebaseBuildWorkingBranchName(zOrigBranchConst);
-  if( !zReturnBranch || !zWorking || !zOrigBranch ){
-    sqlite3_free(zReturnBranch);
-    sqlite3_free(zWorking);
-    sqlite3_free(zOrigBranch);
-    sqlite3_result_error_code(context, SQLITE_NOMEM);
-    return;
-  }
-
-  rebaseDiscardWorkingBranch(db, zOrigBranch, zWorking);
-  if( cs && zReturnBranch && zReturnBranch[0] ){
-    rc = rebaseRestoreReturnBranchWorkingState(db, zReturnBranch);
-    if( rc!=SQLITE_OK ){
-      sqlite3_free(zReturnBranch);
-      sqlite3_free(zWorking);
-      sqlite3_free(zOrigBranch);
-      sqlite3_result_error_code(context, rc);
-      return;
-    }
-  }
-  rc = doltlitePersistWorkingSet(db);
-  if( rc!=SQLITE_OK ){
-    sqlite3_free(zReturnBranch);
-    sqlite3_free(zWorking);
-    sqlite3_free(zOrigBranch);
-    sqlite3_result_error_code(context, rc);
-    return;
-  }
-
-  rc = doltliteVcSealBranchStyleTxn(db);
-  if( rc!=SQLITE_OK ){
-    sqlite3_free(zReturnBranch);
-    sqlite3_free(zWorking);
-    sqlite3_free(zOrigBranch);
-    sqlite3_result_error_code(context, rc);
-    return;
-  }
-
-  sqlite3_free(zReturnBranch);
-  sqlite3_free(zWorking);
-  sqlite3_free(zOrigBranch);
-  sqlite3_result_text(context, "Interactive rebase aborted", -1, SQLITE_STATIC);
-}
-
-static void doltliteRebaseInteractiveContinue(
-  sqlite3_context *context,
-  sqlite3 *db
-){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  u8 isRebasing = 0;
-  const char *zOrigBranchConst = 0;
-  const char *zReturnBranchConst = 0;
-  char *zReturnBranch = 0;
-  char *zOrigBranch = 0;
-  char *zWorking = 0;
-  RebasePlanRow *aPlan = 0;
-  int nPlan = 0;
-  int rc;
-  int i;
-  int bPlanDropped = 0;
-  int bSkipConstraintDetect = (!db->autoCommit || db->pSavepoint!=0);
-  ProllyHash curCat;
-  ProllyHash curHead;
-  RebaseFinalizeRefsCtx refsCtx;
-  char *zPlanErr = 0;
-
-  memset(&curCat, 0, sizeof(curCat));
-  memset(&curHead, 0, sizeof(curHead));
-
-  doltliteGetSessionRebaseState(db, &isRebasing, 0, 0,
-                                &zOrigBranchConst, &zReturnBranchConst);
-  if( !isRebasing || !zOrigBranchConst || !zReturnBranchConst ){
-    sqlite3_result_error(context, "no rebase in progress", -1);
-    return;
-  }
-  zOrigBranch = sqlite3_mprintf("%s", zOrigBranchConst);
-  zReturnBranch = sqlite3_mprintf("%s", zReturnBranchConst);
-  zWorking = rebaseBuildWorkingBranchName(zOrigBranchConst);
-  if( !zReturnBranch || !zWorking || !zOrigBranch ){ rc = SQLITE_NOMEM; goto abort_err; }
-
-  (void)sqlite3_exec(db, "SELECT 1 FROM main.dolt_rebase LIMIT 0", 0, 0, 0);
-
-  rc = doltliteValidateRebasePlanTable(db, &zPlanErr);
-  if( rc!=SQLITE_OK ){
-    if( zPlanErr ) sqlite3_result_error(context, zPlanErr, -1);
-    sqlite3_free(zPlanErr);
-    goto abort_err_silent;
-  }
-
-  rc = rebaseReadPlan(db, &aPlan, &nPlan);
-  if( rc!=SQLITE_OK ) goto abort_err;
-
-  rc = doltliteVcSealBranchStyleTxnMaybeKeepTopLevelSavepoint(db);
-  if( rc!=SQLITE_OK ) goto abort_err;
-
-  rc = doltliteEnsureWriteTxnAndSavepoints(db);
-  if( rc!=SQLITE_OK ) goto abort_err;
-
-  /* Reject any unknown plan verb (e.g. a typo in dolt_rebase.action) instead
-  ** of silently treating it as a pick; abort_err restores the pre-rebase
-  ** branch and reports the failure. */
-  for(i=0; i<nPlan; i++){
-    const char *zAct = aPlan[i].zAction;
-    if( strcmp(zAct,"pick")!=0 && strcmp(zAct,"reword")!=0
-     && strcmp(zAct,"squash")!=0 && strcmp(zAct,"fixup")!=0
-     && strcmp(zAct,"drop")!=0 ){
-      rc = SQLITE_ERROR;
-      goto abort_err;
-    }
-  }
-
-  i = 0;
-  while( i < nPlan && strcmp(aPlan[i].zAction, "drop")==0 ) i++;
-  if( i < nPlan
-   && strcmp(aPlan[i].zAction, "pick")!=0
-   && strcmp(aPlan[i].zAction, "reword")!=0 ){
-    rc = SQLITE_ERROR;
-    sqlite3_result_error(context,
-      "first non-drop action must be pick or reword", -1);
-    goto abort_err_silent;
-  }
-
-  rc = sqlite3_exec(db, "DROP TABLE main.dolt_rebase", 0, 0, 0);
-  if( rc!=SQLITE_OK ) goto abort_err;
-  bPlanDropped = 1;
-
-  rc = doltliteFlushCatalogToHash(db, &curCat);
-  if( rc!=SQLITE_OK ) goto abort_err;
-  doltliteGetSessionHead(db, &curHead);
-
-  i = 0;
-  while( i < nPlan ){
-    int j;
-
-    while( i < nPlan && strcmp(aPlan[i].zAction, "drop")==0 ) i++;
-    if( i >= nPlan ) break;
-
-    rc = rebaseReplayPlanGroup(db, aPlan, nPlan, i, &curCat, &curHead, &j,
-                               bSkipConstraintDetect);
-    if( rc==SQLITE_CONSTRAINT ) goto abort_err_conflict;
-    if( rc!=SQLITE_OK ) goto abort_err;
-    i = j;
-  }
-
-  memset(&refsCtx, 0, sizeof(refsCtx));
-  refsCtx.zOrigBranch = zOrigBranch;
-  refsCtx.zWorkingBranch = zWorking;
-  refsCtx.pCurHead = &curHead;
-  refsCtx.pCurCat = &curCat;
-  rc = doltliteMutateRefs(db, rebaseFinalizeContinueRefs, &refsCtx);
-  if( rc!=SQLITE_OK ) goto abort_err;
-
-  doltliteClearSessionRebaseState(db);
-  rc = doltlitePersistWorkingSet(db);
-  if( rc!=SQLITE_OK ) goto abort_err;
-
-  rc = doltliteCheckoutBranchForRebase(db, zOrigBranch);
-  if( rc!=SQLITE_OK ) goto abort_err;
-  rc = doltliteMutateRefs(db, rebaseDeleteWorkingBranchRefs, zWorking);
-  if( rc!=SQLITE_OK ) goto abort_err;
-  rc = rebaseRestoreReturnBranchWorkingState(db, zReturnBranch);
-  if( rc!=SQLITE_OK ) goto abort_err;
-  rc = doltlitePersistWorkingSet(db);
-  if( rc!=SQLITE_OK ) goto abort_err;
-  rc = doltliteVcSealBranchStyleTxnMaybeKeepTopLevelSavepoint(db);
-  if( rc!=SQLITE_OK ) goto abort_err;
-
-  rebaseFreePlan(aPlan, nPlan);
-  {
-    char *zMsg = sqlite3_mprintf(
-      "Successfully rebased and updated refs/heads/%s", zOrigBranch);
-    sqlite3_free(zOrigBranch);
-    sqlite3_free(zReturnBranch);
-    sqlite3_free(zWorking);
-    if( zMsg ) sqlite3_result_text(context, zMsg, -1, sqlite3_free);
-    else sqlite3_result_text(context, "Successfully rebased", -1, SQLITE_STATIC);
-  }
-  return;
-
-abort_err_conflict:
-  rebaseFreePlan(aPlan, nPlan);
-  rebaseAbortConflictedContinue(db, zOrigBranch, zReturnBranch, zWorking);
-  if( doltliteVcTxnMode(db)==DOLTLITE_VC_TXN_AUTOCOMMIT_LIKE ){
-    (void)sqlite3_exec(db, "COMMIT", 0, 0, 0);
-  }
-  sqlite3_free(zOrigBranch);
-  sqlite3_free(zReturnBranch);
-  sqlite3_free(zWorking);
-  sqlite3_result_error(context,
-    "data conflicts from rebase — rebase has been aborted", -1);
-  return;
-
-abort_err:
-  rebaseFreePlan(aPlan, nPlan);
-  if( bPlanDropped ){
-    rebaseDiscardWorkingBranch(db, zOrigBranch ? zOrigBranch : "main", zWorking);
-    if( cs && zReturnBranch && zReturnBranch[0] ){
-      (void)rebaseRestoreReturnBranchWorkingState(db, zReturnBranch);
-    }
-  }
-  sqlite3_free(zOrigBranch);
-  sqlite3_free(zReturnBranch);
-  sqlite3_free(zWorking);
-  sqlite3_result_error(context, "rebase failed — branch restored to pre-rebase state", -1);
-  return;
-
-abort_err_silent:
-  rebaseFreePlan(aPlan, nPlan);
-  sqlite3_free(zOrigBranch);
-  sqlite3_free(zReturnBranch);
-  sqlite3_free(zWorking);
-}
-
-static void doltliteRebaseFunc(
-  sqlite3_context *context,
-  int argc,
-  sqlite3_value **argv
-){
-  sqlite3 *db = sqlite3_context_db_handle(context);
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  const char *zArg0;
-  int sealTopLevel = db->pSavepoint!=0 && db->nSavepoint==0;
-  int keepTopLevelSavepoint = 0;
-
-  if( !cs ){ sqlite3_result_error(context, "no database", -1); goto rebase_cleanup; }
-  if( argc<1 ){
-    sqlite3_result_error(context, "usage: dolt_rebase('upstream_branch')", -1);
-    goto rebase_cleanup;
-  }
-
-  zArg0 = (const char*)sqlite3_value_text(argv[0]);
-  if( !zArg0 ){
-    sqlite3_result_error(context, "upstream ref required", -1);
-    goto rebase_cleanup;
-  }
-
-  if( strcmp(zArg0, "--abort")==0 ){
-    doltliteRebaseInteractiveAbort(context, db);
-    goto rebase_cleanup;
-  }
-  if( strcmp(zArg0, "--continue")==0 ){
-    keepTopLevelSavepoint = 1;
-    doltliteRebaseInteractiveContinue(context, db);
-    goto rebase_cleanup;
-  }
-  if( strcmp(zArg0, "-i")==0 || strcmp(zArg0, "--interactive")==0 ){
-    const char *zUpstream;
-    keepTopLevelSavepoint = 1;
-    if( argc<2 ){
-      sqlite3_result_error(context,
-        "interactive rebase requires upstream branch: "
-        "dolt_rebase('-i', 'upstream')", -1);
-      goto rebase_cleanup;
-    }
-    if( argc!=2 ){
-      sqlite3_result_error(context,
-        "interactive rebase takes exactly one upstream branch", -1);
-      goto rebase_cleanup;
-    }
-    zUpstream = (const char*)sqlite3_value_text(argv[1]);
-    if( !zUpstream ){
-      sqlite3_result_error(context, "upstream ref required", -1);
-      goto rebase_cleanup;
-    }
-    doltliteRebaseInteractiveStart(context, db, zUpstream);
-    goto rebase_cleanup;
-  }
-
-  if( zArg0[0]=='-' ){
-    char *zErr = sqlite3_mprintf("unknown option `%s`", zArg0);
-    if( zErr ){
-      sqlite3_result_error(context, zErr, -1);
-      sqlite3_free(zErr);
-    }else{
-      sqlite3_result_error_nomem(context);
-    }
-    goto rebase_cleanup;
-  }
-  if( argc!=1 ){
-    sqlite3_result_error(context,
-      "too many positional arguments to dolt_rebase", -1);
-    goto rebase_cleanup;
-  }
-
-  {
-    char *zFinalMessage = 0;
-    int rc = doltliteRebaseLinearReplay(db, context, zArg0, &zFinalMessage);
-    if( rc==SQLITE_OK && zFinalMessage ){
-      sqlite3_result_text(context, zFinalMessage, -1, sqlite3_free);
-    }
-  }
-
-rebase_cleanup:
-  if( sealTopLevel && !keepTopLevelSavepoint ){
-    (void)doltliteVcSealTopLevelSavepointTxn(db);
-  }
-}
-
-static void doltliteConfigFunc(sqlite3_context *context, int argc, sqlite3_value **argv){
-  sqlite3 *db = sqlite3_context_db_handle(context);
-  const char *zKey;
-
-  if( argc<1 ){
-    sqlite3_result_error(context, "usage: dolt_config(key [, value])", -1);
-    return;
-  }
-  if( argc>2 ){
-    sqlite3_result_error(context,
-      "too many positional arguments to dolt_config", -1);
-    return;
-  }
-  zKey = (const char*)sqlite3_value_text(argv[0]);
-  if( !zKey ){
-    sqlite3_result_error(context, "key required", -1);
-    return;
-  }
-
-  if( argc==1 ){
-
-    if( strcmp(zKey, "user.name")==0 ){
-      sqlite3_result_text(context, doltliteGetAuthorName(db), -1, SQLITE_TRANSIENT);
-    }else if( strcmp(zKey, "user.email")==0 ){
-      sqlite3_result_text(context, doltliteGetAuthorEmail(db), -1, SQLITE_TRANSIENT);
-    }else{
-      sqlite3_result_error(context, "unknown config key (valid: user.name, user.email)", -1);
-    }
-  }else{
-
-    const char *zVal = (const char*)sqlite3_value_text(argv[1]);
-    if( strcmp(zKey, "user.name")==0 ){
-      doltliteSetAuthorName(db, zVal);
-      sqlite3_result_int(context, 0);
-    }else if( strcmp(zKey, "user.email")==0 ){
-      doltliteSetAuthorEmail(db, zVal);
-      sqlite3_result_int(context, 0);
-    }else{
-      sqlite3_result_error(context, "unknown config key (valid: user.name, user.email)", -1);
-    }
-  }
-}
-
-static void doltliteVersionFunc(sqlite3_context *ctx, int argc, sqlite3_value **argv){
-  (void)argv;
-  if( argc!=0 ){
-    sqlite3_result_error(ctx,
-        "dolt_version() takes exactly zero arguments", -1);
-    return;
-  }
-  sqlite3_result_text(ctx, DOLTLITE_VERSION, -1, SQLITE_STATIC);
-}
-
-static int mutateDefaultBranch(sqlite3 *db, ChunkStore *cs, void *pArg){
-  const char *zName = (const char*)pArg;
-  ProllyHash unused;
-  int rc;
-  (void)db;
-  rc = chunkStoreFindBranch(cs, zName, &unused);
-  if( rc!=SQLITE_OK ) return rc;
-  return chunkStoreSetDefaultBranch(cs, zName);
-}
-
-static void doltliteDefaultBranchFunc(
-  sqlite3_context *ctx, int argc, sqlite3_value **argv
-){
-  sqlite3 *db = sqlite3_context_db_handle(ctx);
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  if( !cs ){
-    sqlite3_result_error(ctx, "no chunk store", -1);
-    return;
-  }
-  if( argc==0 ){
-    const char *zDef = chunkStoreGetDefaultBranch(cs);
-    sqlite3_result_text(ctx, zDef ? zDef : "main", -1, SQLITE_TRANSIENT);
-    return;
-  }
-  if( argc==1 ){
-    const char *zNew;
-    int rc;
-    if( sqlite3_value_type(argv[0])!=SQLITE_TEXT ){
-      sqlite3_result_error(ctx,
-        "dolt_default_branch(name): name must be text", -1);
-      return;
-    }
-    zNew = (const char*)sqlite3_value_text(argv[0]);
-    if( !zNew || !zNew[0] ){
-      sqlite3_result_error(ctx, "branch name required", -1);
-      return;
-    }
-    rc = doltliteMutateRefs(db, mutateDefaultBranch, (void*)zNew);
-    if( rc==SQLITE_NOTFOUND ){
-      char *zErr = sqlite3_mprintf("branch '%s' not found", zNew);
-      sqlite3_result_error(ctx, zErr ? zErr : "branch not found", -1);
-      sqlite3_free(zErr);
-      return;
-    }
-    if( rc!=SQLITE_OK ){
-      sqlite3_result_error_code(ctx, rc);
-      return;
-    }
-    sqlite3_result_int(ctx, 0);
-    return;
-  }
-  sqlite3_result_error(ctx,
-    "dolt_default_branch() takes 0 or 1 arguments", -1);
-}
-
-static void doltliteInternalMaterializeDefaultColumnFunc(
-  sqlite3_context *ctx,
-  int argc,
-  sqlite3_value **argv
-){
-  sqlite3 *db = sqlite3_context_db_handle(ctx);
-  const char *zDb;
-  const char *zTable;
-  const char *zColumn;
-  char *zSql;
-  int rc;
-
-  if( argc!=3 ){
-    sqlite3_result_error(ctx,
-        "doltlite_internal_materialize_default_column() takes 3 arguments", -1);
-    return;
-  }
-  zDb = (const char*)sqlite3_value_text(argv[0]);
-  zTable = (const char*)sqlite3_value_text(argv[1]);
-  zColumn = (const char*)sqlite3_value_text(argv[2]);
-  if( !zDb || !zTable || !zColumn ){
-    sqlite3_result_null(ctx);
-    return;
-  }
-
-  zSql = sqlite3_mprintf("UPDATE \"%w\".\"%w\" SET \"%w\"=\"%w\"",
-                         zDb, zTable, zColumn, zColumn);
-  if( !zSql ){
-    sqlite3_result_error_nomem(ctx);
-    return;
-  }
-  rc = sqlite3_exec(db, zSql, 0, 0, 0);
-  sqlite3_free(zSql);
-  if( rc!=SQLITE_OK ){
-    sqlite3_result_error_code(ctx, rc);
-    return;
-  }
-  sqlite3_result_int(ctx, 0);
-}
-
-static int doltliteMaybeSeedRepo(sqlite3 *db){
-  ChunkStore *cs = doltliteGetChunkStore(db);
-  ProllyHash emptyParent;
-  ProllyHash emptyCatalog;
-  ProllyHash seedHash;
-  int rc;
-
-  if( !cs ) return SQLITE_OK;
-  if( refsTableBranchCount(&cs->refs) > 0 ) return SQLITE_OK;
-  if( sqlite3_db_readonly(db, "main")==1 ) return SQLITE_OK;
-
-  memset(&emptyParent, 0, sizeof(emptyParent));
-  memset(&emptyCatalog, 0, sizeof(emptyCatalog));
-
-  rc = doltliteCreateAndStoreCommit(db, &emptyParent, &emptyCatalog,
-      "Initialize data repository", NULL, NULL, 0, 0, &seedHash);
-  if( rc!=SQLITE_OK ) return rc;
-
-  return doltliteAdvanceBranch(db, &seedHash, &emptyCatalog, 0);
-}
+extern int doltliteHashofRegister(sqlite3 *db);
+extern int doltliteConstraintViolationsRegister(sqlite3 *db);
 
 int doltliteRegister(sqlite3 *db){
   int rc;
-  rc = sqlite3_create_function(db, "dolt_commit", -1,
-                               DOLTLITE_COMMAND_FUNC_FLAGS, 0,
-                               doltliteCommitFunc, 0, 0);
-  if( rc==SQLITE_OK ) rc = sqlite3_create_function(db, "dolt_add", -1,
-                                                   DOLTLITE_COMMAND_FUNC_FLAGS, 0,
-                                                   doltliteAddFunc, 0, 0);
-  if( rc==SQLITE_OK ) rc = sqlite3_create_function(db, "dolt_reset", -1,
-                                                   DOLTLITE_COMMAND_FUNC_FLAGS, 0,
-                                                   doltliteResetFunc, 0, 0);
-  if( rc==SQLITE_OK ) rc = sqlite3_create_function(db, "dolt_merge", -1,
-                                                   DOLTLITE_COMMAND_FUNC_FLAGS, 0,
-                                                   doltliteMergeFunc, 0, 0);
-  if( rc==SQLITE_OK ) rc = sqlite3_create_function(db, "dolt_cherry_pick", -1,
-                                                   DOLTLITE_COMMAND_FUNC_FLAGS, 0,
-                                                   doltliteCherryPickFunc, 0, 0);
-  if( rc==SQLITE_OK ) rc = sqlite3_create_function(db, "dolt_revert", -1,
-                                                   DOLTLITE_COMMAND_FUNC_FLAGS, 0,
-                                                   doltliteRevertFunc, 0, 0);
-  if( rc==SQLITE_OK ) rc = sqlite3_create_function(db, "dolt_rebase", -1,
-                                                   DOLTLITE_COMMAND_FUNC_FLAGS, 0,
-                                                   doltliteRebaseFunc, 0, 0);
-  if( rc==SQLITE_OK ) rc = sqlite3_create_function(db, "dolt_config", -1,
-                                                   DOLTLITE_COMMAND_FUNC_FLAGS, 0,
-                                                   doltliteConfigFunc, 0, 0);
-  if( rc==SQLITE_OK ) rc = sqlite3_create_function(db, "dolt_version", 0, SQLITE_UTF8, 0,
-                                                   doltliteVersionFunc, 0, 0);
-  if( rc==SQLITE_OK ) rc = sqlite3_create_function(db, "dolt_default_branch", -1,
-                                                   DOLTLITE_COMMAND_FUNC_FLAGS, 0,
-                                                   doltliteDefaultBranchFunc, 0, 0);
-  if( rc==SQLITE_OK ) rc = sqlite3_create_function(db,
-                                                   "doltlite_internal_materialize_default_column",
-                                                   3, DOLTLITE_COMMAND_FUNC_FLAGS, 0,
-                                                   doltliteInternalMaterializeDefaultColumnFunc,
-                                                   0, 0);
-  if( rc!=SQLITE_OK ) return rc;
+  if( (rc = doltliteCommitCmdRegister(db))!=SQLITE_OK ) return rc;
+  if( (rc = doltliteAddRegister(db))!=SQLITE_OK ) return rc;
+  if( (rc = doltliteResetRegister(db))!=SQLITE_OK ) return rc;
+  if( (rc = doltliteMergeCmdRegister(db))!=SQLITE_OK ) return rc;
+  if( (rc = doltliteCherryPickRegister(db))!=SQLITE_OK ) return rc;
+  if( (rc = doltliteRevertRegister(db))!=SQLITE_OK ) return rc;
+  if( (rc = doltliteRebaseRegister(db))!=SQLITE_OK ) return rc;
+  if( (rc = doltliteConfigRegister(db))!=SQLITE_OK ) return rc;
   if( (rc = doltliteLogRegister(db))!=SQLITE_OK ) return rc;
   if( (rc = doltliteCommitAncestorsRegister(db))!=SQLITE_OK ) return rc;
   if( (rc = doltliteStatusRegister(db))!=SQLITE_OK ) return rc;
@@ -5785,15 +54,8 @@ int doltliteRegister(sqlite3 *db){
   if( (rc = doltliteSchemasRegister(db))!=SQLITE_OK ) return rc;
   if( (rc = doltliteDiffStatRegister(db))!=SQLITE_OK ) return rc;
   if( (rc = doltliteRemoteSqlRegister(db))!=SQLITE_OK ) return rc;
-  {
-    extern int doltliteHashofRegister(sqlite3*);
-    if( (rc = doltliteHashofRegister(db))!=SQLITE_OK ) return rc;
-  }
-  {
-    extern int doltliteConstraintViolationsRegister(sqlite3*);
-    if( (rc = doltliteConstraintViolationsRegister(db))!=SQLITE_OK ) return rc;
-  }
-
+  if( (rc = doltliteHashofRegister(db))!=SQLITE_OK ) return rc;
+  if( (rc = doltliteConstraintViolationsRegister(db))!=SQLITE_OK ) return rc;
   return doltliteMaybeSeedRepo(db);
 }
