+++ tool/mksqlite3c.tcl
@@ -614,7 +614,10 @@ proc emit_doltlite_engine_block {} {
     prolly_btree.c prolly_btree_catalog.c prolly_btree_cursor.c
     prolly_btree_mutation.c prolly_btree_orig.c prolly_btree_state.c
     prolly_btree_txn.c pager_shim.c sortkey.c
-    doltlite.c doltlite_commit.c doltlite_ref.c doltlite_log.c
+    doltlite.c doltlite_core.c doltlite_add.c doltlite_commit_cmd.c
+    doltlite_reset.c doltlite_merge_cmd.c doltlite_cherry_pick.c
+    doltlite_revert.c doltlite_rebase.c doltlite_config.c
+    doltlite_commit.c doltlite_ref.c doltlite_log.c
     doltlite_commit_ancestors.c doltlite_status.c doltlite_diff.c doltlite_diff_table.c
     doltlite_workspace.c
     doltlite_branch.c doltlite_tag.c doltlite_ancestor.c doltlite_merge.c
